### 2021-12-01 #3968 @whedon
[REVIEW]: retworkx: A High-Performance Graph Library for Python
**Submitting author:** @IvanIsCoding (<a href="http://orcid.org/0000-0002-8257-2103">Ivan Carvalho</a>)
**Repository:** https://github.com/Qiskit/rustworkx
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.12.0
