### 2021-12-01 #3968 @szhorvat
Since JOSS reviews are so different from the typical peer review, I'd like to let you know in advance how I plan to proceed, just to make sure we are all on the same page. Editors, authors, please let me know if I should change this approach.

I will prefix the titles of issues I open in the retworkx repository with `[joss]`. The fact that I opened an issue doesn't mean that it must be fixed before I would recommend the paper for publication. It will sometimes be meant as a discussion starter, to help me understand the library and its limitations. I understand that not every suggestion I might make would be reasonable to implement before publication. I would however appreciate responses to all the issues I open.

I will post things that concern only the paper (and not the library) directly here. The same with things that I believe concern the editors.

I will make it clear here when I consider the review completed. I won't be able to do it in a single go. Instead, I expect to have an ongoing conversation between the reviewers and authors for a couple of weeks. I will sometimes wait for answers to my comments/questions before proceeding. This is a large library and it will take time to get to know it.

It is the first time that I write such an introduction. I decided to do it because it has happened in the past that people seemed to have differing concepts about how reviews are carried out, and communication wasn't as effective as it could have been.
