### 2021-12-01 #3968 @szhorvat
And thus we arrive to my first comment. This checklist item isn't fulfilled at the moment:

> * **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Can you please let me know how I can seek support? I expect I will have lots of questions which you may not want to see in the issue tracker (not bug reports or feature suggestions). Please also make sure that support channels are clearly communicated to users.
