### 2021-12-01 #3968 @mtreinish
> And thus we arrive to my first comment. This checklist item isn't fulfilled at the moment:
>
> > * **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> Can you please let me know how I can seek support? I expect I will have lots of questions which you may not want to see in the issue tracker (not bug reports or feature suggestions). Please also make sure that the proper support channels are clearly communicated to users.

We have a public slack channel as well as an IRC channel on oftc (the irc channel is empty except for me right now). I'll push a PR that adds a section to the bottom of the readme shortly that mentions these.
