### 2021-12-01 #3968 @mtreinish
I just remembered we do briefly mention the slack channel in our contributing documentation already: https://github.com/Qiskit/retworkx/blob/main/CONTRIBUTING.md#module-directories-when-a-single-file-is-not-enough (which also gets put in our published documentation: https://qiskit.org/documentation/retworkx/CONTRIBUTING.html?highlight=slack#module-directories-when-a-single-file-is-not-enough) but I think that should still probably be mentioned in the README as well.
