### 2021-12-01 d2a1778
Add section to README about slack and irc channels
This commit adds a section to the bottom of the README that mentions our
slack and irc channels for retworkx if people would like to reach out
about the library.
