### 2021-12-01 #3968 @mtreinish
I opened the PR for this here: https://github.com/Qiskit/retworkx/pull/501
