### 2021-12-01 #3968 @szhorvat
@mtreinish Thanks for the response. I will join the Slack space, but I would prefer to also have the option of using some asynchronous communication channel that's public and keeps a record of the conversation (for the sake of the review). Is there a place where I can ask my questions that way? Should I just use the issue tracker? Have you considered GitHub Discussions? It's just an idea.

> I just remembered we do briefly mention the slack channel in our contributing documentation already

The support channels should be communicated clearly to normal users, not just would-be contributors. The README is a much better place for this. The PR looks good!
