### 2021-12-02 #3968 @szhorvat
For reference, the paper is here on arXiv: https://arxiv.org/abs/2110.15221
