### 2021-12-02 #3968 @szhorvat
Regarding the name `retworkx`: it obviously references `networkx`, which is by far the most popular graph library in Python. People are going to assume some kind of relationship, but I don't see what it is. Can you please clarify?

The API is different (not a replacement), the projects appear unrelated, and retworkx doesn't appear to be closer to networkx than any of the other Python network libraries. I am wondering if you have contacted the networkx project regarding the naming, and if they are happy with it.

I am not sure if this is within the scope of a JOSS review, but I thought it important to mention.
