### 2021-12-02 #3968 @szhorvat
I have now read the software paper, and I believe it needs a bit of improvement. In my view, the purpose of a JOSS paper is so that readers can get answers to questions such as:

 - What does this library do? What is already implemented and what are the long-term goals? Who is the (narrower or broader) target audience?
 - What novel contributions does it bring? What new things does it make possible?
 - Is this library useful for me? Will it be suitable for my project?
 - How does it differ from other similar libraries? What are its strengths and weaknesses? When and why should I choose this one over another?

As the paper currently stands, these questions are not fully answered.

## Statement of need

Here, the authors state _their own_ need, when working on the Qiskit project (networkx was used originally, but it was too slow, so retworkx was created to replace it).  I would expect to read about the target audience's need.  The broader target audience was set in the abstract: retworkx is "a flexible graph and network analysis library". I believe that the majority of people who are interested in such libraries will not have heard of Qiskit, thus talking about the needs of Qiskit will not be meaningful to them.

In short: What so-far unmet need does retworkx aim to fulfil in the field of graph and network analysis?

## Related work

In this section, the authors list a few (but not all) alternative libraries, and simply state that they were not suitable for the needs of Qiskit. Some of the issues are:
 - Once again, the focus is inwards facing, talking about the authors' own needs and Qiskit.
 - It is merely stated that other libraries fell short, but it is not explained how. Readers are not given any information that would help them decide about when to choose retworkx over another library for a specific project.

I should note that the landscape of graph libraries is very crowded in Python. It might seem like it is hard to stand out. Yet, as a co-developer of igraph, it is quite clear to me that it is nearly impossible to create a truly all-purpose graph library. There will always be tradeoffs, and between the many library choices there are gaps to fill in. But which of the gaps does retworkx fit into? I cannot tell after reading the paper.

To give some examples of what could be discussed, some libraries are targeted at graph theory researchers (mathematics). SageMath should be mentioned here. Some others (including most of those mentioned in this section) are mainly for complex network analysis, a very different field. In both these fields, users are likely to work with static graphs and use existing algorithms instead of implementing new ones. In contrast, some libraries may strive to provide a data structure for implementing efficient graph algorithms. There are many possible representation of graphs on a computer, all coming with their own tradeoffs: e.g. efficient modification is in conflict efficient graph traversal. Some libraries even provide multiple underlying representations in an attempt to deal with this.

Please clarify these questions in this section. SageMath and NetworKit should be mentioned. I would like to bring the authors' attention to [LEMON graph](https://lemon.cs.elte.hu/trac/lemon) as well, which stands apart from most other libraries with its focus on optimization and enabling the implementation of efficient algorithms. LEMON is C++ but has a (now abandoned?) Python interface. The decision to write everything from scratch, as retworkx seems to do, is a major one. It will take a very long time to catch up with other libraries, functionality-wise. It would be good to discuss why this decision was made instead of building on top of an existing solid library such as LEMON.

## Graph data structures

This is a good section because it aims to explain some of the conceptual foundations. It could be improved by adding some examples (e.g. to clarify what "callbacks" are). I suggest to take the opportunity to also explain the core capabilities of the library: libraries may support simple graphs, multigraphs with self-loops, directed/undirected or even mixed graphs, weighted graphs, vertex and edge attributes. Which of these features are or aren't supported?

It would be nice to touch on the tradeoffs I mentioned in the previous section, such as tradeoffs between efficient modification and traversal. How is the graph represented internally?

## Use Cases

This section is in fact about benchmarks that compare with igraph, graph-tool and networkx.  It would be good to also discuss some actual potential use cases, and mention examples of what tasks one may solve with the library.

The benchmarks are hosted in a different repository under a different user: https://github.com/mtreinish/retworkx-comparison-benchmarks  If the paper gets accepted, the current contents of the main repository will be archived along with the paper. With this organization, it looks like the benchmarks might be left out, which will be a problem for future reproducibility. This should be remedied. (@drvinceknight, correct me if I'm wrong about this.)

In fact, some of the benchmarks are flawed (I will open issues for this later—but in which repository should I do this?). If a reader would only see the benchmark results, without an opportunity to look at the benchmark code as well, they will come away with an incorrect impression ... all the more reason to bundle the code.

It's a bit unclear what is being compared between libraries. How fast they accomplish a task (with whatever algorithm they have implemented)? Or how fast the implementation of a specific algorithm is? Sometimes different algorithms are used (e.g. VF2 vs VF2++).

The benchmark intention should be made clearer.  The benchmark tasks should be described in more detail, e.g. with shortest paths: Weighted or unweighted? Directed or undirected? These do make a difference in some libraries. Single-source shortest path (implying all other vertices are targets) or the path between one source and one target (the search may be terminated as soon as the target is reached)?

There are only three benchmark tasks. What do we learn about the libraries from these? Why were these specific tasks chosen, and do you expect the performance findings to generalize to other tasks? In the case of subgraph isomorphism, I would say no. For shortest paths, they might, as many typical calculations on graphs rely on the same kind of graph traversal as in shortest path finding. However, the shortest path benchmark might be backed up with some others that exercise the library in similar ways (e.g., betweenness).

SNAP is omitted from the comparison "because its Python wrapper did not contain the required functions" (see footnote). Does this refer to VF2? If so, I do not think this is a fair reason to exclude it. Subgraph isomorphism is rarely needed for practical network analysis, and the fact that it's not implemented shouldn't immediately disqualify a library even from the other benchmarks.
