### 2021-12-02 Qiskit/retworkx#507 @szhorvat
[joss] Awkward phrasings in the documentation
Here I will keep posting various documentation problems that are too small to deserve their own GitHub issue.
