### 2021-12-02 Qiskit/retworkx#507 @szhorvat
In `betweenness_centrality`: https://qiskit.org/documentation/retworkx/stubs/retworkx.betweenness_centrality.html#retworkx.betweenness_centrality

> normalized (bool) – If true, normalize the betweenness scores by the number of distinct paths between all pairs of nodes.

I think what you mean is normalizing simply by the number of pairs? This is the usual way and this is what the function does, so no problems with the functionality.

Note: There are usually a very large number of paths. There are fewer shortest paths, but still potentially more than the number of pairs. By experimentation, I can see that it doesn't even use the number of pairs between which there is a path. Paths are not relevant to normalization here.
