### 2021-12-02 Qiskit/retworkx#507 @szhorvat
`mesh_graph` is a very unusual name for a concept that is normally called a "complete graph". It will likely be confusing to users. I am curious if there are any uses of this term out there, other than in retworkx.
