### 2021-12-02 Qiskit/retworkx#507 @szhorvat
The signature of `undirected_gnm_random_graph` is `undirected_gnm_random_graph(num_nodes, probability, seed=None, /)`. Th second argument is called `probability`. It should be `num_edges` or similar.
