### 2021-12-02 Qiskit/retworkx#507 @szhorvat
`retworkx.directed_gnm_random_graph` is documented as:

"Return a G_nm of a directed graph"

I know what the function does, but this description not meaningful. A graph does not have a "G_nm".

On a related note, the original paper of Erdős and Rényi deals with the G(n,m) model, although recently G(n,p) is also often referred to as "Erdős-Rényi". The retworkx docs apply this name only to G(n,p), which is the opposite of the historical naming.
