### 2021-12-02 Qiskit/retworkx#507 @szhorvat
"Return a new PyGraph object for an edge induced subgrapph of this graph", note "subgrapph" with double p. https://qiskit.org/documentation/retworkx/stubs/retworkx.PyGraph.edge_subgraph.html#retworkx.PyGraph.edge_subgraph
