### 2021-12-02 Qiskit/retworkx#507 @szhorvat
https://qiskit.org/documentation/retworkx/stubs/retworkx.PyGraph.neighbors.html#retworkx.PyGraph.neighbors

> The index of the node to get the **neibhors** of

>     A list of the neighbor node **indicies**
