### 2021-12-02 Qiskit/retworkx#509 @szhorvat
[joss] How to get edges incident to a vertex?
I was wondering about how one might implement certain graph algorithms, such as basic graph traversal, or obtaining a cycle basis in terms of edges, perhaps compute the edge-cycle matrix (see #505).

For such operations, it is necessary to be able to access the incident edges of a node (similar to `.neighbors()` or `.adj()` but returning edges). How can I do this?
