### 2021-12-02 #3968 @szhorvat
## Graph Creation benchmark

This benchmark is severely flawed.

The `igraph` "graph creation" benchmark actually first builds a `networkx` graph then converts that to an igraph object. This doesn't actually measure igraph's performance, and it's no wonder that igraph ends up looking slower than networkx ...  https://github.com/mtreinish/retworkx-comparison-benchmarks/blob/main/igraph_bench/gr_parser.py#L65

With igraph, the correct way to create a large graph like this is to first build an edge list as a list of tuples of vertex indices, then call `igraph.Graph(edges, n=vertex_count)` in one go. Do not call `.add_edge()` repeatedly, as this operation takes time proportional to the size of the graph. igraph is designed to analyse large static network efficiently, but it does not currently support efficient modification.

Note that building the graph this way may take less time than reading the entire huge gzipped file line by line using Python. Thus, it makes sense to separate the timings for reading the file vs building the graph. This brings us to the question: what does it even mean to time "creating a graph from an edge list"? The performance will depend on how that edge list was originally stored, and whether it first had to be converted to a suitable input format for a given library.
