### 2021-12-03 eb60a4a
Improve docstring for lexicographical topological sort
As was pointed out in #506 the docstring for the
lexicographical_topological_sort() function was not really clear as to
what it was or how it differed from the normal topological_sort()
function. This commit expands the docstring for the function to outline
specifically how it behaves.

Fixes #506
