### 2021-12-03 e2cd9bf
Add Cartesian Product (#470)
This commit adds the Cartesian Product for PyDiGraph and
PyGraph. After this commit it is possible to rewrite the
grid_graph generator to generate N-dimensional graphs.

Related to #440

* add cartesian product

* add edge weights

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* add weight handling

* lint

* fmt

* add capacity for the final_graph

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* fix digraph_cartesian_product doc

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com
[…]
