### 2021-12-03 Qiskit/retworkx#507 @szhorvat
The naming of the `.edges()` and `.nodes()` methods in the graph classes is confusing: these do not return edges and nodes, only the data associated with the edges and nodes.
