### 2021-12-03 Qiskit/retworkx#512 @szhorvat
[joss] What is the canonical way to represent edges?
As I see it, one of the critical design issues for any graph library is the representation of edges. How do I refer to an edge in situations such as:

 - Removing an edge
 - Changing the attributes/payload of an edge
 - Functions that return sets of edges
 - Functions that take edge sets as attributes

Edge representation becomes critical when it comes to the handling of multigraphs and distinguishing paralell edges.

There are generally two coherent approaches I have seen:

 1. Edges are represented as `(u, v)` pairs, where `u` and `v` are their endpoints. This alone does not make it possible to distinguish several edges connecting the same `u` and `v`. The solution is to add an extra "key": `(u, v, key)`.

    The only system that I am aware of that does this well is networkx. Mathematica also uses this paradigm, but support for keys was added only recently, and even now ensuring that keys are unique is a pain. Before Mathematica got support for edge keys, it was virtually impossible to perform certain operations on multigraphs.

    All other systems that I know of that prefer the `(u,v)` representation do not support multigraphs (e.g. Maple).

 2. Edges are represented through their indices within a consistently ordered edge list. Then the index can be used as a "handle" to the edge. This approach seems to be more common. The disadvantage is that the indices change when the graph itself is changed (e.g. edges are removed). Thus it takes effort to keep track of which edge was which after graph modification, when this is necessary.  (Some systems actually use opaque "handle" objects without letting the user know what's inside. This is similar enough to indices that I won't mention it separately.)

----

I would like to suggest to think about this topic in advance, and come up with a coherent design. What is the canonical way to refer to edges? Index? `(u, v, payload)` tuple?  Right now, retworkx seems to use a mishmash of approaches without ever spelling out what users should go with.

There are several custom return types such as `EdgeIndices`, `EdgeList`, `WeightedEdgeList`.  Functions that return edges (e.g. spanning tree) seem to prefer one of the "edge list" types, but this does not ensure that parallel edges are distinguishable. You might say that the payload serves to distinguish edges and that when the payload is the same, the difference between edges does not matter for any practical application. This is not true: consider e.g. a graph traversal such as DFS or BFS. How do I mark edges which have already been traversed? Or one you start to support operations on planar graphs, how will you represent a [combinatorial embedding](https://en.wikipedia.org/wiki/Rotation_system)?

Overall, the issue is that the `(u,v)` approach is the most common in retworkx, but this will not allow you to properly support multigraphs in the future.
