### 2021-12-03 Qiskit/retworkx#512 @szhorvat
This ties back to my comments in #505. While the API is unsuitable for supporting multigraphs, what would a suitable API actually look like? And how would you implement finding a fundamental cycle bases in pure Python in retworkx, without the ability to distinguish parallel edges? This would require a graph traversal which considers edges, not only nodes. Another missing piece for implementing this is #509.
