### 2021-12-03 Qiskit/retworkx#512 @szhorvat
A few questions which relate to the above:

What is the correct way to change the payload of the `i`th edge? Should I use `g.edges()[i] = ...`?

How do I get all edges between vertices `u` and `v` and remove the one(s) with the smallest weight (assuming numeric weight)? This illustrates a situation when it would be good to have a canonical edge representation. How will the returned edges be represented, so that we could do just about anything with them (find their endpoints, access their payload, remove them, etc.)?

Is there a function to "simplify" a graph, i.e. convert a multigraph to a simple graph by merging parallel edges and removing self-loops? How would payloads be combined?
