### 2021-12-03 Qiskit/retworkx#513 @szhorvat
[joss] Who uses retworkx and for what purpose?
I am wondering if you could show a few uses cases (outside of Qiskit) that illustrate what kinds retworkx is _currently_ suitable for. Do you have some short but non-trivial examples of analysing complex networks? Demonstrating some graph theory problems? Implementing a simple graph algorithm?

This ties into [my comments about the software paper](https://github.com/openjournals/joss-reviews/issues/3968#issuecomment-984648615).

One of the items to check off for JOSS reads:

> **Documentation:**
>
> Example usage: Do the authors include examples of how to use the software (**ideally to solve real-world analysis problems**).
