### 2021-12-03 Qiskit/retworkx#513 @szhorvat
It would also be interesting to see a couple of specific usages within Qiskit (I assume some of these are representative of other general use cases as well). This will give an idea of what tasks the library is already good at. What I mean is to actually describe the task that is being solved rather than just say that "it's used within Qiskit" without specifying how.
