### 2021-12-04 7db9c1a
Fixes for PyGraph sub nodes and tests.
