+++ tests/graph/test_substitute_nodes_with_node.py
@@ -15,69 +15,10 @@ import unittest
 import retworkx
 
 
-class TestSubstituteNodesCheckCycleSwitch(unittest.TestCase):
-    def setUp(self):
-        super().setUp()
-        self.dag = retworkx.PyDAG()
-        self.node_a = self.dag.add_node("a")
-        self.node_b = self.dag.add_child(self.node_a, "b", "b")
-        self.node_c = self.dag.add_child(self.node_b, "c", "c")
-
-        self.new_weight = "m"
-
-    def do_substitution(self, **kwargs):
-        """
-           ┌─┐                         ┌─┐
-         ┌─┤a│               ┌─────────┤m│
-         │ └─┘               │         └▲┘
-        ┌▼┐                 ┌▼┐         │
-        │b│          ───►   │b├─────────┘
-        └┬┘                 └─┘
-         │  ┌─┐
-         └─►┤c│
-            └─┘
-        """
-        self.dag.substitute_nodes_with_node(
-            [self.node_a, self.node_c], self.new_weight, **kwargs
-        )
-
-    def test_cycle_check_enable_local(self):
-        # Disable at class level.
-        self.dag.check_cycle = False
-
-        # Check removal is not allowed with explicit check_cycle=True.
-        self.assertRaises(
-            retworkx.DAGWouldCycle, self.do_substitution, check_cycle=True
-        )
-
-    def test_cycle_check_disable_local(self):
-        # Enable at class level.
-        self.dag.check_cycle = True
-
-        # Check removal is allowed for check_cycle=False
-        self.do_substitution(check_cycle=False)
-        self.assertEqual(set(self.dag.nodes()), {"b", "m"})
-
-    def test_cycle_check_inherit_class_enable(self):
-        # Enable at class level.
-        self.dag.check_cycle = True
-
-        # Check removal is not allowed.
-        self.assertRaises(retworkx.DAGWouldCycle, self.do_substitution)
-
-    def test_cycle_check_inherit_class_disable(self):
-        # Disable at class level.
-        self.dag.check_cycle = False
-
-        # Check removal is allowed.
-        self.do_substitution()
-        self.assertEqual(set(self.dag.nodes()), {"b", "m"})
-
-
 class TestSubstituteNodes(unittest.TestCase):
     def test_empty_nodes(self):
         """Replacing empty nodes is functionally equivalent to add_node."""
-        dag = retworkx.PyDAG()
+        dag = retworkx.PyGraph()
         dag.substitute_nodes_with_node([], "m")
 
         self.assertEqual(set(dag.nodes()), {"m"})
@@ -87,7 +28,7 @@ class TestSubstituteNodes(unittest.TestCase):
         Replacing all unknown nodes is functionally equivalent to add_node,
         since unknown nodes should be ignored.
         """
-        dag = retworkx.PyDAG()
+        dag = retworkx.PyGraph()
         dag.substitute_nodes_with_node([0, 1, 2], "m")
 
         self.assertEqual(set(dag.nodes()), {"m"})
@@ -96,32 +37,30 @@ class TestSubstituteNodes(unittest.TestCase):
         """
             ┌─┐              ┌─┐
          ┌4─┤a├─1┐           │m├──1───┐
-         │  └─┘  │           └▲┘      │
-        ┌▼┐     ┌▼┐           │      ┌▼┐
+         │  └─┘  │           └┬┘      │
+        ┌┴┐     ┌┴┐           │      ┌┴┐
[…]
