### 2021-12-04 Qiskit/retworkx#514 @szhorvat
[joss] How to find connected components in undirected graphs?
How can I find the connected components in an undirected graph?

For directed graphs, there are `retworkx.strongly_connected_components` and `retworkx.weakly_connected_components`, but these throw an error on undirected graphs.

This is one of the most fundamental operations so I expect there must be a function that I'm just missing?
