### 2021-12-04 Qiskit/retworkx#507 @szhorvat
On this page:

https://qiskit.org/documentation/retworkx/stubs/retworkx.BFSSuccessors.html#retworkx.BFSSuccessors

note `nex` instead of `next` on the last line of the example.
