### 2021-12-04 Qiskit/retworkx#514 @georgios-ts
No, currently such function is missing for undirected graphs but it should be a simple addition.
In general, `PyDiGraph` has received more attention due to its extensive use in qiskit-terra, so there might be some functionality gaps between undirected/directed graphs that we should definitely start closing.
