### 2021-12-04 a8474b5
Removed the weights options from barbell graph.
