+++ src/generators.rs
@@ -2307,9 +2307,8 @@ pub fn generalized_petersen_graph(
 /// Generate an undirected barbell graph where two identical mesh graphs are
 /// connected by a path.
 ///
-/// If neither ``num_path_nodes`` nor ``path_weights`` (both described
-/// below) are specified then this is equivalent to two mesh graphs joined
-/// together.
+/// If ``num_path_nodes`` (described below) is not specified then this is
+/// equivalent to two mesh graphs joined together.
 ///
 /// :param int num_mesh_nodes: The number of nodes to generate the mesh graphs
 ///     with. Node weights will be None if this is specified. If both
@@ -2319,12 +2318,6 @@ pub fn generalized_petersen_graph(
 ///     with. Node weights will be None if this is specified. If both
 ///     ``num_path_nodes`` and ``path_weights`` are set this will be ignored and
 ///     ``path_weights`` will be used.
-/// :param list mesh_weights: A list of node weights for the mesh graph. If both
-///     ``num_mesh_nodes`` and ``mesh_weights`` are set ``num_mesh_nodes`` will
-///     be ignored and ``mesh_weights`` will be used.
-/// :param list path_weights: A list of node weights for the path. If both
-///     ``num_path_nodes`` and ``path_weights`` are set ``num_path_nodes`` will
-///     be ignored and ``path_weights`` will be used.
 /// :param bool multigraph: When set to False the output
 ///     :class:`~retworkx.PyGraph` object will not be not be a multigraph and
 ///     won't  allow parallel edges to be added. Instead
@@ -2332,7 +2325,7 @@ pub fn generalized_petersen_graph(
 ///
 /// :returns: The generated barbell graph
 /// :rtype: PyGraph
-/// :raises IndexError: If neither ``num_mesh_nodes`` or ``mesh_weights`` are specified
+/// :raises IndexError: If ``num_mesh_nodes`` is not specified
 ///
 /// .. jupyter-execute::
 ///
@@ -2344,69 +2337,72 @@ pub fn generalized_petersen_graph(
 ///
 #[pyfunction(multigraph = true)]
 #[pyo3(
-    text_signature = "(/, num_mesh_nodes=None, num_path_nodes=None, mesh_weights=None, path_weights=None, multigraph=True)"
+    text_signature = "(/, num_mesh_nodes=None, num_path_nodes=None, multigraph=True)"
 )]
 pub fn barbell_graph(
     py: Python,
     num_mesh_nodes: Option<usize>,
     num_path_nodes: Option<usize>,
-    mesh_weights: Option<Vec<PyObject>>,
-    path_weights: Option<Vec<PyObject>>,
     multigraph: bool,
 ) -> PyResult<graph::PyGraph> {
-    let mut left_mesh =
-        mesh_graph(py, num_mesh_nodes, mesh_weights, multigraph)?;
-    let right_mesh = left_mesh.clone();
-    let meshlen = left_mesh.num_nodes();
+    if num_mesh_nodes.is_none() {
+        return Err(PyIndexError::new_err("num_mesh_nodes not specified"));
+    }
 
-    if num_path_nodes.is_some() || path_weights.is_some() {
-        let path_nodes: Vec<NodeIndex> = match path_weights {
-            Some(path_weights) => path_weights
-                .into_iter()
-                .map(|node| left_mesh.graph.add_node(node))
-                .collect(),
-            None => (0..num_path_nodes.unwrap())
-                .map(|_| left_mesh.graph.add_node(py.None()))
-                .collect(),
-        };
-
-        let pathlen = path_nodes.len();
-        if pathlen > 0 {
-            left_mesh.graph.add_edge(
-                NodeIndex::new(meshlen - 1),
-                NodeIndex::new(meshlen),
[…]
