### 2021-12-05 Qiskit/retworkx#507 @szhorvat
In the [shortest paths section of the documentation](https://qiskit.org/documentation/retworkx/api.html#shortest-paths), we see:

> retworkx.all_pairs_dijkstra_shortest_paths() | Find the shortest path from all nodes
>
> retworkx.all_pairs_dijkstra_path_lengths() | Find the shortest path from a node

This suggests that the difference between the functions is that one computes paths from a single source, while the other from all sources. This is not the case.
