### 2021-12-05 1f7509e
Fix `PyGraph.degree` if self-loops. (#518)
In an undirected graph, self-loops should increment by two (and not by one)
the degree of a node.
