### 2021-12-05 598a1e8
Fix dispatch function of `dfs_edges` and `union`. (#520)
* Fix dispatch function of `dfs_edges` and `union`.

This commit updates `dfs_edges` to set `source` argument to `None`
by default and fixes an oversight in `union` where user-defined values
of `merge_nodes` and `merge_edges` arguments were being ingored.

* add release note
