### 2021-12-05 Qiskit/retworkx#507 @szhorvat
The argument list of `.update_edge_by_index()` is wrong in the documentation.

https://qiskit.org/documentation/retworkx/stubs/retworkx.PyGraph.update_edge_by_index.html#retworkx.PyGraph.update_edge_by_index
