### 2021-12-06 Qiskit/retworkx#512 @szhorvat
Clarification: When I wrote this I did not realize that "indices" stay stable after removing vertices/edges. I assumed that they are integers that can be used to index into a vector (vector of nodes/edges).

Could you please let me know where the documentation explains the nature of these "indices"? It would be useful to have an introductory section that explains the basic concepts.
