### 2021-12-06 Qiskit/retworkx#512 @georgios-ts
> Could you please let me know where the documentation explains the nature of these "indices"? It would be useful to have an introductory section that explains the basic concepts.

It's briefly described in the first paragraph here https://qiskit.org/documentation/retworkx/stubs/retworkx.PyDiGraph.html#retworkx.PyDiGraph
