### 2021-12-06 Qiskit/retworkx#512 @szhorvat
Indeed it's there, I'm not sure how I missed it.

I find the term "index" to be a bit confusing. Perhaps another term, such as "ID" could avoid misunderstandings?

"Index" might be interpreted as "position in an array" by some people.
