### 2021-12-06 2e932ca
Rename substitute_nodes_with_node to contract_nodes.
