+++ tests/graph/test_contract_nodes.py
@@ -1,0 +1,216 @@
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+import unittest
+
+import retworkx
+
+
+class TestContractNodes(unittest.TestCase):
+    def test_empty_nodes(self):
+        """Replacing empty nodes is functionally equivalent to add_node."""
+        dag = retworkx.PyGraph()
+        dag.contract_nodes([], "m")
+
+        self.assertEqual(set(dag.nodes()), {"m"})
+
+    def test_unknown_nodes(self):
+        """
+        Replacing all unknown nodes is functionally equivalent to add_node,
+        since unknown nodes should be ignored.
+        """
+        dag = retworkx.PyGraph()
+        dag.contract_nodes([0, 1, 2], "m")
+
+        self.assertEqual(set(dag.nodes()), {"m"})
+
+    def test_cycle_path_len_gt_1(self):
+        """
+            ┌─┐              ┌─┐
+         ┌4─┤a├─1┐           │m├──1───┐
+         │  └─┘  │           └┬┘      │
+        ┌┴┐     ┌┴┐           │      ┌┴┐
+        │d│     │b│   ───►    │      │b│
+        └┬┘     └┬┘           │      └┬┘
+         │  ┌─┐  2            │  ┌─┐  2
+         └3─┤c├──┘            └3─┤c├──┘
+            └─┘                  └─┘
+        """
+        dag = retworkx.PyGraph()
+        node_a = dag.add_node("a")
+        node_b = dag.add_node("b")
+        node_c = dag.add_node("c")
+        node_d = dag.add_node("d")
+
+        dag.add_edge(node_a, node_b, 1)
+        dag.add_edge(node_b, node_c, 2)
+        dag.add_edge(node_c, node_d, 3)
+        dag.add_edge(node_a, node_d, 4)
+
+        node_m = dag.contract_nodes([node_a, node_d], "m")
+
+        self.assertEqual([node_b, node_c, node_m], dag.node_indexes())
+        self.assertEqual(
+            {(node_b, node_c), (node_c, node_m), (node_b, node_m)},
+            set(dag.edge_list()),
+        )
+
+    def test_multiple_paths_would_cycle(self):
+        """
+            ┌─┐     ┌─┐                  ┌─┐     ┌─┐
+         ┌3─┤c│     │e├─5┐            ┌──┤c│     │e├──┐
+         │  └┬┘     └┬┘  │            │  └┬┘     └┬┘  │
+        ┌┴┐  2  ┌─┐  4  ┌┴┐           │   2  ┌─┐  4   │
+        │d│  └──┤b├──┘  │f│   ───►    │   └──┤b├──┘   │
+        └─┘     └┬┘     └─┘           3      └┬┘      5
+                 1                    │       1       │
+                ┌┴┐                   │      ┌┴┐      │
+                │a│                   └──────┤m├──────┘
+                └─┘                          └─┘
+        """
+        dag = retworkx.PyGraph()
+        node_a = dag.add_node("a")
+        node_b = dag.add_node("b")
+        node_c = dag.add_node("c")
+        node_d = dag.add_node("d")
[…]

[…]
