### 2021-12-06 2fde9ce
Support undirected graphs in connected components.
This PR introduces several functions: `connected_components`, `number_connected_components`,
`node_connected_component` and `is_connected` for finding connected component in an
undirected graph. Previously, only directed graphs were supported.
