+++ src/connectivity/mod.rs
@@ -12,6 +12,7 @@
 
 #![allow(clippy::float_cmp)]
 
+mod connected_components;
 mod core_number;
 
 use super::{
@@ -28,7 +29,9 @@ use pyo3::Python;
 use petgraph::algo;
 use petgraph::graph::NodeIndex;
 use petgraph::unionfind::UnionFind;
-use petgraph::visit::{EdgeRef, IntoEdgeReferences, NodeCount, NodeIndexable};
+use petgraph::visit::{
+    EdgeRef, IntoEdgeReferences, NodeCount, NodeIndexable, Visitable,
+};
 
 use ndarray::prelude::*;
 use numpy::IntoPyArray;
@@ -227,6 +230,82 @@ pub fn digraph_find_cycle(
     EdgeList { edges: cycle }
 }
 
+/// Find the number of connected components in an undirected graph.
+///
+/// :param PyGraph graph: The graph to find the number of connected
+///     components on.
+///
+/// :returns: The number of connected components in the graph
+/// :rtype: int
+#[pyfunction]
+#[pyo3(text_signature = "(graph, /)")]
+fn number_connected_components(graph: &graph::PyGraph) -> usize {
+    connected_components::number_connected_components(&graph.graph)
+}
+
+/// Find the connected components in an undirected graph
+///
+/// :param PyGraph graph: The graph to find the connected components.
+///
+/// :returns: A list of sets where each set is a connected component of
+///     the graph
+/// :rtype: list
+#[pyfunction]
+#[pyo3(text_signature = "(graph, /)")]
+pub fn connected_components(graph: &graph::PyGraph) -> Vec<BTreeSet<usize>> {
+    connected_components::connected_components(&graph.graph)
+}
+
+/// Returns the set of nodes in the component of graph containing `node`.
+///
+/// :param PyGraph graph: The graph to be used.
+/// :param int node: A node in the graph.
+///
+/// :returns: A set of nodes in the component of graph containing `node`.
+/// :rtype: set
+///
+/// :raises InvalidNode: When an invalid node index is provided.
+#[pyfunction]
+#[pyo3(text_signature = "(graph, node, /)")]
+pub fn node_connected_component(
+    graph: &graph::PyGraph,
+    node: usize,
+) -> PyResult<BTreeSet<usize>> {
+    let node = NodeIndex::new(node);
+
+    if !graph.graph.contains_node(node) {
+        return Err(InvalidNode::new_err(
+            "The input index for 'node' is not a valid node index",
+        ));
+    }
+
+    Ok(connected_components::bfs_undirected(
+        &graph.graph,
+        node,
+        &mut graph.graph.visit_map(),
+    ))
+}
+
+/// Check if the graph is connected.
+///
+/// :param PyGraph graph: The graph to check if it is connected.
+///
+/// :returns: Whether the graph is connected or not
+/// :rtype: bool
+///
+/// :raises NullGraph: If an empty graph is passed in
+#[pyfunction]
+#[pyo3(text_signature = "(graph, /)")]
+pub fn is_connected(graph: &graph::PyGraph) -> PyResult<bool> {
+    match graph.graph.node_indices().next() {
+        Some(node) => {
+            let component = node_connected_component(graph, node.index())?;
+            Ok(component.len() == graph.graph.node_count())
+        }
+        None => Err(NullGraph::new_err("Invalid operation on a NullGraph")),
+    }
+}
+
 /// Find the number of weakly connected components in a directed graph
 ///
 /// :param PyDiGraph graph: The graph to find the number of weakly connected
@@ -254,7 +333,7 @@ fn number_weakly_connected_components(graph: &digraph::PyDiGraph) -> usize {
[…]
