### 2021-12-06 3c18100
Improve contract_nodes docstring.
