### 2021-12-06 Qiskit/retworkx#507 @szhorvat
The naming is inconsistent between `.node_indexes()` and `.edge_indices()`.
