### 2021-12-08 d8be17d
Raise if negative weights in shortest path algos.
Some shortest path algorithms, like `dijkstra_shortest_paths`,
`dijkstra_shortest_path_lengths`, `all_pairs_dijkstra_path_lengths`,
`all_pairs_dijkstra_shortest_paths`, `astar_shortest_path`,
`k_shortest_path_lengths` give incorrect results if there are negative
edge weights. This commit adds a check that raises if a user-defined weight
function returns a negative or NaN value.
