+++ tests/graph/test_k_shortest_path.py
@@ -11,6 +11,7 @@
 # under the License.
 
 import unittest
+import numpy
 
 import retworkx
 
@@ -54,3 +55,16 @@ class TestKShortestpath(unittest.TestCase):
             graph, start=1, k=1, edge_cost=lambda _: 1, goal=3
         )
         self.assertEqual({3: 2}, res)
+
+    def test_graph_k_shortest_path_with_invalid_weight(self):
+        graph = retworkx.generators.path_graph(4)
+        for invalid_weight in [numpy.nan, -1]:
+            with self.subTest(invalid_weight=invalid_weight):
+                with self.assertRaises(ValueError):
+                    retworkx.graph_k_shortest_path_lengths(
+                        graph,
+                        start=1,
+                        k=1,
+                        edge_cost=lambda _: invalid_weight,
+                        goal=3,
+                    )
