+++ tests/digraph/test_dijkstra.py
@@ -11,6 +11,7 @@
 # under the License.
 
 import unittest
+import numpy
 
 import retworkx
 
@@ -286,3 +287,45 @@ class TestDijkstraDiGraph(unittest.TestCase):
             expected,
             retworkx.digraph_all_pairs_dijkstra_shortest_paths(graph, float),
         )
+
+    def dijkstra_with_invalid_weights(self):
+        graph = retworkx.generators.directed_path_graph(2)
+        for invalid_weight in [numpy.nan, -1]:
+            for as_undirected in [False, True]:
+                with self.subTest(
+                    invalid_weight=invalid_weight, as_undirected=as_undirected
+                ):
+                    with self.assertRaises(ValueError):
+                        retworkx.digraph_dijkstra_shortest_paths(
+                            graph,
+                            source=0,
+                            weight_fn=lambda _: invalid_weight,
+                            as_undirected=as_undirected,
+                        )
+
+    def dijkstra_lengths_with_invalid_weights(self):
+        graph = retworkx.generators.directed_path_graph(2)
+        for invalid_weight in [numpy.nan, -1]:
+            with self.subTest(invalid_weight=invalid_weight):
+                with self.assertRaises(ValueError):
+                    retworkx.digraph_dijkstra_shortest_path_lengths(
+                        graph, node=0, edge_cost_fn=lambda _: invalid_weight
+                    )
+
+    def all_pairs_dijkstra_with_invalid_weights(self):
+        graph = retworkx.generators.directed_path_graph(2)
+        for invalid_weight in [numpy.nan, -1]:
+            with self.subTest(invalid_weight=invalid_weight):
+                with self.assertRaises(ValueError):
+                    retworkx.digraph_all_pairs_dijkstra_shortest_paths(
+                        graph, edge_cost_fn=lambda _: invalid_weight
+                    )
+
+    def all_pairs_dijkstra_lenghts_with_invalid_weights(self):
+        graph = retworkx.generators.directed_path_graph(2)
+        for invalid_weight in [numpy.nan, -1]:
+            with self.subTest(invalid_weight=invalid_weight):
+                with self.assertRaises(ValueError):
+                    retworkx.digraph_all_pairs_dijkstra_path_lengths(
+                        graph, edge_cost_fn=lambda _: invalid_weight
+                    )
