+++ releasenotes/notes/negative-weights-in-shortest-path-algos-b205b63bd12d1a26.yaml
@@ -1,0 +1,11 @@
+---
+fixes:
+  - |
+    Fixes an oversight in shortest path algorithms like:
+    :func:`~retworkx.dijkstra_shortest_paths`,
+    :func:`~retworkx.dijkstra_shortest_path_lengths`,
+    :func:`~retworkx.all_pairs_dijkstra_path_lengths`,
+    :func:`~retworkx.all_pairs_dijkstra_shortest_paths`,
+    :func:`~retworkx.astar_shortest_path`,
+    :func:`~retworkx.k_shortest_path_lengths`, that would previously
+    accept edge weights with negative or NaN values.
