### 2021-12-08 ac6825a
Apply suggestions from code review
Co-authored-by: Jake Lishman <jake@binhbar.com>
