+++ MANIFEST.in
@@ -4,4 +4,4 @@ include Cargo.toml
 include Cargo.lock
 recursive-include src *
 recursive-include retworkx-core *
-recursive-include retworkx *.pyi *.typed
+recursive-include retworkx *.pyi py.typed
