+++ releasenotes/notes/graph-annotations-1d436930bf60c5c2.yaml
@@ -3,4 +3,4 @@ features:
   - |
     Added type annotations to :class:`~retworkx.PyDiGraph` and
     :class:`~retworkx.PyGraph`. They can now be statically type checked
-    with `mypy <http://mypy-lang.org>`_
+    with `mypy <http://mypy-lang.org>`__.
