### 2021-12-09 604a50f
use python built-in NaN value in tests
