+++ tests/digraph/test_dijkstra.py
@@ -11,7 +11,6 @@
 # under the License.
 
 import unittest
-import numpy
 
 import retworkx
 
@@ -290,7 +289,7 @@ class TestDijkstraDiGraph(unittest.TestCase):
 
     def dijkstra_with_invalid_weights(self):
         graph = retworkx.generators.directed_path_graph(2)
-        for invalid_weight in [numpy.nan, -1]:
+        for invalid_weight in [float("nan"), -1]:
             for as_undirected in [False, True]:
                 with self.subTest(
                     invalid_weight=invalid_weight, as_undirected=as_undirected
@@ -305,7 +304,7 @@ class TestDijkstraDiGraph(unittest.TestCase):
 
     def dijkstra_lengths_with_invalid_weights(self):
         graph = retworkx.generators.directed_path_graph(2)
-        for invalid_weight in [numpy.nan, -1]:
+        for invalid_weight in [float("nan"), -1]:
             with self.subTest(invalid_weight=invalid_weight):
                 with self.assertRaises(ValueError):
                     retworkx.digraph_dijkstra_shortest_path_lengths(
@@ -314,7 +313,7 @@ class TestDijkstraDiGraph(unittest.TestCase):
 
     def all_pairs_dijkstra_with_invalid_weights(self):
         graph = retworkx.generators.directed_path_graph(2)
-        for invalid_weight in [numpy.nan, -1]:
+        for invalid_weight in [float("nan"), -1]:
             with self.subTest(invalid_weight=invalid_weight):
                 with self.assertRaises(ValueError):
                     retworkx.digraph_all_pairs_dijkstra_shortest_paths(
@@ -323,7 +322,7 @@ class TestDijkstraDiGraph(unittest.TestCase):
 
     def all_pairs_dijkstra_lenghts_with_invalid_weights(self):
         graph = retworkx.generators.directed_path_graph(2)
-        for invalid_weight in [numpy.nan, -1]:
+        for invalid_weight in [float("nan"), -1]:
             with self.subTest(invalid_weight=invalid_weight):
                 with self.assertRaises(ValueError):
                     retworkx.digraph_all_pairs_dijkstra_path_lengths(
