+++ tests/digraph/test_k_shortest_path.py
@@ -11,7 +11,6 @@
 # under the License.
 
 import unittest
-import numpy
 
 import retworkx
 
@@ -77,7 +76,7 @@ class TestKShortestpath(unittest.TestCase):
 
     def test_digraph_k_shortest_path_with_invalid_weight(self):
         graph = retworkx.generators.directed_path_graph(4)
-        for invalid_weight in [numpy.nan, -1]:
+        for invalid_weight in [float("nan"), -1]:
             with self.subTest(invalid_weight=invalid_weight):
                 with self.assertRaises(ValueError):
                     retworkx.digraph_k_shortest_path_lengths(
