### 2021-12-09 20d904e
Fix `unweighted_average_shortest_path_length` if disconnected graph.
Fixes the way `unweighted_average_shortest_path_length` handles
disconnected graphs. Disconnected pairs of nodes are now assumed
to have an infinite distance. In addition, an extra kwarg ``disconnected``
was added where, if set to ``True``, the average is taken only over
connected node pairs. By default, it's set to ``False``.
