+++ retworkx/__init__.py
@@ -180,19 +180,21 @@ def _graph_distance_matrix(graph, parallel_threshold=300, null_value=0.0):
 
 
 @functools.singledispatch
-def unweighted_average_shortest_path_length(graph, parallel_threshold=300):
+def unweighted_average_shortest_path_length(
+    graph, parallel_threshold=300, disconnected=False
+):
     r"""Return the average shortest path length with unweighted edges.
 
     The average shortest path length is calculated as
 
     .. math::
 
-        a =\sum_{s,t \in V} \frac{d(s, t)}{n(n-1)}
+        a =\sum_{s,t \in V, s \ne t} \frac{d(s, t)}{n(n-1)}
 
     where :math:`V` is the set of nodes in ``graph``, :math:`d(s, t)` is the
     shortest path length from :math:`s` to :math:`t`, and :math:`n` is the
-    number of nodes in ``graph``. This also assumes that :math:`d(s, t) = 0`
-    if :math:`t` cannot be reached from :math:`s`.
+    number of nodes in ``graph``. If ``disconnected`` is set to ``True``,
+    the average will be taken only between connected nodes.
 
     This function is also multithreaded and will run in parallel if the number
     of nodes in the graph is above the value of ``parallel_threshold`` (it
@@ -201,17 +203,20 @@ def unweighted_average_shortest_path_length(graph, parallel_threshold=300):
     By default it will use all available CPUs if the environment variable is
     not specified.
 
-    :param PyDiGraph graph: The graph to compute the average shortest path length
-        for
+    :param graph: The graph to compute the average shortest path length for,
+        can be either a :class:`~retworkx.PyGraph` or :class:`~retworkx.PyDiGraph`.
     :param int parallel_threshold: The number of nodes to calculate the
         the distance matrix in parallel at. It defaults to 300, but this can
         be tuned to any number of nodes.
     :param bool as_undirected: If set to ``True`` the input directed graph
         will be treated as if each edge was bidirectional/undirected while
         finding the shortest paths. Default: ``False``.
+    :param bool disconnected: If set to ``True`` only connected vertex pairs
+        will be included in the calculation. If ``False``, infinity is returned
+        for disconnected graphs. Default: ``False``.
 
-    :returns: The average shortest path length. If the graph is empty this
-        will return NaN and if there is a single node 0 will be returned.
+    :returns: The average shortest path length. If no vertex pairs can be included
+        in the calculation this will return NaN.
 
     :rtype: float
     """
@@ -220,19 +225,22 @@ def unweighted_average_shortest_path_length(graph, parallel_threshold=300):
 
 @unweighted_average_shortest_path_length.register(PyDiGraph)
 def _digraph_unweighted_average_shortest_path_length(
-    graph, parallel_threshold=300, as_undirected=False
+    graph, parallel_threshold=300, as_undirected=False, disconnected=False
 ):
     return digraph_unweighted_average_shortest_path_length(
         graph,
         parallel_threshold=parallel_threshold,
         as_undirected=as_undirected,
+        disconnected=disconnected,
     )
 
 
 @unweighted_average_shortest_path_length.register(PyGraph)
-def _graph_unweighted_shortest_path_length(graph, parallel_threshold=300):
+def _graph_unweighted_shortest_path_length(
+    graph, parallel_threshold=300, disconnected=False
+):
     return graph_unweighted_average_shortest_path_length(
-        graph, parallel_threshold=parallel_threshold
+        graph, parallel_threshold=parallel_threshold, disconnected=disconnected
     )
 
 
