+++ src/shortest_path/mod.rs
@@ -1094,12 +1094,12 @@ pub fn graph_distance_matrix(
 ///
 /// .. math::
 ///
-///     a =\sum_{s,t \in V} \frac{d(s, t)}{n(n-1)}
+///     a =\sum_{s,t \in V, s \ne t} \frac{d(s, t)}{n(n-1)}
 ///
 /// where :math:`V` is the set of nodes in ``graph``, :math:`d(s, t)` is the
 /// shortest path length from :math:`s` to :math:`t`, and :math:`n` is the
-/// number of nodes in ``graph``. This also assumes that
-/// :math:`d(s, t) = 0` if :math:`t` cannot be reached from :math:`s`.
+/// number of nodes in ``graph``. If ``disconnected`` is set to ``True``,
+/// the average will be taken only between connected nodes.
 ///
 /// This function is also multithreaded and will run in parallel if the number
 /// of nodes in the graph is above the value of ``parallel_threshold`` (it
@@ -1116,32 +1116,48 @@ pub fn graph_distance_matrix(
 /// :param bool as_undirected: If set to ``True`` the input directed graph
 ///     will be treated as if each edge was bidirectional/undirected while
 ///     finding the shortest paths. Default: ``False``.
+/// :param bool disconnected: If set to ``True`` only connected vertex pairs
+///     will be included in the calculation. If ``False``, infinity is returned
+///     for disconnected graphs. Default: ``False``.
 ///
-/// :returns: The average shortest path length. If the graph is empty this
-///     will return NaN and if there is a single node 0 will be returned.
+/// :returns: The average shortest path length. If no vertex pairs can be included
+///     in the calculation this will return NaN.
 /// :rtype: float
-#[pyfunction(parallel_threshold = "300", as_undirected = "false")]
+#[pyfunction(
+    parallel_threshold = "300",
+    as_undirected = "false",
+    disconnected = "false"
+)]
 #[pyo3(
-    text_signature = "(graph, /, parallel_threshold=300, as_undirected=False)"
+    text_signature = "(graph, /, parallel_threshold=300, as_undirected=False, disconnected=False)"
 )]
 pub fn digraph_unweighted_average_shortest_path_length(
     graph: &digraph::PyDiGraph,
     parallel_threshold: usize,
     as_undirected: bool,
+    disconnected: bool,
 ) -> f64 {
     let n = graph.node_count();
-    if n == 0 {
+    if n <= 1 {
         return std::f64::NAN;
     }
-    if n == 1 {
-        return 0.0;
-    }
-    let sum = average_length::compute_distance_sum(
+
+    let (sum, conn_pairs) = average_length::compute_distance_sum(
         &graph.graph,
         parallel_threshold,
         as_undirected,
-    ) as f64;
-    sum / (n * (n - 1)) as f64
+    );
+
+    let tot_pairs = n * (n - 1);
+    if disconnected && conn_pairs == 0 {
+        return std::f64::NAN;
+    }
+
+    if !disconnected && conn_pairs < tot_pairs {
+        return std::f64::INFINITY;
+    }
+
+    (sum as f64) / (conn_pairs as f64)
 }
 
 /// Return the average shortest path length for a :class:`~retworkx.PyGraph`
@@ -1151,12 +1167,12 @@ pub fn digraph_unweighted_average_shortest_path_length(
 ///
 /// .. math::
 ///
-///     a =\sum_{s,t \in V} \frac{d(s, t)}{n(n-1)}
+///     a =\sum_{s,t \in V, s \ne t} \frac{d(s, t)}{n(n-1)}
 ///
 /// where :math:`V` is the set of nodes in ``graph``, :math:`d(s, t)` is the
 /// shortest path length from node :math:`s` to node :math:`t`, and :math:`n`
[…]

[…]
