### 2021-12-09 Qiskit/retworkx#521 @georgios-ts
Here we need to improve documentation and state more explicitly that this function returns the data payload of parent/children node, and add support for undirected graphs.

@mtreinish But there might be also a "bug" in this function since it returns *all* adjacent nodes of a parent node, and *not* the children nodes in the BFS tree. This might be unexpected since in PR #17, was stated that this function is equivalent to https://networkx.org/documentation/stable/reference/algorithms/generated/networkx.algorithms.traversal.breadth_first_search.bfs_successors.html#networkx.algorithms.traversal.breadth_first_search.bfs_successors, which it isn't.
