### 2021-12-09 Qiskit/retworkx#521 @mtreinish
Hmm, yeah that's a good point, I agree it's probably broken. I added that for [`DAGCircuit.bfs_successors()`](https://github.com/Qiskit/qiskit-terra/blob/main/qiskit/dagcircuit/dagcircuit.py#L1462)  which is not widely used so I could see that being in an issue.
