### 2021-12-10 #3968 @szhorvat
@drvinceknight: This is my final report.

----

In the past few days, I have been thinking about how to go forward with the review. My opinion is that the submission to JOSS was premature. This package certainly shows promise, and I think it would make sense to resubmit it in the future, maybe in a year's time. However, trying to work out all the issues within the timeframe of the review does not seem realistic. It would put too much pressure on the authors, as well as on the reviewers.

### Software paper

Please refer to my comments above. In short, as currently written, the paper does not provide enough information to readers to decide whether the library would be useful for them.

I leave the checklist items for "Software Paper" unchecked for now (quality of writing and references would need to be re-checked anyway one the paper is updated).

### Benchmarks

The library appears to be fast, but there are two issues with the benchmarks, as presented:

 - They do not demonstrate the performance adequately. The relevance of benchmarks to various functionality areas should be discussed.
 - The first benchmark is not performed correctly on other libraries (such as igraph) and misrepresents their performance. I would like to ask the authors to correct this soon, as the benchmarks are already being discussed on forums](https://discourse.julialang.org/t/retworkx-new-high-performance-python-graph-library/71109), and [the arXiv version](https://arxiv.org/abs/2110.15221) of the paper containing them is linked from [the retworkx GitHub page](https://github.com/Qiskit/retworkx/#authors-and-citation)

More comments are above. I left the Performance item unchecked on the list for now.

### The library

According to the submission, retworkx aims to be a new general-purpose Python library for working with graphs. In my comments about the paper, I discussed how different use cases could be targeted within this broader goal. Possibilities are:

 - analysing real-world network data (a network science tool)
 - a tool for graph theory (for mathematicians, like SageMath)
 - a tool to prototype new graph algorithms

It seems to me that at this time, the library is not practically usable for either of these three tasks. There are too many functionality gaps. For example, even [decomposing to connected components is not implemented for undirected graphs](https://github.com/Qiskit/retworkx/issues/514) (only for directed ones). I quickly found bugs in fundamental functionality, such as computing [degrees](https://github.com/Qiskit/retworkx/issues/517) (now fixed). The core API misses features critical for implementing graph algorithms (e.g. [getting incident edges](https://github.com/Qiskit/retworkx/issues/509)). Other examples are in the issues I opened.

I could not check off the Functionality item at this point.

The documentation is [somewhat haphazardly written](https://github.com/Qiskit/retworkx/issues/507), with many typos and sometimes [unintelligible descriptions](https://github.com/Qiskit/retworkx/issues/521). The same applies to the naming used in the API, e.g. `node_indexes` vs `edge_indices`, a continual conflation of the terms "weight", "payload", "edge" (all three of which sometimes refer to the very same thing), etc. The impression one gets is that even fundamental parts library were put together in a bit of a rush, in response to an immediate need for a specific application within Qiskit.

I could not check off several of the Documentation items.

All that said, the library is built on solid foundations (the `petgraph` Rust library), and has good potential to grow and progress.

### Advice to the authors

While the landscape of Python graph libraries is crowded, there is definitely space for newcomers. However, new libraries should have a _vision_ that distinguishes them. Which gap are you trying to fill?  Think about what you are trying to achieve, and instead of letting the library grow organically (one might say _haphazardly_), think about a coherent design that fits the proposed use case.

It seem to me that retworkx's biggest problem is that it has a single driver: being a component in Qiskit. If it does just that—serve very specific needs within a _single_ project that does not even deal with networks—it is unlikely to become useful to a broader audience, and would not be publishable. I elaborated on this a bit in my comments about the software paper. Let actual general use cases be the driver behind development. Would you like it to be useful for analysing network data? Then you, the developers, should complete a number of network analysis projects. That will show you what is missing (e.g. can you do the first step, import data from common formats?) Talk to people who work in network science. Do you want it to be useful for prototyping graph algorithms? Then do that: implement algorithms not in Rust, but in pure Python using retworkx. Graph theory? Read Math.StackExchange or even MathOverflow, find computational tasks, see if you can solve them.

Finally, you have the advantage of starting from scratch. You are not carrying all the heavy historical baggage of other graph libraries like igraph or networkx. First, understand the other libraries' limitations, understand the mistakes they made, learn from them. **Come up with a better design.** I see functionality being piled on, but very little _design_. Why are you foregoing this unique opportunity?  To be frank, the software paper gave me the impression that retworkx was born out of a "not invented here" mindset, not from a detailed understanding of what is missing in existing Python graph libraries.
