### 2021-12-11 #3968 @mtreinish
First, I want to thank you for your detailed look at the retworkx so far. I was hoping the paper review would serve as a chance for people with different sets of experience and viewpoints to review the library and point out gaps and all the issues you've opened will help us make a better library. One of my personal goals with retworkx recently is to expand it's usefulness for more domains including a more traditional network analysis and getting your perspective on this has been quite valuable and you've identified a lot of places we can make improvements. But, I think because your critique here is coming from a completely different viewpoint you're inadvertently missing the use cases for which retworkx was developed and the value it's providing. What you're attributing to a lack of good design are just different priorities and design decisions we've made in the development of retworkx. This doesn't preclude making the changes you're looking for, but it also doesn't invalidate what has been done so far or how retworkx is used today.

In the development of retworkx so far we've been more focused on applications that use graph or network analysis as a component or tool for a specific larger goal or application. This is why we've designed retworkx to be dynamic and flexible in its use. While retworkx was originally designed to meet Qiskit's needs from a graph library, Qiskit is not a single application or exclusively what retworkx can be used for. More specifically, Qiskit is a constellation of different software projects each working on a different aspect of quantum computing, including applications of quantum computing. The original use case for retworkx was in the internal use by the compiler where a directed acyclic graph is used to [represent a quantum circuit ](https://github.com/Qiskit/qiskit-terra/blob/0.19.1/qiskit/dagcircuit/dagcircuit.py) (i.e. a program for a quantum computer). The compiler analyzes that graph and performs transformations on it to fit and optimize the input program to the target QPU. This is where we hit scaling limits, which prompted the development of retworkx. In this application dynamically building and modifying large graphs with thousands of nodes fast is very important as it's the bottleneck for compilation. Another application of graphs in Qiskit's compiler uses a directed graph to [model the connectivity of qubits on a QPU ](https://github.com/Qiskit/qiskit-terra/blob/main/qiskit/transpiler/coupling.py), which is one of the hardware constraints the compiler needs to adapt an input program to. An example application of this connectivity graph we recently just developed was a new compiler technique applying subgraph isomorphism to attempt to [find a perfect mapping](https://github.com/Qiskit/qiskit-terra/blob/0.19.1/qiskit/transpiler/passes/layout/vf2_layout.py) of the qubits used in a quantum circuit to the physical qubits in a QPU. But this is just in the base layer of Qiskit (the interface between software and hardware) and there are higher level applications of quantum computing which are using retworkx in different manners. For example, recently the Qiskit Nature project, which is concerned with applications of quantum computing for natural sciences, started using retworkx recently to create [lattice models](https://gist.github.com/k-tamuraphys/6da54c468fffca3b4d18d9837de81aaf) to perform physics simulations.

Looking outside of Qiskit there are additional applications currently using retworkx such as [qtcodes](https://github.com/yaleqc/qtcodes), which is a downstream user of Qiskit and retworkx directly. qtcodes is using retworkx to research new techniques for quantum error correction (you can see some examples of this here: https://github.com/yaleqc/qtcodes/blob/master/tutorials/rep/2-fitters.ipynb but I'm not an expert in QEC and won't be able to provide an adequate explanation of how it works). Another research project using retworkx is [atompack](https://github.com/seatonullberg/atompack) (who were a very early user of retworkx) where they use retworkx to create graph models of atomic structures. Both of these (which we cited in the paper) are using retworkx to generate graphs dynamically as part of a larger application.

As for things you highlight like inconsistencies in the API, bugs in certain functions, or issues in the documentation, many of those are valid things to fix, and highlight why getting multiple sets of eyes on software is so useful, and why I value this review process. But, for some of the issues you've outlined I think this is another place that highlights where we have a difference in priorities. One of the things we value strongly in retworkx is backwards compatibility. While the project may have only existed for a couple of years we already have a large user base both via Qiskit but also outside of it and we can't simply change an API to make naming consistent or change documented behavior without potentially breaking people's applications. For example, one of the quirks you've highlighted, the difference between `node_indexes()` and `edge_indices()`, is a known issue and dates back to one of the earliest retworkx releases. We've settled on the name `indices` in the API and documentation and that's what we've been using whenever we add new functionality that consumes or returns multiple identifiers. But, the `node_indexes()` method existed prior to that decision and simply renaming that function would break current users which isn't acceptable to us. We have to make any change like that gradually to not disrupt existing working applications and enabling people to migrate to a new `node_indices()` method. But, making that transition hasn't been a priority as it has real cost for users and most people can accept this and understand what both methods do.

For the paper, part of why we didn't go into too much detail there is partially because of the length limits. Reading the expectations of the JOSS paper we were cautious to try and avoid explaining all the internal details of the library and saving that for the documentation. The other aspect was we were trying to avoid putting any project in a negative light. We felt it was important to showcase how retworkx's performance compares to other popular graph libraries for python because the we felt the performance characteristics of the library is one of the key reasons you would use retworkx. But, we weren't out to paint any particular library in a negative light as they're all useful for different applications and tried to avoid going into too deep an analysis or comparison because the purpose of the paper is not be a survey comparing in detail different graph libraries in Python. But, it is also difficult to explain all these nuances in just ~1000 words. We'll try to tweak the paper based on your suggestions and update the review here with an updated draft when it's ready.

For the specific thing about the creation benchmark, you're right we shouldn't have used the networkx converter there with igraph. This was done mainly because creating a graph with igraph in the same manner as all the other libraries benchmarked was exceedingly slow (taking hours to create the graph). As we're not experts in igraph we referred to the documentation and there was no reference that we could find mentioning what you mentioned with the performance characteristics of the `add_edge()` method. We used networkx and converted to igraph because we wanted to have numbers in the same ballpark (instead of having it show as orders of magnitude slower). But you're right that's not a good demonstration of how igraph performs. As for the validity of the benchmark I have to disagree with you there, the benchmark is a valid use case and one we value heavily in Qiskit's use of retworkx. Being able to iteratively create a graph quickly is important in some applications, especially those that are using networks as part of a larger computation for analysis where the graph is dynamic. Not every application is able to statically create a graph all at once from a static edge list and then perform analysis. Looping over an input file like that and creating the graph one edge at a time showcases the time it takes to build a graph iteratively. But I agree we didn't do the best job explaining this in the paper, we'll update the wording to make it a bit more clear what we're trying to show with that benchmark. As for igraph we'll update the graphs to just not include igraph and mention it's not suitable for applications where you need to dynamically create large graphs iteratively.
