### 2021-12-12 #3968 @szhorvat
Let me sum up my recommendation, to avoid misunderstandings.

The library shows promise, but it does not look mature enough. When trying to use it for various tasks, its limits are reached very quickly and the functionality gaps become apparent. It was easy to find bugs, which again suggests that it has not been thoroughly tested through real-world usage not related to Qiskit and its dependents. I think resubmission at a more mature stage would make sense.

At that time, it would help to (1) state the scope clearly (2) demonstrate use cases within the scope. As a comparison, another library I reviewed recently [has several worked examples in its documentation](https://cellpylib.org/) which shows that it can be used to solve new tasks within its scope. The paper word limit should not be an issues since many things are better placed in the documentation anyway.

All the issues I brought up are fixable, and I hope to see retworkx succeed and eventually published in JOSS. However, in my opinion, this would take much more time than what is reasonable during a review. I do not think that it is in the authors' interest to try to address all concerns under time pressure, especially since parts of the library already make the impression that they were created in a rush. Also consider that retworkx, like most graph libraries (and perhaps unlike most JOSS submissions), has a very large scope, thus it simply takes more time and effort to bring it to a reasonable level of completion.

I will address some of the specifics in @mtreinish's response at a later time.
