### 2021-12-12 #3968 @inakleinbottle
@drvinceknight I don't seem to have the ability to update the checklist in the initial issue post.

I have completed my initial review of the software and paper.

# General comments
I'm glad to see Rust being used as a safe and fast back-end for serious computation, and I have no doubt that this fulfills the needs of the Qiskit project. However, I too share some of the concerns of the other reviewer about the utility of the library for a more general audience, and to the benchmark methodology.

A cursory look at the reference seems to suggest that retworkx has some support for other graph algorithms such as cutting and colouring, but there are not mentioned at all in your benchmarks, and the support looks relatively limited as compared to networkx, igraph, or graph-tool. To be clear, I would not expect feature parity with these very mature libraries and this is not a criticism that I put much weight behind, but I would perhaps like to see benchmarks for all of the functionality you have (and that are stable) against these other libraries to show some direct benefit of retworkx over other tools. The reason I mention this is that the name of the library "retworkx" implies a fast alternative to "networkx", but in reality only a small subset of the functionality is represented as of yet.

# Documentation
The statement of need suggests a need for a drop-in replacement for the networkx library in which the core algorithms and data structures are passed to a compiled binary to do the actual work, thus improving the performance with minimal changes to the Python API. I agree that there is a place for such a library in the current ecosystem, and I am happy to accept that the external API of the other graph libraries mentioned do not fulfill the requirements for which retworkx was designed. I cannot comment on whether more general users of graph libraries in Python would find this to be true.

The installation instructions provided in the documentation are excellent and, since this is distributed as a wheel containing the compiled binaries on my architecture and OS, I had no problems installing the library for testing. (This should be the case for most users, as wheels are available for Python 3.6-3.9 on most of the major architectures. However, since Python3.10 was release recently, the authors may wish to consider adding a wheel for Python 3.10 too.)
As for compiling the library from source, I have some concerns that I will detail below.

The API documentation is fine, and the core classes have plenty of example usage, though I have not checked exhaustively.  However, the algorithm functions have rather sparse documentation. It would be good to have at least one example usage in each of the main algorithm methods, even if this is a trivial example, so users can see what a typical call might look like. I would like to see more examples from a top level that walk through how to use the library to construct, manipulate, and use graph objects in a more complete pipeline; I have always found such "tutorial" style documentation to be a great way to cover a great deal of the usage without users needing to dig into API documentation.

The library has a suite of tests, which as far as I can see test results against networkx (again, I have not checked every test file). This is fine, assuming that networkx does not contain any bugs. I would suggest, if there are not already, adding tests that explicitly construct correct results to compare computed results against rather than relying on networkx to test the correctness.

You have documentation about how people might contributed to the library, but nothing to suggest how users should raise issues. I assume, since the project is hosted on GitHub, that they should create an issue on the repository. I would add a small section to the end of the Readme document to confirm this, or point users to the place that they can file a bug report. (This is set on the "bug tracker" attribute on PyPI.)

# Functionality
As I mentioned, I had no problems installing the library from wheel. However, I noticed that the setup.py file requires the rust-specific setuptools addon yet this is not listed as a build-time dependency. Rather, the setup.py file attempts to load this addon, and if it fails uses a subprocess call to pip to install the necessary package and then attempts to proceed. I understand why the authors have structured the code in this way, but I don't think this is a good idea. First, the pip-install step might fail and then users would get a cryptic error message when the setup resumes. Second, some users might be using tools to scan dependencies, and this would be missed. This is something the setup.py file format doesn't address well, and I can't offer a better solution.

As for functionality, there are some small inconsistencies between the interfaces of retworkx and it's inspiration networkx that users might find a little frustrating (though nothing, as far as I can see, that would cause problems). At present, retworkx does not cover the same range of functionality as its competition, though that could hardly be expected (as mentioned above, those libraries are very mature).

I am in agreement that the benchmarks on performance leave something to be desired, though I have not delved into them quite as deep as the other reviewer. I won't reiterate the comments about construction benchmarks, though these do need to be addressed. Instead, I will mention that the benchmarks should cover a wider range of use cases. The graphs used in benchmarking are generally pretty large. I would suggest that a large number of graph library users would primarily work with graphs that are relatively small by comparison (a few hundred or fewer nodes). Your benchmarks don't address this end of the spectrum at all. I would consider adding benchmarks for these cases too, although I expect the performance gap to be considerably closer here.

# Software paper
Generally I don't see any issues in the software paper. The summary and statement of need are clear. The code for the benchmarks should be included in the main repository, so it is more strongly linked to the version of retworkx. I suggest making a folder called "benchmarks" and placing the code in there.
