### 2021-12-14 7bafabf
add ProductNodeMap to tensor product
