### 2021-12-15 1959cf8
Make endpoint ordering of edge tups unspecified.
