+++ tests/graph/test_contract_nodes.py
@@ -14,6 +14,22 @@ import unittest
 
 import retworkx
 
+class UE(tuple):
+    """An edge tuple wrapper for comparing expected edges with actual graph
+    edge lists where endpoint order doesn't matter (undirected). Supports
+    both edges and weighted edges.
+
+    For example, the following become true:
+    ``UE((2, 3)) == UE((3, 2))`` 
+    ``UE((4, 5, "a")) == UE((5, 4, "a"))``
+    
+    """
+    def __eq__(self, o: object) -> bool:
+        return (frozenset(self[:2]), tuple(self[2:])) == (frozenset(o[:2]), tuple(o[2:]))
+
+    def __hash__(self) -> int:
+        return hash((frozenset(self[:2]), tuple(self[2:])))
+
 
 class TestContractNodes(unittest.TestCase):
     def test_empty_nodes(self):
@@ -60,8 +76,8 @@ class TestContractNodes(unittest.TestCase):
 
         self.assertEqual([node_b, node_c, node_m], dag.node_indexes())
         self.assertEqual(
-            {(node_b, node_c), (node_c, node_m), (node_b, node_m)},
-            set(dag.edge_list()),
+            {UE((node_b, node_c)), UE((node_c, node_m)), UE((node_b, node_m))},
+            set(UE(e) for e in dag.edge_list()),
         )
 
     def test_multiple_paths_would_cycle(self):
@@ -98,13 +114,13 @@ class TestContractNodes(unittest.TestCase):
         )
         self.assertEqual(
             {
-                (node_b, node_c),
-                (node_c, node_m),
-                (node_e, node_m),
-                (node_b, node_e),
-                (node_b, node_m),
+                UE((node_b, node_c)),
+                UE((node_c, node_m)),
+                UE((node_e, node_m)),
+                UE((node_b, node_e)),
+                UE((node_b, node_m)),
             },
-            set(dag.edge_list()),
+            set(UE(e) for e in dag.edge_list()),
         )
 
     def test_replace_node_no_neighbors(self):
@@ -137,8 +153,8 @@ class TestContractNodes(unittest.TestCase):
 
         # Note that target is *always* the new node (m).
         self.assertEqual(
-            {(node_a, node_m, 1), (node_a, node_m, 2)},
-            set(dag.weighted_edge_list()),
+            {UE((node_a, node_m, 1)), UE((node_a, node_m, 2))},
+            set(UE(e) for e in dag.weighted_edge_list()),
         )
 
 
