### 2021-12-17 #3968 @mtreinish
>  However, since Python3.10 was release recently, the authors may wish to consider adding a wheel for Python 3.10 too.

We've already setup CI on the main branch for 3.10 support (https://github.com/Qiskit/retworkx/pull/447) and we're going to be publishing wheels for 3.10 in the next releases (which has unfortunately been delayed a bit, we were originally hoping to have it out a month or 2 ago).

> As I mentioned, I had no problems installing the library from wheel. However, I noticed that the setup.py file requires the rust-specific setuptools addon yet this is not listed as a build-time dependency. Rather, the setup.py file attempts to load this addon, and if it fails uses a subprocess call to pip to install the necessary package and then attempts to proceed. I understand why the authors have structured the code in this way, but I don't think this is a good idea. First, the pip-install step might fail and then users would get a cryptic error message when the setup resumes. Second, some users might be using tools to scan dependencies, and this would be missed. This is something the setup.py file format doesn't address well, and I can't offer a better solution.

This is mostly a holdover for users on older versions of pip, we have `setuptools-rust` listed as a build system requirement in the `pyproject.toml`: https://github.com/Qiskit/retworkx/blob/main/pyproject.toml#L2 (which hopefully scan tools look at) which is the only place we can list it and get it installed early enough (`setup_requires` in the `setup.py` is too late). But for users on older pip  versions that didn't fully support PEP 517 this was insufficient as pip wouldn't install the build system requirements and that would result in a cryptic import error. So the compromise to avoid this was to pip install in the `setup.py`. I agree pip installing as part of the setup.py isn't a great approach here and the need for this has mostly disappeared since RHEL/Centos 6 is EoL which is mostly where we got users with ancient pip versions anyway. I think we can probably get away with removing it from the `setup.py` and improving the documentation in the readme to explicitly list upgrading pip or manually installing `setuptools-rust`. I've pushed a PR proposing these changes here: https://github.com/Qiskit/retworkx/pull/533
