### 2021-12-17 ee41df9
Remove optional dependency on pydot
Previously we were relying on the pydot library to call graphviz to
generate visualizations. This was done for simplicity since pydot is all
about providing a python API wrapper around graphviz. However, it does
this by parsing the dot files into it's own internal representation of
the graph description to provide a python api for interacting with it
prior to calling out to graphviz. For retworkx this is unnecessary as we
have a native function for generating dot files and we just need to pass
that directly to graphviz. Additionally, this parsing adds a plac
[…]
