+++ tox.ini
@@ -51,7 +51,7 @@ whitelist_externals=cargo
 commands =
   black --check --diff {posargs} '../retworkx' '../tests'
   flake8 --per-file-ignores='../retworkx/__init__.py:F405,F403' ../setup.py ../retworkx .
-  cargo fmt -- --check
+  cargo fmt --all -- --check
 
 [testenv:docs]
 basepython = python3
