### 2021-12-17 3de84b8
Remove missed references to pydot
This commit removes missed references to pydot in the README and the
unittests.
