### 2021-12-17 78f0d21
Add `contract_nodes` for graphs (#487)
Adds a new method contract_nodes to PyDiGraph and PyGraph to support replacing a set of nodes with an equivalent node.

Resolves #476 

* Start porting Python method from Terra.

* Use internal methods.

* First pass impl, no error handling.

* Pre-refactor.

* Implement contraction validation.

* Run cycle check if enabled.

* Code cleanup.

* Add local force_check_cycle override.

* Unit test check_cycle override.

* Add ascii graph representations.

* Test unknown indices and corresponding fixes.

* Formatting.

* Collapse if.

* Change return statemen
[…]
