### 2021-12-17 34c0632
Document minimum pip version that works with install from sdist
Document that at least pip 19.0.0 is needed to install from sdist.
Version 19.0.0 added pep 517 support:

https://pip.pypa.io/en/stable/news/#v19-0

which is the required feature for pip to parse the pyproject.toml and
properly install setuptools-rust before processing the setup.py.
