+++ tox.ini
@@ -1,6 +1,7 @@
 [tox]
 minversion = 2.1
 envlist = py36, py37, py38, py39, py310, lint
+isolated_build = true
 
 [testenv]
 install_command = pip install -c{toxinidir}/constraints.txt -U {opts} {packages}
