### 2021-12-17 Qiskit/retworkx#509 @IvanIsCoding
For directed graphs, the methods you are looking for are:
* [`PyDiGraph.out_edges`](https://qiskit.org/documentation/retworkx/stubs/retworkx.PyDiGraph.out_edges.html#retworkx.PyDiGraph.out_edges): for outgoing edges
* [`PyDiGraph.in_edges`](https://qiskit.org/documentation/retworkx/stubs/retworkx.PyDiGraph.in_edges.html#retworkx.PyDiGraph.in_edges): for incident edges

To achieve an equivalent behaviour on undirected graphs, the current best way is to:
* Call [`PyGraph.neighbors`](https://qiskit.org/documentation/retworkx/stubs/retworkx.PyGraph.neighbors.html#retworkx.PyGraph.neighbors), as you pointed out
* Then call [`PyGraph.get_edge_data`](https://qiskit.org/documentation/retworkx/stubs/retworkx.PyGraph.get_edge_data.html#retworkx.PyGraph.get_edge_data) with the node indices
