### 2021-12-18 d0af86a
Fix docstring
