### 2021-12-18 3147c3b
Fix docstring
