### 2021-12-18 Qiskit/retworkx#513 @IvanIsCoding
I will argue that retworkx is currently suitable for shortest paths, maximum weight matching, and subgraph isomorphism on the algortihmic side; and manipulation of the dynamic DAGs on the graph representation side. Most of the examples will have to do with analysing graphs that come out of quantum computing applications, but other disciplines with the same subproblems (e.g. manipulation of DAGs for static analysis in Computer Security) can still reuse our code.

[qtcodes](https://github.com/yaleqc/qtcodes) is a major example of non-trivial analysis of networks. They analyse the syndrome graphs that arise in Quantum Error Correction (QEC):
* Using retworkx to calculate and count the number of unweighted shortest paths  as in [1](https://github.com/yaleqc/qtcodes/blob/fbbe82ec255232d5b95486c41d5d10c64ec4ed80/qtcodes/fitters/lattice_decoder.py#L115-L120) [2](https://github.com/yaleqc/qtcodes/blob/fbbe82ec255232d5b95486c41d5d10c64ec4ed80/qtcodes/fitters/lattice_decoder.py#L220-L252)
* Using retworkx to find a minimum-weight perfect matching to generate surface codes as in [3](https://github.com/yaleqc/qtcodes/blob/fbbe82ec255232d5b95486c41d5d10c64ec4ed80/qtcodes/fitters/lattice_decoder.py#L272-L293)
* Using retworkx to visualize syndrome graphs in general as in [4](https://github.com/yaleqc/qtcodes/blob/fbbe82ec255232d5b95486c41d5d10c64ec4ed80/qtcodes/fitters/lattice_decoder.py#L422-L470)

Some uses within Qiskit that are worth sharing as well are:
* Use retworkx to calculate the depth of a quantum circuit (this one is trivial but is the most popular operation) as in [5](https://github.com/Qiskit/qiskit-terra/blob/413ec5e8650437ea700a0039a0e56a1cd79a4d0b/qiskit/dagcircuit/dagcircuit.py#L880-L891)
* Using retworkx to simplify chains of 1-qubit and 2-qubit operators on the DAGs that represent quantum circuits as in [6](https://github.com/Qiskit/qiskit-terra/blob/4cf3491362e682e1f82ad4e9ea492ab79a60a9cb/qiskit/transpiler/passes/optimization/collect_1q_runs.py), [7](https://github.com/Qiskit/qiskit-terra/blob/413ec5e8650437ea700a0039a0e56a1cd79a4d0b/qiskit/dagcircuit/dagcircuit.py#L1624-L1645), [8](https://github.com/Qiskit/qiskit-terra/blob/413ec5e8650437ea700a0039a0e56a1cd79a4d0b/qiskit/dagcircuit/dagcircuit.py#L1663-L1687)
* Using retworkx to implement [Noise-Adaptive Compiler Mappings](https://arxiv.org/pdf/1901.11054.pdf) as in [9](https://github.com/Qiskit/qiskit-terra/blob/2eee56616d50a9e26756f855ef4aa0135920ad78/qiskit/transpiler/passes/layout/noise_adaptive_layout.py)
* Using retworkx and subgraph isomorphism to map quantum circuits to quantum compute architectures as in [10](https://github.com/Qiskit/qiskit-terra/blob/e42e90bbe13c1fe0673d63654fe0682f31ff8acf/qiskit/transpiler/passes/layout/vf2_layout.py)
