### 2021-12-18 44a1598
Adding full rary tree generator (#463)
* added full_rary_tree generator function and tests

* update rary_tree tests

* argument mandatory, missing comma and typo in comments

* fixing pr comments

* lint

* Apply suggestions from code review

* Fix docstring

* Add full_rary_tree to api.rst

* Fix docstring

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
Co-authored-by: Ivan Carvalho <ivancarv@student.ubc.ca>
