### 2021-12-18 3828fb8
switch to `HashSet`
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
