### 2021-12-18 b50b46c
switch to `HashSet` in `weakly_connected_components`
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
