### 2021-12-18 #3968 @szhorvat
As promised, I will respond to some of the specific points that @mtreinish made. Regarding [the reminder](https://github.com/openjournals/joss-reviews/issues/3968#issuecomment-995197715), my review should be considered complete.

----

> But, I think because your critique here is coming from a completely different viewpoint you're inadvertently missing the use cases for which retworkx was developed and the value it's providing.

I think this illustrates well why the statement of need must be improved both in the paper and in the documentation.

> What you're attributing to a lack of good design are just different priorities and design decisions we've made in the development of retworkx.

Let me clarify my concerns about design. A good example is the lack of consistency in the representation of edges, see https://github.com/Qiskit/retworkx/issues/512 for details. There are fundamental problems here that will have long-term consequences if you plan to support multigraphs. I am not talking about priorities, but about whether certain things are possible at all.

For example, can you implement a function that computes a cycle basis _in Python, not Rust_? It seems to me that this is at best very awkward and at worst impossible, partly due to https://github.com/Qiskit/retworkx/issues/509  How about implementing an undirected Dijkstra shortest path finder which returns a sequence of edges (not nodes) whose attributes can later be retrieved?

I am writing this with the assumption that retworkx aims to be suitable for prototyping algorithms. Correct me if this is not the case. (That would bring us back to the statement of need and the description of the library's purpose.)

Regarding current use cases for the library: It would help if some example uses would be presented in the documentation.

> For the paper, part of why we didn't go into too much detail there is partially because of the length limits.

Many of the things that are missing from the paper could be included in the documentation instead.

Consider spending the word limit on what matters. The "Use Cases" section, which takes up half of the paper in page count, does not actually discuss any use cases. It just presents benchmarks, and re-states in words some of the things we already see in the figures. This is of course your choice, but I would not include the benchmarks in the paper at all (you can still refer to them), unless you think that they are important to answer the questions I listed at [the beginning this comment](https://github.com/openjournals/joss-reviews/issues/3968#issuecomment-984648615).

> The other aspect was we were trying to avoid putting any project in a negative light.

I do not think that this is a problem. It should be stated clearly what advances and novel capabilities this new library brings. This is not possible without referring to existing libraries.

> Being able to iteratively create a graph quickly is important in some applications, especially those that are using networks as part of a larger computation for analysis where the graph is dynamic.

Yes, you are right. This is the type of analysis I expected to see in the statement of need. It would be even better if you can mention specific use cases where the graph must be built incrementally (i.e. first collecting edges, and then adding them to the graph all at once, is not an option).
