+++ docs/source/tutorial/index.rst
@@ -1,0 +1,12 @@
+##################
+retworkx Tutorials
+##################
+
+This section contains tutorials on using retworkx.
+
+.. toctree::
+   :maxdepth: 1
+
+   introduction
+   dags
+   betweenness_centrality
