### 2021-12-19 2c25806
Move DistanceMap to its own file
