### 2021-12-19 67432c9
Use DistanceMap in k_shortest_path
