### 2021-12-19 63f410b
Support HashMap in DistanceMap
