### 2021-12-19 68db1a8
Remove type casting
