### 2021-12-19 cf6dff7
Remove unnecessary IndexType
