### 2021-12-19 f6ab452
Improve dfs_edges documentation
