### 2021-12-19 9e4ec08
Remove part of docstring that was supposed to be code
