### 2021-12-19 b5f0451
Fix sphinx syntax error in incident_edges() docstring
