+++ src/graph.rs
@@ -311,8 +311,8 @@ impl PyGraph {
     /// Return the list of edge indices incent to a provided node
     ///
     /// :param int node: The node index to get incident edges from. If
-    /// this node index is not present in the graph this method will
-    /// return an empty list and not error.
+    ///     this node index is not present in the graph this method will
+    ///     return an empty list and not error.
     ///
     /// :returns: A list of the edge indices incident to a node in the graph
     /// :rtype: EdgeIndices
