### 2021-12-19 7e1a0a0
Use centrality as node label in graphviz visualization
