### 2021-12-19 ef01317
Finish first pass at introduction
