### 2021-12-19 d025a66
Fix heavy hex graphviz visualization example
