### 2021-12-19 bdb4ecc
Add edge utility methods to PyGraph and PyDiGraph
There were some gaps in working with edges in the graph class API.
Mainly when working with indices there was no method to acquire the
edge's data payload or endpoints easily. Additionally for the PyGraph
class didn't have a simple method to find incident edges directly
instead a user had to rely on neighbors() and get the edges from that
directly. This commit adds methods to fill these API gaps to make
working with edges a bit easier.
