+++ releasenotes/notes/edge-index-methods-427f7301c720f565.yaml
@@ -1,0 +1,17 @@
+---
+features:
+  - |
+    Added a new method, :meth:`~retworkx.PyGraph.incident_edges`, to the
+    :class:`~retworkx.PyGraph` class. This method returns a list of edge
+    indices for edges incident to a provided node.
+  - |
+    Added a new method, :meth:`~retworkx.PyGraph.get_edge_data_by_index`,
+    to the :class:`~retworkx.PyGraph` and :class:`~retworkx.PyGraph` classes.
+    This method returns the data payload for an edge in the graph from its
+    index.
+  - |
+    Added a new method, :meth:`~retworkx.PyGraph.get_edge_endpoints_by_index`,
+    to the :class:`~retworkx.PyGraph` and :class:`~retworkx.PyGraph` classes.
+    This method returns the edge's endpoint tuple for an edge in the graph from
+    its index.
+
