### 2021-12-19 fa3f719
Update pseudocode
