+++ src/traversal/mod.rs
@@ -35,22 +35,18 @@ use crate::iterators::EdgeList;
 ///
 /// ::
 ///
-///     DFS(G, s)
-///       for each vertex u in V
-///           color[u] := WHITE
-///       end for
-///       STACK := Ø
-///       color[s] := GRAY
-///       PUSH(STACK, s)                           start at vertex s
-///       while (STACK != Ø)
-///           u := POP(STACK)
-///           for each vertex v in Adj[u]          (u,v) is a tree edge
-///               if (color[v] = WHITE)
-///                   color[v] = GRAY
-///                   PUSH(STACK, v)
-///           end for
-///           color[u] := BLACK                    finish vertex u
-///       end while
+///     DFS(G, v) is
+///         let S be a stack
+///         PUSH(S, (v, iterator of G.edges(v)))
+///         while (S != Ø)
+///             let (v, iterator) := LAST(S)
+///             if hasNext(iterator) then
+///                 w := next(iterator)
+///                 if w is not labeled as discovered then
+///                     label w as discovered                   # (u, w) is a tree edge
+///                     PUSH(S, (w, iterator of G.edges(w)))
+///             else
+///                 POP(S)
 ///
 /// :param PyDiGraph graph: The graph to get the DFS edge list from
 /// :param int source: An optional node index to use as the starting node
@@ -80,22 +76,18 @@ fn digraph_dfs_edges(
 ///
 /// ::
 ///
-///     DFS(G, s)
-///       for each vertex u in V
-///           color[u] := WHITE
-///       end for
-///       STACK := Ø
-///       color[s] := GRAY
-///       PUSH(STACK, s)                           start at vertex s
-///       while (STACK != Ø)
-///           u := POP(STACK)
-///           for each vertex v in Adj[u]          (u,v) is a tree edge
-///               if (color[v] = WHITE)
-///                   color[v] = GRAY
-///                   PUSH(STACK, v)
-///           end for
-///           color[u] := BLACK                    finish vertex u
-///       end while
+///     DFS(G, v) is
+///         let S be a stack
+///         PUSH(S, (v, iterator of G.edges(v)))
+///         while (S != Ø)
+///             let (v, iterator) := LAST(S)
+///             if hasNext(iterator) then
+///                 w := next(iterator)
+///                 if w is not labeled as discovered then
+///                     label w as discovered                   # (u, w) is a tree edge
+///                     PUSH(S, (w, iterator of G.edges(w)))
+///             else
+///                 POP(S)
 ///
 /// .. note::
 ///
