+++ src/traversal/mod.rs
@@ -43,7 +43,7 @@ use crate::iterators::EdgeList;
 ///             if hasNext(iterator) then
 ///                 w := next(iterator)
 ///                 if w is not labeled as discovered then
-///                     label w as discovered                   # (u, w) is a tree edge
+///                     label w as discovered                   # (v, w) is a tree edge
 ///                     PUSH(S, (w, iterator of G.edges(w)))
 ///             else
 ///                 POP(S)
@@ -84,7 +84,7 @@ fn digraph_dfs_edges(
 ///             if hasNext(iterator) then
 ///                 w := next(iterator)
 ///                 if w is not labeled as discovered then
-///                     label w as discovered                   # (u, w) is a tree edge
+///                     label w as discovered                   # (v, w) is a tree edge
 ///                     PUSH(S, (w, iterator of G.edges(w)))
 ///             else
 ///                 POP(S)
