### 2021-12-20 7cbf8e2
relax trait bounds in `dfs_edges` in retworkx-core
