### 2021-12-20 6ac1a5d
Fix docstring for new methods
