### 2021-12-20 188c699
Minor corrections in dfs_edge_neighbors docstring (#540)
