+++ retworkx/__init__.py
@@ -778,14 +778,15 @@ def dfs_edges(graph, source=None):
 
         DFS(G, v)
           let S be a stack
-          PUSH(S, (v, iterator of G.edges(v)))
+          label v as discovered
+          PUSH(S, (v, iterator of G.neighbors(v)))
           while (S != Ø)
               let (v, iterator) := LAST(S)
               if hasNext(iterator) then
                   w := next(iterator)
                   if w is not labeled as discovered then
                       label w as discovered                   # (v, w) is a tree edge
-                      PUSH(S, (w, iterator of G.edges(w)))
+                      PUSH(S, (w, iterator of G.neighbors(w)))
               else
                   POP(S)
           end while
