+++ retworkx-core/src/traversal/dfs_edges.rs
@@ -27,14 +27,15 @@ use petgraph::visit::{
 /// ```norust
 /// DFS(G, v)
 ///   let S be a stack
-///   PUSH(S, (v, iterator of G.edges(v)))
+///   label v as discovered
+///   PUSH(S, (v, iterator of G.neighbors(v)))
 ///   while (S != Ø)
 ///       let (v, iterator) := LAST(S)
 ///       if hasNext(iterator) then
 ///           w := next(iterator)
 ///           if w is not labeled as discovered then
 ///               label w as discovered                   # (v, w) is a tree edge
-///               PUSH(S, (w, iterator of G.edges(w)))
+///               PUSH(S, (w, iterator of G.neighbors(w)))
 ///       else
 ///           POP(S)
 ///   end while
