### 2021-12-20 c59e4da
Apply suggestions from code review
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
