+++ src/digraph.rs
@@ -666,7 +666,7 @@ impl PyDiGraph {
         Ok(data)
     }
 
-    /// Return the edge data for the edge by it's given index
+    /// Return the edge data for the edge by its given index
     ///
     /// :param int edge_index: The edge index to get the data for
     ///
@@ -690,7 +690,7 @@ impl PyDiGraph {
         Ok(data)
     }
 
-    /// Return the edge endpoints for the edge by it's given index
+    /// Return the edge endpoints for the edge by its given index
     ///
     /// :param int edge_index: The edge index to get the endpoints for
     ///
