### 2021-12-21 e737ca7
Add all_edges kwarg to to PyDigraph.incident_edges()
This commit adds a flag to the directed graph version of
incident_edges() to enable users to get both the incoming and outgoing
edges if they so desire.
