### 2021-12-22 12f479d
Add incident_edge_index_map method
