+++ src/digraph.rs
@@ -1600,6 +1600,72 @@ impl PyDiGraph {
         }
     }
 
+    /// Return the index map of edges incident to a provided node
+    ///
+    /// By default this method will only return the outgoing edges of
+    /// the provided ``node``. If you would like to access both the
+    /// incoming and outgoing edges you can set the ``all_edges``
+    /// kwarg to ``True``.
+    ///
+    /// :param int node: The node index to get incident edges from. If
+    ///     this node index is not present in the graph this method will
+    ///     return an empty list and not error.
+    /// :param bool all_edges: If set to ``True`` both incoming and outgoing
+    ///     edges to ``node`` will be returned.
+    ///
+    /// :returns: A mapping of incident edge indices to the tuple
+    ///     ``(source, target, data)``
+    /// :rtype: EdgeIndexMap
+    #[pyo3(text_signature = "(self, node, /, all_edges=False)")]
+    #[args(all_edges = "false")]
+    pub fn incident_edge_index_map(
+        &self,
+        py: Python,
+        node: usize,
+        all_edges: bool,
+    ) -> EdgeIndexMap {
+        let node_index = NodeIndex::new(node);
+        if all_edges {
+            EdgeIndexMap {
+                edge_map: self
+                    .graph
+                    .edges_directed(node_index, petgraph::Direction::Outgoing)
+                    .chain(self.graph.edges_directed(
+                        node_index,
+                        petgraph::Direction::Incoming,
+                    ))
+                    .map(|edge| {
+                        (
+                            edge.id().index(),
+                            (
+                                edge.source().index(),
+                                edge.target().index(),
+                                edge.weight().clone_ref(py),
+                            ),
+                        )
+                    })
+                    .collect(),
+            }
+        } else {
+            EdgeIndexMap {
+                edge_map: self
+                    .graph
+                    .edges(node_index)
+                    .map(|edge| {
+                        (
+                            edge.id().index(),
+                            (
+                                edge.source().index(),
+                                edge.target().index(),
+                                edge.weight().clone_ref(py),
+                            ),
+                        )
+                    })
+                    .collect(),
+            }
+        }
+    }
+
     /// Get the index and edge data for all parents of a node.
     ///
     /// This will return a list of tuples with the parent index the node index
