+++ docs/source/tutorial/dags.rst
@@ -125,7 +125,159 @@ Qiskit's Compiler
 
 Another application using directed acyclic graphs is the compiler in
 `Qiskit <https://qiskit.org>`__. Qiskit is an SDK for working with
-quantum computing.
+quantum computing. Qiskit's
+`compiler <https://qiskit.org/documentation/apidoc/transpiler.html>`__
+internally represents a quantum circuit as a
+`directed acyclic graph <https://qiskit.org/documentation/stubs/qiskit.dagcircuit.DAGCircuit.html>`__.
+Retworkx was originally started to accelerate the performance of the Qiskit
+compiler's use of directed acyclic graphs for the compiler.
 
-Retworkx was originally written to accelerate the performance of the Qiskit
-compiler's use of directed acyclic graphs.
+To examine how Qiskit's we first need to look at a quantum circuit. A
+quantum is a computation routine consisting of coherent quantum operations on
+quantum data. It is an ordered squence of quantum gates, measurements and
+resets which may be conditioned on real-time classical computation. A
+quantum circuit is represented graphically like:
+
+.. parsed-literal::
+
+            ┌───┐      ░ ┌─┐
+       q_0: ┤ H ├──■───░─┤M├───
+            └───┘┌─┴─┐ ░ └╥┘┌─┐
+       q_1: ─────┤ X ├─░──╫─┤M├
+                 └───┘ ░  ║ └╥┘
+    meas: 2/══════════════╩══╩═
+                          0  1
+
+The specifics of this circuit aren't important here beyond the fact that
+we have 2 qubits, ``q_0`` and ``q_1``, 2 classical bits, ``c_0`` and ``c_1``,
+and a series of operations on those qubits with a depedency ordering. The last
+operation on each qubit is a measurement on ``q_0`` that is stored in ``c_0``
+and ``q_1`` that is stored in ``c_1``.
+
+We can represent this quantum circuit as a directed acyclic graph like qiskit
+does internally with:
+
+.. jupyter-execute::
+
+    dag = rx.PyDiGraph()
+    # Input nodes:
+    in_nodes = dag.add_nodes_from(["q_0", "q_1", "c_0", "c_1"])
+    # Output nodes
+    out_nodes = dag.add_nodes_from(["q_0", "q_1", "c_0", "c_1"])
+    # Add H gate
+    h_gate = dag.add_child(in_nodes[0], "h", "q_0")
+    # Add CX Gate
+    cx_gate = dag.add_child(h_gate, "cx", "q_0")
+    dag.add_edge(in_nodes[1], cx_gate, "q_1")
+    # Add measure Gates
+    meas_q0 = dag.add_child(cx_gate, "measure", "q_0")
+    meas_q1 = dag.add_child(cx_gate, "measure", "q_1")
+    # Measure q0 instruction edges
+    dag.add_edge(meas_q0, out_nodes[0], "q_0")
+    dag.add_edge(in_nodes[2], meas_q0, "c_0")
+    dag.add_edge(meas_q0, out_nodes[2], "c_0")
+    # Measure q1 instruction edges
+    dag.add_edge(meas_q1, out_nodes[1], "q_1")
+    dag.add_edge(in_nodes[3], meas_q1, "c_1")
+    dag.add_edge(meas_q1, out_nodes[3], "c_1")
+
+    from retworkx.visualization import graphviz_draw
+
+    graphviz_draw(
+        dag,
+        node_attr_fn=lambda node: {"label": str(node)},
+        edge_attr_fn=lambda edge: {"label": str(edge)}
+    )
+
+In this representation of the circuit the flow of data through the bits is
+modeled by edges. The first set of nodes are input nodes and the last set
+are output nodes representing the beginning state and end state of each
[…]
