### 2021-12-23 f7e4421
Update tutorials based on review comments
