### 2021-12-23 97dcf16
Assert endpoint order for PyGraph.incident_edge_index_map()
Previously, the endpoint order for PyGraph.incident_edge_index_map()
wasn't guaranteed. In practice it appears that the source node was
always the user provided node, but that wasn't actually defined
behavior. This commit updates the internal petgraph calls to ensure we
always return the endpoints in a fixed order and adds that to the
documented behavior of the method.
