+++ releasenotes/notes/add-node-indices-3c7c37ed4ea75a2c.yaml
@@ -1,0 +1,11 @@
+---
+features:
+  - |
+    Added a new method, :meth:`~retworkx.PyDiGraph.node_indices`, to the
+    :class:`~retworkx.PyDigraph` and :class:`~retworkx.PyGraph` classes.
+    This method is identical to the previously existing
+    :meth:`~retworkx.PyDiGraph.node_indexes` method but changes the name
+    to be consistent with the use of "indices" throughout the rest of
+    retworkx. The :meth:`~retworkx.PyDiGraph.node_indexes` will likely
+    be deprecated in a future release prior to it's removal in an eventual
+    1.0 release.
