### 2021-12-24 025d90c
Small fixes to the qiskit section of dag tutorial
