### 2022-01-03 2dc3552
Start migration from node_indexes -> node_indices (#541)
Since the earliest retworkx releases the method to return the list of
node indices in the graph was named "node_indexes". However as the
retworkx api evolved we standardized on the use of "indices" everywhere
else in the API. For backwards compatibility we kept the "node_indexes"
method name. This commit starts the process of migrating away from
"node_indexes" a new method "node_indices" is being added. The
"node_indexes" method is now just a backwards compatibility alias for
node_indices. In a future release when we start preparing for 
[…]
