+++ src/digraph.rs
@@ -75,9 +75,9 @@ use super::dag_algo::is_directed_acyclic_graph;
 ///        graph.add_nodes_from(list(range(5)))
 ///        graph.add_nodes_from(list(range(2)))
 ///        graph.remove_node(2)
-///        print("After deletion:", graph.node_indexes())
+///        print("After deletion:", graph.node_indices())
 ///        res_manual = graph.add_parent(6, None, None)
-///        print("After adding a new node:", graph.node_indexes())
+///        print("After adding a new node:", graph.node_indices())
 ///
 /// Additionally, each node and edge contains an arbitrary Python object as a
 /// weight/data payload. You can use the index for access to the data payload
@@ -475,15 +475,30 @@ impl PyDiGraph {
             .collect()
     }
 
-    /// Return a list of all node indexes.
+    /// Return a list of all node indices.
     ///
-    /// :returns: A list of all the node indexes in the graph
+    /// :returns: A list of all the node indices in the graph
     /// :rtype: NodeIndices
     #[pyo3(text_signature = "(self)")]
-    pub fn node_indexes(&self) -> NodeIndices {
+    pub fn node_indices(&self) -> NodeIndices {
         NodeIndices {
             nodes: self.graph.node_indices().map(|node| node.index()).collect(),
         }
+    }
+
+    /// Return a list of all node indices.
+    ///
+    /// .. note::
+    ///
+    ///     This is identical to :meth:`.node_indices()`, which is the
+    ///     preferred method to get the node indices in the graph. This
+    ///     exists for backwards compatibility with earlier releases.
+    ///
+    /// :returns: A list of all the node indices in the graph
+    /// :rtype: NodeIndices
+    #[pyo3(text_signature = "(self)")]
+    pub fn node_indexes(&self) -> NodeIndices {
+        self.node_indices()
     }
 
     /// Return True if there is an edge from node_a to node_b.
@@ -513,11 +528,11 @@ impl PyDiGraph {
             .graph
             .neighbors_directed(index, petgraph::Direction::Outgoing);
         let mut succesors: Vec<&PyObject> = Vec::new();
-        let mut used_indexes: HashSet<NodeIndex> = HashSet::new();
+        let mut used_indices: HashSet<NodeIndex> = HashSet::new();
         for succ in children {
-            if !used_indexes.contains(&succ) {
+            if !used_indices.contains(&succ) {
                 succesors.push(self.graph.node_weight(succ).unwrap());
-                used_indexes.insert(succ);
+                used_indices.insert(succ);
             }
         }
         succesors
@@ -536,11 +551,11 @@ impl PyDiGraph {
             .graph
             .neighbors_directed(index, petgraph::Direction::Incoming);
         let mut predec: Vec<&PyObject> = Vec::new();
-        let mut used_indexes: HashSet<NodeIndex> = HashSet::new();
+        let mut used_indices: HashSet<NodeIndex> = HashSet::new();
         for pred in parents {
-            if !used_indexes.contains(&pred) {
+            if !used_indices.contains(&pred) {
                 predec.push(self.graph.node_weight(pred).unwrap());
-                used_indexes.insert(pred);
+                used_indices.insert(pred);
             }
         }
         predec
@@ -568,7 +583,7 @@ impl PyDiGraph {
     ) -> PyResult<Vec<&PyObject>> {
         let index = NodeIndex::new(node);
         let mut succesors: Vec<&PyObject> = Vec::new();
-        let mut used_indexes: HashSet<NodeIndex> = HashSet::new();
+        let mut used_indices: HashSet<NodeIndex> = HashSet::new();
 
         let filter_edge = |edge: &PyObject| -> PyResult<bool> {
[…]
