+++ paper.bib
@@ -210,17 +210,6 @@
   url = {https://github.com/PyO3/pyo3}
 }
 
-@book{Demetrescu2009,
-  doi = {10.1090/dimacs/074},
-  url = {https://doi.org/10.1090/dimacs/074},
-  year = {2009},
-  month = {jul},
-  publisher = {American Mathematical Society},
-  address = {Providence, RI, USA},
-  editor = {Camil Demetrescu and Andrew V. Goldberg and David Johnson},
-  title = {The Shortest Path Problem: Ninth DIMACS Implementation Challenge}
-}
-
 @Article{Dijkstra1959ANO,
   title={A note on two problems in connexion with graphs},
   author={Edsger W. Dijkstra},
@@ -229,25 +218,6 @@
   volume={1},
   doi={10.1007/BF01386390},
   pages={269-271}
-}
-
-@article{DeSanto2003,
-author = {De Santo, M. and Foggia, P. and Sansone, C. and Vento, M.},
-title = {A large database of graphs and its use for benchmarking graph isomorphism algorithms},
-journal = {Pattern Recogn. Lett.},
-issue_date = {May 2003},
-volume = {24},
-number = {8},
-month = may,
-year = {2003},
-issn = {0167-8655},
-pages = {1067–1079},
-numpages = {13},
-url = {http://dx.doi.org/10.1016/S0167-8655(02)00253-2},
-doi = {10.1016/S0167-8655(02)00253-2},
-acmid = {763457},
-publisher = {Elsevier Science Inc.},
-address = {New York, NY, USA},
 }
 
 @ARTICLE{Cordella2004,
@@ -302,11 +272,46 @@ pages = {1–13},
 numpages = {13}
 }
 
-@techreport{pep427,
-  author  = {Daniel Holth},
-  title   = {The Wheel Binary Package Format 1.0},
-  year    = {2012},
-  type    = {PEP},
-  number  = {427},
-  url     = {https://www.python.org/dev/peps/pep-0427/},
+@article{Staudt2016, 
+  title={NetworKit: A tool suite for large-scale complex network analysis}, 
+  volume={4}, 
+  DOI={10.1017/nws.2016.20}, 
+  number={4}, 
+  journal={Network Science}, 
+  publisher={Cambridge University Press}, 
+  author={STAUDT, CHRISTIAN L. and SAZONOVS, ALEKSEJS and MEYERHENKE, HENNING}, 
+  year={2016}, 
+  pages={508–530}
+}
+
+@inbook{Ji2021,
+author = {Liu, Ji and Bello, Luciano and Zhou, Huiyang},
+title = {Relaxed Peephole Optimization: A Novel Compiler Optimization for Quantum Circuits},
+year = {2021},
+isbn = {9781728186139},
+publisher = {IEEE Press},
+url = {https://doi.org/10.1109/CGO51591.2021.9370310},
[…]
