+++ paper.bib
@@ -40,11 +40,11 @@
                   Christa Zoufal and
                   Donny Greenberg and
                   Stefan Woerner},
-  title        = {Qiskit/qiskit-terra: Qiskit Terra  0.18.3},
+  title        = {Qiskit/qiskit-terra: Qiskit Terra  0.19.1},
   month        = sep,
   year         = 2021,
   publisher    = {Zenodo},
-  version      = {0.18.3},
+  version      = {0.19.1},
   doi          = {10.5281/zenodo.2583252},
   url          = {https://doi.org/10.5281/zenodo.2583252}
 }
@@ -314,4 +314,16 @@ numpages = {14}
       eprint={1811.04968},
       archivePrefix={arXiv},
       primaryClass={quant-ph}
+}
+
+@misc{Stone2021,
+  author = {
+    Josh Stone 
+    and Nicholas D. Matsakis
+    },
+  title = {Rayon: A data parallelism library for Rust},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/rayon-rs/rayon}
 }
