### 2022-01-04 371962d
Remove release notes
