+++ releasenotes/notes/graph-indices-6fe5425b6e2b35d4.yaml
@@ -1,25 +1,0 @@
----
-features:
-  - |
-    Added a new method, :meth:`~retworkx.PyDiGraph.node_indices`, to the
-    :class:`~retworkx.PyDigraph` and :class:`~retworkx.PyGraph` classes.
-    This method is identical to the previously existing
-    :meth:`~retworkx.PyDiGraph.node_indexes` method but changes the name
-    to be consistent with the use of "indices" throughout the rest of
-    retworkx. The :meth:`~retworkx.PyDiGraph.node_indexes` will likely
-    be deprecated in a future release prior to it's removal in an eventual
-    1.0 release.
-  - |
-    Added a new method, :meth:`~retworkx.PyGraph.incident_edges`, to the
-    :class:`~retworkx.PyGraph` class. This method returns a list of edge
-    indices for edges incident to a provided node.
-  - |
-    Added a new method, :meth:`~retworkx.PyGraph.get_edge_data_by_index`,
-    to the :class:`~retworkx.PyGraph` and :class:`~retworkx.PyGraph` classes.
-    This method returns the data payload for an edge in the graph from its
-    index.
-  - |
-    Added a new method, :meth:`~retworkx.PyGraph.get_edge_endpoints_by_index`,
-    to the :class:`~retworkx.PyGraph` and :class:`~retworkx.PyGraph` classes.
-    This method returns the edge's endpoint tuple for an edge in the graph from
-    its index.
