### 2022-01-04 f19833f
Update releasenotes/notes/add-tensor-product-651db371a96d650c.yaml
