### 2022-01-05 b7e388a
Apply suggestions from code review
Co-authored-by: Kevin Krsulich <kevin@krsulich.net>
