+++ paper.bib
@@ -279,7 +279,7 @@ numpages = {13}
   number={4}, 
   journal={Network Science}, 
   publisher={Cambridge University Press}, 
-  author={STAUDT, CHRISTIAN L. and SAZONOVS, ALEKSEJS and MEYERHENKE, HENNING}, 
+  author={Staudt, Christian L. and Sazonovs, Aleksejs and Meyerhenke, Henning}, 
   year={2016}, 
   pages={508–530}
 }
