+++ retworkx/__init__.py
@@ -605,9 +605,9 @@ def _graph_dijkstra_shortest_path(
 
 @functools.singledispatch
 def all_pairs_dijkstra_shortest_paths(graph, edge_cost_fn):
-    """Find the shortest path from all nodes
+    """Find the shortest path from all nodes in a graph
 
-    This function will generate the shortest path from all nodes int the graph
+    This function will generate the shortest path from all nodes in the graph
     using Dijkstra's algorithm. This function is multithreaded and will run
     launch a thread pool with threads equal to the number of CPUs by default.
     You can tune the number of threads with the ``RAYON_NUM_THREADS``
@@ -648,12 +648,12 @@ def _graph_all_pairs_dijkstra_shortest_path(graph, edge_cost_fn):
 
 @functools.singledispatch
 def all_pairs_dijkstra_path_lengths(graph, edge_cost_fn):
-    """Find the shortest path from a node
+    """Calculate the shortest path lengths from all nodes in a graph
 
-    This function will generate the shortest path from a source node using
-    Dijkstra's algorithm. This function is multithreaded and will run
-    launch a thread pool with threads equal to the number of CPUs by default.
-    You can tune the number of threads with the ``RAYON_NUM_THREADS``
+    This function will generate the shortest path lengths from all nodes in the
+    graph using Dijkstra's algorithm. This function is multithreaded and will
+    run launch a thread pool with threads equal to the number of CPUs by
+    default. You can tune the number of threads with the ``RAYON_NUM_THREADS``
     environment variable. For example, setting ``RAYON_NUM_THREADS=4`` would
     limit the thread pool to 4 threads.
 
