+++ paper.bib
@@ -284,19 +284,6 @@ numpages = {13}
   pages={508–530}
 }
 
-@inbook{Ji2021,
-author = {Liu, Ji and Bello, Luciano and Zhou, Huiyang},
-title = {Relaxed Peephole Optimization: A Novel Compiler Optimization for Quantum Circuits},
-year = {2021},
-isbn = {9781728186139},
-publisher = {IEEE Press},
-url = {https://doi.org/10.1109/CGO51591.2021.9370310},
-abstract = {As in classical computing, compilers play an important role in quantum computing. Quantum processors typically support a limited set of primitive operations or quantum gates and have certain hardware-related limitations. A quantum compiler is responsible for adapting a quantum program to these constraint environments and decomposing quantum gates into a sequence of the primitive ones. During the compilation process, it is also critical for the compiler to optimize the quantum circuits in order to reduce the noise in the computation results. Since the noise is introduced by operations and decoherence, reducing the gate count is the key for improving performance.In this paper, we propose a novel quantum compiler optimization, named relaxed peephole optimization (RPO) for quantum computers. RPO leverages the single-qubit state information that can be determined statically by the compiler. We define that a qubit is in a basis state when, at a given point in time, its state is either in the X-, Y-, or Z-basis (|+〉 / |-〉, |L〉 / |R〉 and |0〉 / |1〉). When basis qubits are used as inputs to quantum gates, there exist opportunities for strength reduction, which replaces quantum operations with equivalent but less expensive ones. Compared to the existing peephole optimization for quantum programs, the difference is that our proposed optimization does not require an identical unitary matrix, thereby named 'relaxed' peephole optimization. We also extend our approach to optimize the quantum gates when some input qubits are in known pure states. Both optimizations, namely the Quantum Basis-state Optimization (QBO) and the Quantum Pure-state Optimization (QPO), are implemented in the IBM's Qiskit transpiler. Our experimental results show that our proposed optimization pass is fast and effective. The circuits optimized with our compiler optimizations obtain up to 18.0% (11.7% on average) fewer CNOT gates and up to 8.2% (7.1% on average) lower transpilation time than that of the most aggressive optimization level in the Qiskit compiler. When running on real quantum computers, the success rates of 3-qubit quantum phase estimation algorithm improve by 2.30X due to the reduced gate counts.},
-booktitle = {Proceedings of the 2021 IEEE/ACM International Symposium on Code Generation and Optimization},
-pages = {301–314},
-numpages = {14}
-}
-
 @manual{Sagemath2020,
   Key          = {SageMath},
   Author       = {{The Sage Developers}},
@@ -326,4 +313,13 @@ numpages = {14}
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/rayon-rs/rayon}
+}
+
+@misc{Cross2021,
+      title={OpenQASM 3: A broader and deeper quantum assembly language}, 
+      author={Andrew W. Cross and Ali Javadi-Abhari and Thomas Alexander and Niel de Beaudrap and Lev S. Bishop and Steven Heidel and Colm A. Ryan and John Smolin and Jay M. Gambetta and Blake R. Johnson},
+      year={2021},
+      eprint={2104.14722},
+      archivePrefix={arXiv},
+      primaryClass={quant-ph}
 }
