+++ paper.bib
@@ -210,16 +210,6 @@
   url = {https://github.com/PyO3/pyo3}
 }
 
-@Article{Dijkstra1959ANO,
-  title={A note on two problems in connexion with graphs},
-  author={Edsger W. Dijkstra},
-  journal={Numerische Mathematik},
-  year={1959},
-  volume={1},
-  doi={10.1007/BF01386390},
-  pages={269-271}
-}
-
 @ARTICLE{Cordella2004,
   author={Cordella, L.P. and Foggia, P. and Sansone, C. and Vento, M.},
   journal={IEEE Transactions on Pattern Analysis and Machine Intelligence}, 
@@ -243,33 +233,6 @@ url = {https://www.sciencedirect.com/science/article/pii/S0166218X18300829},
 author = {Alpár Jüttner and Péter Madarasi},
 keywords = {Computational biology, Subgraph isomorphism problem},
 abstract = {This paper presents a largely improved version of the VF2 algorithm for the Subgraph Isomorphism Problem. The improvements are twofold. Firstly, it is based on a new approach for determining the matching order of the nodes, and secondly, more efficient — nevertheless easier to compute — cutting rules are proposed. They together reduce the search space significantly. In addition to the usual Subgraph Isomorphism Problem, the paper also presents specialized algorithms for the Induced Subgraph Isomorphism and for the Graph Isomorphism Problems. Finally, an extensive experimental evaluation is provided using a wide range of inputs, including both real-life biological and chemical datasets and standard randomly generated graph series. The results show major and consistent running time improvements over the other known methods. The C++ implementations of the algorithms are available open-source as part of the LEMON graph and network optimization library.}
-}
-
-@article{Johnson1977,
-author = {Johnson, Donald B.},
-title = {Efficient Algorithms for Shortest Paths in Sparse Networks},
-year = {1977},
-issue_date = {Jan. 1977},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-volume = {24},
-number = {1},
-issn = {0004-5411},
-url = {https://doi.org/10.1145/321992.321993},
-doi = {10.1145/321992.321993},
-abstract = {Algorithms for finding shortest paths are presented which are faster than algorithms
-previously known on networks which are relatively sparse in arcs. Known results which
-the results of this paper extend are surveyed briefly and analyzed. A new implementation
-for priority queues is employed, and a class of “arc set partition” algorithms is
-introduced. For the single source problem on networks with nonnegative arcs a running
-time of O(min(n1+1/k + e, n + e) log n)) is achieved, where there are n nodes and
-e arcs, and k is a fixed integer satisfying k > 0. This bound is O(e) on dense networks.
-For the single source and all pairs problem on unrestricted networks the running time
-is O(min(n2+1/k + ne, n2 log n + ne log n).},
-journal = {J. ACM},
-month = jan,
-pages = {1–13},
-numpages = {13}
 }
 
 @article{Staudt2016, 
@@ -322,4 +285,19 @@ numpages = {13}
       eprint={2104.14722},
       archivePrefix={arXiv},
       primaryClass={quant-ph}
+}
+
+@article{Li2021,
+   title={Qubit Mapping Based on Subgraph Isomorphism and Filtered Depth-Limited Search},
+   volume={70},
+   ISSN={2326-3814},
+   url={http://dx.doi.org/10.1109/TC.2020.3023247},
+   DOI={10.1109/tc.2020.3023247},
+   number={11},
+   journal={IEEE Transactions on Computers},
+   publisher={Institute of Electrical and Electronics Engineers (IEEE)},
+   author={Li, Sanjiang and Zhou, Xiangzhen and Feng, Yuan},
+   year={2021},
+   month={Nov},
+   pages={1777–1788}
 }
