+++ paper.bib
@@ -301,3 +301,12 @@ abstract = {This paper presents a largely improved version of the VF2 algorithm 
    month={Nov},
    pages={1777–1788}
 }
+
+@misc{Bharti2021,
+      title={Noisy intermediate-scale quantum (NISQ) algorithms}, 
+      author={Kishor Bharti and Alba Cervera-Lierta and Thi Ha Kyaw and Tobias Haug and Sumner Alperin-Lea and Abhinav Anand and Matthias Degroote and Hermanni Heimonen and Jakob S. Kottmann and Tim Menke and Wai-Keong Mok and Sukin Sim and Leong-Chuan Kwek and Alán Aspuru-Guzik},
+      year={2021},
+      eprint={2101.08448},
+      archivePrefix={arXiv},
+      primaryClass={quant-ph}
+}
