### 2022-01-08 7b217ae
Fix binomial tree graph generator docs
