+++ docs/source/api.rst
@@ -173,6 +173,7 @@ Generators
     retworkx.generators.grid_graph
     retworkx.generators.directed_grid_graph
     retworkx.generators.binomial_tree_graph
+    retworkx.generators.directed_binomial_tree_graph
     retworkx.generators.hexagonal_lattice_graph
     retworkx.generators.directed_hexagonal_lattice_graph
     retworkx.generators.heavy_square_graph
