### 2022-01-09 dd87f1a
Add image for dag circuit
