### 2022-01-10 2ee0d1b
Add graph connectivity image
