### 2022-01-10 a79a194
Tweak image size
