### 2022-01-10 bcad945
Tweak isomorphism image size
