### 2022-01-10 301053a
Tweak image size
