### 2022-01-10 96abcf6
Yet another image size tweak
