### 2022-01-10 0f2ae8b
More image resizing
