### 2022-01-10 3197f66
Bump numpy from 0.15.0 to 0.15.1
Bumps [numpy](https://github.com/PyO3/rust-numpy) from 0.15.0 to 0.15.1.
- [Release notes](https://github.com/PyO3/rust-numpy/releases)
- [Changelog](https://github.com/PyO3/rust-numpy/blob/main/CHANGELOG.md)
- [Commits](https://github.com/PyO3/rust-numpy/compare/v0.15.0...v0.15.1)

---
updated-dependencies:
- dependency-name: numpy
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
