+++ Cargo.toml
@@ -22,7 +22,7 @@ crate-type = ["cdylib"]
 ahash = { version = "0.7.6", default-features = false }
 petgraph = "0.6.0"
 fixedbitset = "0.4.1"
-numpy = "0.15.0"
+numpy = "0.15.1"
 rand = "0.8"
 rand_pcg = "0.3"
 rayon = "1.5"
