### 2022-01-10 f230431
deduplicate macro and mention petgraph DfsEvent struct
