### 2022-01-11 a4ec9e9
Use Vector in dijkstra_shortest_paths and all_pairs_dijkstra_path_lengths
