+++ src/shortest_path/all_pairs_dijkstra.rs
@@ -80,17 +80,18 @@ pub fn all_pairs_dijkstra_path_lengths<Ty: EdgeType + Sync>(
     let out_map: DictMap<usize, PathLengthMapping> = node_indices
         .into_par_iter()
         .map(|x| {
-            let path_lenghts: PyResult<DictMap<NodeIndex, f64>> =
+            let path_lenghts: PyResult<Vec<Option<f64>>> =
                 dijkstra(graph, x, None, |e| edge_cost(e.id()), None);
             let out_map = PathLengthMapping {
                 path_lengths: path_lenghts
                     .unwrap()
-                    .iter()
-                    .filter_map(|(index, cost)| {
-                        if *index == x {
-                            None
+                    .into_iter()
+                    .enumerate()
+                    .filter_map(|(index, opt_cost)| {
+                        if index != x.index() {
+                            opt_cost.map(|cost| (index, cost))
                         } else {
-                            Some((index.index(), *cost))
+                            None
                         }
                     })
                     .collect(),
