### 2022-01-11 d46f800
add cycle check
