+++ tests/digraph/test_toposort.py
@@ -64,3 +64,19 @@ class TestTopologicalSorter(unittest.TestCase):
         sorter.done([0])
         with self.assertRaises(ValueError):
             sorter.done([0])
+
+    def test_topo_sort_raises_if_graph_has_cycle(self):
+        graph = retworkx.generators.directed_cycle_graph(5)
+        with self.assertRaises(retworkx.DAGHasCycle):
+            _ = retworkx.TopologicalSorter(graph)
+
+    def test_topo_sort_progress_if_graph_has_cycle_and_cycle_check_disabled(self):
+        graph = retworkx.generators.directed_cycle_graph(5)
+        starting_node = graph.add_node('starting node')
+        graph.add_edge(starting_node, 0, 'starting edge')
+
+        sorter = retworkx.TopologicalSorter(graph, check_cycle=False)
+        nodes = sorter.get_ready()
+        self.assertEqual(nodes, [starting_node])
+        sorter.done(nodes)
+        self.assertFalse(sorter.is_active())
