### 2022-01-11 f58617a
an example reproducing the old behavior in release note
