### 2022-01-11 8cc0011
typo in release note code
