### 2022-01-11 9f29268
Add flag for attribute conversion when converting from networkx
