+++ tests/test_converters.py
@@ -18,90 +18,109 @@ import networkx
 class TestNetworkxConverter(unittest.TestCase):
     def test_undirected_gnm_graph(self):
         g = networkx.gnm_random_graph(10, 10, seed=42)
-        out_graph = retworkx.networkx_converter(g)
-        self.assertIsInstance(out_graph, retworkx.PyGraph)
-        self.assertEqual(out_graph.nodes(), list(g.nodes))
-        self.assertEqual(
-            out_graph.weighted_edge_list(), list(g.edges(data=True))
-        )
-        self.assertEqual(out_graph.multigraph, g.is_multigraph())
+        for keep_attributes in [True, False]:
+            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            self.assertIsInstance(out_graph, retworkx.PyGraph)
+            self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
+            self.assertEqual(
+                out_graph.weighted_edge_list(), list(g.edges(data=True))
+            )
+            self.assertEqual(out_graph.multigraph, g.is_multigraph())
 
     def test_directed_gnm_graph(self):
         g = networkx.gnm_random_graph(10, 10, seed=42, directed=True)
-        out_graph = retworkx.networkx_converter(g)
-        self.assertIsInstance(out_graph, retworkx.PyDiGraph)
-        self.assertEqual(out_graph.nodes(), list(g.nodes))
-        self.assertEqual(
-            out_graph.weighted_edge_list(), list(g.edges(data=True))
-        )
-        self.assertEqual(out_graph.multigraph, g.is_multigraph())
+        for keep_attributes in [True, False]:
+            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            self.assertIsInstance(out_graph, retworkx.PyDiGraph)
+            self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
+            self.assertEqual(
+                out_graph.weighted_edge_list(), list(g.edges(data=True))
+            )
+            self.assertEqual(out_graph.multigraph, g.is_multigraph())
 
     def test_empty_graph(self):
         g = networkx.Graph()
-        out_graph = retworkx.networkx_converter(g)
-        self.assertIsInstance(out_graph, retworkx.PyGraph)
-        self.assertEqual(out_graph.nodes(), list(g.nodes))
-        self.assertEqual(
-            out_graph.weighted_edge_list(), list(g.edges(data=True))
-        )
-        self.assertEqual(out_graph.multigraph, g.is_multigraph())
+        for keep_attributes in [True, False]:
+            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            self.assertIsInstance(out_graph, retworkx.PyGraph)
+            self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
+            self.assertEqual(
+                out_graph.weighted_edge_list(), list(g.edges(data=True))
+            )
+            self.assertEqual(out_graph.multigraph, g.is_multigraph())
 
     def test_empty_multigraph(self):
         g = networkx.MultiGraph()
-        out_graph = retworkx.networkx_converter(g)
-        self.assertIsInstance(out_graph, retworkx.PyGraph)
-        self.assertEqual(out_graph.nodes(), list(g.nodes))
-        self.assertEqual(
-            out_graph.weighted_edge_list(), list(g.edges(data=True))
-        )
-        self.assertEqual(out_graph.multigraph, g.is_multigraph())
+        for keep_attributes in [True, False]:
+            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            self.assertIsInstance(out_graph, retworkx.PyGraph)
[…]
