+++ retworkx/__init__.py
@@ -1292,7 +1292,7 @@ def _graph_spring_layout(
     )
 
 
-def networkx_converter(graph):
+def networkx_converter(graph, keep_attributes: bool = False):
     """Convert a networkx graph object into a retworkx graph object.
 
     .. note::
@@ -1303,6 +1303,8 @@ def networkx_converter(graph):
         independently.
 
     :param networkx.Graph graph: The networkx graph to convert.
+    :param bool keep_attributes: If `True`, add networkx attributes to retworkx 
+        object.
 
     :returns: A retworkx graph, either a :class:`~retworkx.PyDiGraph` or a
         :class:`~retworkx.PyGraph` based on whether the input graph is directed
@@ -1321,6 +1323,13 @@ def networkx_converter(graph):
             for x in graph.edges(data=True)
         ]
     )
+
+    if keep_attributes:
+        for node, node_index in node_indices.items():
+            attributes = graph.nodes[node]
+            attributes["__networkx_name__"] = node
+            new_graph[node_index] = attributes
+
     return new_graph
 
 
