+++ tests/test_converters.py
@@ -19,7 +19,9 @@ class TestNetworkxConverter(unittest.TestCase):
     def test_undirected_gnm_graph(self):
         g = networkx.gnm_random_graph(10, 10, seed=42)
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            out_graph = retworkx.networkx_converter(
+                g, keep_attributes=keep_attributes
+            )
             self.assertIsInstance(out_graph, retworkx.PyGraph)
             self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
             self.assertEqual(
@@ -30,7 +32,9 @@ class TestNetworkxConverter(unittest.TestCase):
     def test_directed_gnm_graph(self):
         g = networkx.gnm_random_graph(10, 10, seed=42, directed=True)
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            out_graph = retworkx.networkx_converter(
+                g, keep_attributes=keep_attributes
+            )
             self.assertIsInstance(out_graph, retworkx.PyDiGraph)
             self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
             self.assertEqual(
@@ -41,7 +45,9 @@ class TestNetworkxConverter(unittest.TestCase):
     def test_empty_graph(self):
         g = networkx.Graph()
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            out_graph = retworkx.networkx_converter(
+                g, keep_attributes=keep_attributes
+            )
             self.assertIsInstance(out_graph, retworkx.PyGraph)
             self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
             self.assertEqual(
@@ -52,7 +58,9 @@ class TestNetworkxConverter(unittest.TestCase):
     def test_empty_multigraph(self):
         g = networkx.MultiGraph()
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            out_graph = retworkx.networkx_converter(
+                g, keep_attributes=keep_attributes
+            )
             self.assertIsInstance(out_graph, retworkx.PyGraph)
             self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
             self.assertEqual(
@@ -63,7 +71,9 @@ class TestNetworkxConverter(unittest.TestCase):
     def test_empty_directed_graph(self):
         g = networkx.DiGraph()
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            out_graph = retworkx.networkx_converter(
+                g, keep_attributes=keep_attributes
+            )
             self.assertIsInstance(out_graph, retworkx.PyDiGraph)
             self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
             self.assertEqual(
@@ -74,7 +84,9 @@ class TestNetworkxConverter(unittest.TestCase):
     def test_empty_directed_multigraph(self):
         g = networkx.MultiDiGraph()
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(g, keep_attributes=keep_attributes)
+            out_graph = retworkx.networkx_converter(
+                g, keep_attributes=keep_attributes
+            )
             self.assertIsInstance(out_graph, retworkx.PyDiGraph)
             self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
             self.assertEqual(
[…]
