### 2022-01-11 bd90981
Implement subTest and rename __networkx_name__ for better visibility
- Switch to using subTest to get better test reporting
- Rename __networkx_name__ to __networkx_node__ and update docs

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
