+++ tests/test_converters.py
@@ -19,119 +19,128 @@ class TestNetworkxConverter(unittest.TestCase):
     def test_undirected_gnm_graph(self):
         g = networkx.gnm_random_graph(10, 10, seed=42)
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(
-                g, keep_attributes=keep_attributes
-            )
-            self.assertIsInstance(out_graph, retworkx.PyGraph)
-            self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
-            self.assertEqual(
-                out_graph.weighted_edge_list(), list(g.edges(data=True))
-            )
-            self.assertEqual(out_graph.multigraph, g.is_multigraph())
+            with self.subTest(keep_attributes=keep_attributes):
+                out_graph = retworkx.networkx_converter(
+                    g, keep_attributes=keep_attributes
+                )
+                self.assertIsInstance(out_graph, retworkx.PyGraph)
+                self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
+                self.assertEqual(
+                    out_graph.weighted_edge_list(), list(g.edges(data=True))
+                )
+                self.assertEqual(out_graph.multigraph, g.is_multigraph())
 
     def test_directed_gnm_graph(self):
         g = networkx.gnm_random_graph(10, 10, seed=42, directed=True)
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(
-                g, keep_attributes=keep_attributes
-            )
-            self.assertIsInstance(out_graph, retworkx.PyDiGraph)
-            self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
-            self.assertEqual(
-                out_graph.weighted_edge_list(), list(g.edges(data=True))
-            )
-            self.assertEqual(out_graph.multigraph, g.is_multigraph())
+            with self.subTest(keep_attributes=keep_attributes):
+                out_graph = retworkx.networkx_converter(
+                    g, keep_attributes=keep_attributes
+                )
+                self.assertIsInstance(out_graph, retworkx.PyDiGraph)
+                self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
+                self.assertEqual(
+                    out_graph.weighted_edge_list(), list(g.edges(data=True))
+                )
+                self.assertEqual(out_graph.multigraph, g.is_multigraph())
 
     def test_empty_graph(self):
         g = networkx.Graph()
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(
-                g, keep_attributes=keep_attributes
-            )
-            self.assertIsInstance(out_graph, retworkx.PyGraph)
-            self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
-            self.assertEqual(
-                out_graph.weighted_edge_list(), list(g.edges(data=True))
-            )
-            self.assertEqual(out_graph.multigraph, g.is_multigraph())
+            with self.subTest(keep_attributes=keep_attributes):
+                out_graph = retworkx.networkx_converter(
+                    g, keep_attributes=keep_attributes
+                )
+                self.assertIsInstance(out_graph, retworkx.PyGraph)
+                self.assertEqual(list(out_graph.node_indexes()), list(g.nodes))
+                self.assertEqual(
+                    out_graph.weighted_edge_list(), list(g.edges(data=True))
+                )
+                self.assertEqual(out_graph.multigraph, g.is_multigraph())
 
     def test_empty_multigraph(self):
         g = networkx.MultiGraph()
         for keep_attributes in [True, False]:
-            out_graph = retworkx.networkx_converter(
-                g, keep_attributes=keep_attributes
-            )
[…]
