### 2022-01-12 1cdf76a
Draft bipartition_tree implementation
