+++ src/tree.rs
@@ -11,11 +11,13 @@
 // under the License.
 
 use std::cmp::Ordering;
+use std::collections::VecDeque;
 
 use super::{graph, weight_callable};
 
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
+use pyo3::types::{PyFloat, PyList};
 use pyo3::Python;
 
 use petgraph::prelude::*;
@@ -134,4 +136,148 @@ pub fn minimum_spanning_tree(
     }
 
     Ok(spanning_tree)
+}
+
+/// Find balanced cut edge of the minmum spanning tree of a graph using node
+/// contraction. Assumes that the tree is connected and is a spanning tree.
+///
+/// :param PyGraph graph: Undirected graph
+/// :param weight_fn: A callable object (function, lambda, etc) which
+///     will be passed the edge object and expected to return a ``float``. This
+///     tells retworkx/rust how to extract a numerical weight as a ``float``
+///     for edge object. Some simple examples are::
+///
+///         minimum_spanning_tree(graph, weight_fn: lambda x: 1)
+///
+///     to return a weight of 1 for all edges. Also::
+///
+///         minimum_spanning_tree(graph, weight_fn: float)
+///
+///     to cast the edge object as a float as the weight.
+/// :param float default_weight: If ``weight_fn`` isn't specified this optional
+///     float value will be used for the weight/cost of each edge.
+///
+/// :returns: A set of nodes in one half of the spanning tree
+///
+#[pyfunction]
+#[pyo3(text_signature = "(spanning_tree, pop, target_pop, epsilon)")]
+pub fn balanced_cut_edge(
+    _py: Python,
+    spanning_tree: &graph::PyGraph,
+    py_pops: &PyList,
+    py_pop_target: &PyFloat,
+    py_epsilon: &PyFloat,
+) -> PyResult<Vec<(usize, Vec<usize>)>> {
+    let epsilon = py_epsilon.value();
+    let pop_target = py_pop_target.value();
+
+    let mut pops: Vec<f64> = vec![]; // not sure if the conversions are needed
+    for i in 0..py_pops.len() {
+        pops.push(py_pops.get_item(i).unwrap().extract::<f64>().unwrap());
+    }
+
+    let mut node_queue: VecDeque<NodeIndex> = VecDeque::<NodeIndex>::new();
+    for leaf_node in spanning_tree.graph.node_indices() {
+        // todo: filter expr
+        if spanning_tree.graph.neighbors(leaf_node).count() == 1 {
+            node_queue.push_back(leaf_node);
+        }
+    }
+
+    // eprintln!("leaf nodes: {}", node_queue.len());
+
+    // this process can be multithreaded, if the locking overhead isn't too high
+    // (note: locking may not even be needed given the invariants this is assumed to maintain)
+    let mut balanced_nodes: Vec<(usize, Vec<usize>)> = vec![];
+    let mut same_partition_tracker: Vec<Vec<usize>> =
+        vec![vec![]; spanning_tree.graph.node_count()]; // keeps track of all all the nodes on the same side of the partition
+    let mut seen_nodes: Vec<bool> =
+        vec![false; spanning_tree.graph.node_count()]; // todo: perf test this
+    while node_queue.len() > 0 {
+        let node = node_queue.pop_front().unwrap();
+        let pop = pops[node.index()];
+
+        // todo: factor out expensive clones
+        // Mark as seen; push to queue if only one unseen neighbor
+        let unseen_neighbors: Vec<NodeIndex> = spanning_tree
+            .graph
+            .neighbors(node)
+            .filter(|node| !seen_nodes[node.index()])
+            .collect();
+        if unseen_neighbors.len() == 1 {
+            // this may be false if root
[…]
