### 2022-01-12 078b6c1
Fix algorithm
