### 2022-01-12 224baf1
Deduplicate set combination in DSU function
