### 2022-01-12 8338916
Switch find_set() to an iterative implementation
