### 2022-01-12 460284d
Remove duplicate line in networkx_coverter docstring
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
