### 2022-01-12 20281b0
Working bipartition tree impl
