### 2022-01-12 efa6b7a
Add flag for attribute conversion when converting from networkx (#548)
Currently, converting from a native networkx Graph to a retworkx PyGraph only preserves
adjacency. This PR adds an optional (and by default disabled) flag to copy over the node
attributes, thereby allowing for a more faithful conversion/translation.

* Add flag for attribute conversion when converting from networkx

* Lint with black

* Implement subTest and rename __networkx_name__ for better visibility

- Switch to using subTest to get better test reporting
- Rename __networkx_name__ to __networkx_node__ and update docs


[…]
