### 2022-01-12 a9124b4
Apply suggestions from code review
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
