+++ src/shortest_path/mod.rs
@@ -299,7 +299,7 @@ pub fn digraph_dijkstra_shortest_path_lengths(
 ///
 /// This function will calculate the shortest path lengths from all nodes in the
 /// graph using Dijkstra's algorithm. This function is multithreaded and will
-/// run launch a thread pool with threads equal to the number of CPUs by
+/// launch a thread pool with threads equal to the number of CPUs by
 /// default. You can tune the number of threads with the ``RAYON_NUM_THREADS``
 /// environment variable. For example, setting ``RAYON_NUM_THREADS=4`` would
 /// limit the thread pool to 4 threads.
