### 2022-01-13 f4bd683
Update releasenotes/notes/average-path-length-disconnected-56b118a964af085a.yaml
