### 2022-01-13 eaf7f75
Move and update new release notes
