### 2022-01-13 25f10b2
Apply suggestions from code review
Co-authored-by: Kevin Hartman <kevin@hart.mn>
