### 2022-01-13 6be2f1e
Add footnote defining a trail for dag tutorial
