### 2022-01-13 f85eb52
Release GIL during most of balanced_edge finding code
