### 2022-01-13 1dc199a
Update distancemap.rs
