### 2022-01-14 e8a49a1
Rephrase instructions in Qiskit example
