### 2022-01-14 3a5391f
Rephrase limitations section
