+++ paper.md
@@ -88,7 +88,7 @@ _retworkx_ implements both VF2 [@Cordella2004] and VF2++ [@Juttner2018] algorith
 
 # Limitations
 
-_retworkx_ is not a drop-in replacement for _NetworkX_. Existing code using _NetworkX_ needs to be modified to use _retworkx_. The authors are aware of the limitation and provide a _NetworkX_ to _retworkx_ conversion guide on the documentation.
+_retworkx_ is not a drop-in replacement for _NetworkX_. Existing code using _NetworkX_ needs to be modified to use _retworkx_, which may be a possible limitation for some users. The authors provide a _NetworkX_ to _retworkx_ conversion guide on the documentation to aid in those situations.
 
 # Acknowledgements
 
