### 2022-01-14 3ee7921
deduplicate `try_control` macro
