### 2022-01-14 29fd43c
raise if negative weight
