+++ releasenotes/notes/dijkstra-search-2d1899241f5166ea.yaml
@@ -32,5 +32,5 @@ features:
       graph = retworkx.PyGraph()
       graph.extend_from_weighted_edge_list([(1, 3, 1), (0, 1, 10), (2, 1, 1), (0, 2, 1)])
       vis = DijkstraTreeEdgesRecorder()
-      retworkx.graph_dijkstra_search(self.graph, [0], float, vis)
+      retworkx.graph_dijkstra_search(graph, [0], float, vis)
       print('Tree edges:', vis.edges)
