### 2022-01-14 2102410
Update retworkx-core/src/distancemap.rs
Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>
