### 2022-01-14 0ba99cc
more minor doc fixes
