### 2022-01-14 2bb0e94
Add dijkstra-search (#489)
* Add dijkstra-search

This commit implements a `dijkstra_search` that traverses a graph
using Dijkstra algorithm and provides the ability to insert callback
functions at specified event points. Python users should subclass
`retworkx.visit.DijkstraVisitor` and overwrite the appropriate callback functions.

* remove unnecessary test file

* revert changes in `dijkstra.rs`

* cargo fmt

* deduplicate `try_control` macro

* raise if negative weight

* fix release note

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
