### 2022-01-14 ed7075c
Apply suggestions from code review
Co-authored-by: Kevin Hartman <kevin@hart.mn>
