### 2022-01-14 8d86789
Remove unused declarations
