+++ src/shortest_path/average_length.rs
@@ -15,7 +15,7 @@ use hashbrown::HashSet;
 use petgraph::prelude::*;
 use petgraph::EdgeType;
 
-use rayon::prelude::*;
+// use rayon::prelude::*; Apparently not used, probably because the following is used instead
 use rayon_cond::CondIterator;
 
 use crate::StablePyGraph;
