+++ retworkx-core/src/distancemap.rs
@@ -27,11 +27,11 @@ use hashbrown::HashMap;
 pub trait DistanceMap<K, V> {
     /// Create mapping with support for items between 0 and `num_elements - 1`.
     fn build(num_elements: usize) -> Self;
-    
+
     /// Get the distance to the item at `pos`. If the distance does not exist,
     /// the function returns `None`.
     fn get_item(&self, pos: K) -> Option<&V>;
-    
+
     /// Insert item at position `pos` with distance `V`.
     fn put_item(&mut self, pos: K, val: V);
 }
