### 2022-01-14 5e27c1a
Apply suggestions from code review
Co-authored-by: Kevin Hartman <kevin@hart.mn>
