+++ docs/source/tutorial/introduction.rst
@@ -5,7 +5,7 @@ Introduction to retworkx
 This guide serves as an introduction to working with retworkx. If you're a
 current or past `NetworkX <https:://networkx.org>`__ user who is looking at
 using retworkx as a replacement for NetworkX, you can also refer to
-:ref:`networkx` for a detailed guide comparing using retworkx with NetworkX.
+:ref:`networkx` for a detailed comparison.
 
 Creating a Graph
 ================
@@ -18,22 +18,22 @@ can run the following code:
     import retworkx as rx
     G = rx.PyGraph()
 
-A :class:`~retworkx.PyGraph` is an collection of nodes (vertices)
-along with unordered pairs of nodes (called edges, links, etc). Nodes and
-edges in retworkx are assigned any Python object (e.g. a numeric value, a
-string, an image, an xml object, another graph, a custom node object, etc)
-for a data payload (also referred to as a weight in the API and
-documentation). Nodes and edges are uniquely identified by an integer ``index``
+A :class:`~retworkx.PyGraph` is comprised of nodes (vertices)
+and unordered pairings of nodes (called edges, links, etc). Both nodes and
+edges in retworkx have an assigned data payload (also referred to as a weight
+in the API and documentation) which can be any Python object (e.g. a numeric
+value, a string, an image, an XML object, another graph, a custom node object,
+ etc.). Nodes and edges are uniquely identified by an integer ``index``
 which is stable for the lifetime of that node or edge in the graph. These
-indices are not guaranteed to be a contiguous sequence as removed nodes or
-edges will leave holes in the set of indices and an index can be reused after
-a removal.
+indices are not guaranteed to be contiguous as removed nodes or
+edges will leave holes in the sequence of assigned indices, and an index
+can be reused after a removal.
 
 Nodes
 =====
 
-The graph :math:`G` can be grown in several ways, there are graph generator
-functions and functions to read and write graphs in different foramts.
+The graph :math:`G` can be grown in several ways. There are also graph generator
+functions and functions to read and write graphs in different formats.
 
 To get started we can add a single node:
 
@@ -56,17 +56,15 @@ elements such as a
     indices = G.add_nodes_from(range(5))
     print(indices)
 
-Just as with :meth:`~retworkx.PyGraph.add_node` the
+Just as with :meth:`~retworkx.PyGraph.add_node`, the
 :meth:`~retworkx.PyGraph.add_nodes` method returns the indices for the nodes
 added as a :class:`~retworkx.NodeIndices` object, which is a custom sequence
 type that contains the index of each node in the order it's added from the input
 sequence.
 
-In the above cases we were adding nodes with a data payload integers (``1`` in
-the single :meth:`~retworkx.PyGraph.add_node` case and the sequence ``0``,
-``1``, ``2``, ``3``, ``4`` in the :meth:`~retworkx.PyGraph.add_nodes_from` case)
-however retworkx doesn't place constraints on what the node data payload can
-be so you can use more involved objects including types which are not
+In the above cases, we were adding nodes with a data payload of type integer (e.g. ``G.add_node(1)``).
+However, retworkx doesn't place constraints on what the node data payload can
+be, so you can use more involved objects including types which are not
 `hashable <https://docs.python.org/3/glossary.html#term-hashable>`__. For
 example, we can add a node with a data payload that's a a
 `dict <https://docs.python.org/3/library/stdtypes.html#dict>`__:
@@ -90,18 +88,18 @@ The graph :math:`G` can also be grown by adding one edge at a time
 
     G.add_edge(1, 2, None)
 
-This will add an edge between node index ``1`` and node index ``2`` with the
[…]
