### 2022-01-14 0b542e9
Use `Vector` for intermediate computations in Dijkstra (#493)
* Sort Dijkstra output at the end

* Handle dense cases at the end

* Change condition for sorting output

* Use Vector for intermediate calculations

* Use vector in k_shortest_path

* Address clippy comments

* Fix bug

* Incorporate feedback from PR

* Add quaternary heap

* Fix steiner tree test

* Avoid creating duplicated dictmaps

* Revert "Fix steiner tree test"

This reverts commit 4a17a258ad560f28b702c7aae6ae6437ad4a0b54.

* Add tests and docstring

* Use trait to reduce duplication

* Move DistanceMap to its own file

* U
[…]
