### 2022-01-15 79b3d94
Ensure that unused vars get gc'ed on each loop
