### 2022-01-15 7fd601b
use iterative dfs-search implementation
