+++ src/connectivity/mod.rs
@@ -730,12 +730,11 @@ pub fn digraph_core_number(
 ///
 ///     The function implicitly assumes that there are no parallel edges
 ///     or self loops. It may produce incorrect/unexpected results if the
-///     input graph has self loops or parallel edges. It's also a recursive
-///     implementation and might run out of memory in large graphs.
+///     input graph has self loops or parallel edges.
 ///
 /// :param PyGraph: The undirected graph to be used.
 ///
-/// :returns: A set with node indexes of the articulation points in the graph.
+/// :returns: A set with node indices of the articulation points in the graph.
 /// :rtype: set
 #[pyfunction]
 #[pyo3(text_signature = "(graph, /)")]
@@ -759,8 +758,7 @@ pub fn articulation_points(graph: &graph::PyGraph) -> HashSet<usize> {
 ///
 ///     The function implicitly assumes that there are no parallel edges
 ///     or self loops. It may produce incorrect/unexpected results if the
-///     input graph has self loops or parallel edges. It's also a recursive
-///     implementation and might run out of memory in large graphs.
+///     input graph has self loops or parallel edges.
 ///
 /// :param PyGraph: The undirected graph to be used.
 ///
