### 2022-01-15 6fcf79d
use a custom return type in biconnected components
