+++ src/lib.rs
@@ -432,6 +432,7 @@ fn retworkx(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_class::<iterators::NodesCountMapping>()?;
     m.add_class::<iterators::NodeMap>()?;
     m.add_class::<iterators::ProductNodeMap>()?;
+    m.add_class::<iterators::BiconnectedComponents>()?;
     m.add_wrapped(wrap_pymodule!(generators))?;
     Ok(())
 }
