### 2022-01-17 524c2a6
Bump hashbrown from 0.11.2 to 0.12.0
Bumps [hashbrown](https://github.com/rust-lang/hashbrown) from 0.11.2 to 0.12.0.
- [Release notes](https://github.com/rust-lang/hashbrown/releases)
- [Changelog](https://github.com/rust-lang/hashbrown/blob/master/CHANGELOG.md)
- [Commits](https://github.com/rust-lang/hashbrown/compare/v0.11.2...v0.12.0)

---
updated-dependencies:
- dependency-name: hashbrown
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
