### 2022-01-18 2b171f0
minor changes in tests
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
