### 2022-01-18 4f127b9
Cleanup all pairs docstrings
