### 2022-01-18 399dc82
Add more details to custom return type docstrings
