### 2022-01-18 df4e02a
Algorithms for biconnected graphs. (#445)
* Algorithms for biconnected graphs.

This commit adds two new functions:

1. `retworkx.articulation_points`. An articulation point is any node whose
removal increases the number of connected components of a graph.
2. `retworkx.biconnected_components`. A  biconnected component is a
maximal subgraph such that the removal of a node will not disconnect
the subgraph.

* fix clippy warnings

* use petgraph traits as part of the effort to have generic functions in retworkx

* replace `search` with a `std::iter` build-in func

* move generic function to retwo
[…]
