### 2022-01-18 9e62dfd
Update docs/source/tutorial/betweenness_centrality.rst
Co-authored-by: Kevin Hartman <kevin@hart.mn>
