### 2022-01-18 7f1e744
Documentation fixes and improvements (#544)
* Documentation fixes and improvements

This commit goes through the documentation for retworkx and fixes
various typos and mistakes. It also updates the generated api docs to be
put in an 'apiref' path instead of 'stubs' to clean up the urls a bit.
To maintain compatibility with existing links though a redirect is added
to redirect from stubs/* to apiref/*.

Fixes #507

* Fix binomial tree graph generator docs

* Apply suggestions from code review

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>

* Apply suggestions fro
[…]
