+++ docs/source/requirements.txt
@@ -7,3 +7,4 @@ pillow>=4.2.1
 reno>=3.4.0
 qiskit-sphinx-theme>=1.7
 matplotlib>=3.4
+sphinx-reredirects
