### 2022-01-18 d45f757
Start editing and fixing release notes
