### 2022-01-19 523da87
Fix code example for barbell graph release note
