### 2022-01-19 56ac4e3
Fix issues in the remaining release notes
