+++ releasenotes/notes/0.11/faster-steiner-tree-0066d7cd2df1b28d.yaml
@@ -1,6 +1,8 @@
 ---
 features:
   - |
-    Implements a faster algorithm that achieves the same approximation
-    ratio for the :func:`~retworkx.steiner_tree` problem based on
-    https://doi.org/10.1016/0020-0190(88)90066-X
+    The algorithm used for the :func:`~retworkx.steiner_tree` implementation
+    has been replaced by a faster algorithm based on:
+    https://doi.org/10.1016/0020-0190(88)90066-X This new implementation
+    achieves the same approximation ratio as the algorithm used previously,
+    so there should be no change in functionality.
