+++ releasenotes/notes/0.11/bugfix-node-holes-210d1b4d329df6dc.yaml
@@ -1,6 +1,6 @@
 ---
 fixes:
   - |
-    Fixes a bug where :func:`~retworkx.distance_matrix` and
-    :func:`~retworkx.k_shortest_path` would panic if the input graph
-    contains holes in node indices.
+    Fixed an issue where :func:`~retworkx.distance_matrix` and
+    :func:`~retworkx.k_shortest_path_lengths` would previously panic if the
+    input graph contains holes in the node indices.
