+++ releasenotes/notes/0.11/negative-weights-in-shortest-path-algos-b205b63bd12d1a26.yaml
@@ -1,11 +1,12 @@
 ---
 fixes:
   - |
-    Fixes an oversight in shortest path algorithms like:
+    Fixed an oversight in shortest path algorithms functions, such as:
     :func:`~retworkx.dijkstra_shortest_paths`,
     :func:`~retworkx.dijkstra_shortest_path_lengths`,
     :func:`~retworkx.all_pairs_dijkstra_path_lengths`,
     :func:`~retworkx.all_pairs_dijkstra_shortest_paths`,
-    :func:`~retworkx.astar_shortest_path`,
-    :func:`~retworkx.k_shortest_path_lengths`, that would previously
-    accept edge weights with negative or NaN values.
+    :func:`~retworkx.astar_shortest_path`, and
+    :func:`~retworkx.k_shortest_path_lengths` which would previously incorrectly
+    accept edge weights with negative or ``NaN`` values.
+    Fixed `#525 <https://github.com/Qiskit/retworkx/issues/525>`__
