### 2022-01-19 04a477d
Add note about api stability in retworkx-core to release note
