### 2022-01-19 118b13e
Update docs/source/tutorial/introduction.rst
This changes the layout in the petersen graph example in
the introduction tutorial to be a bit cleaner.

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
