### 2022-01-19 b3fe6f0
Adjust NetworkX capitalization
