### 2022-01-19 257ea54
Add tutorials to documentation (#537)
* Add tutorials to documentation

This commit starts the process of adding tutorials to the documentation.
Tutorials are guides on how you use retworkx to accomplish a specific
task or explain a concept in graph theory. In this commit 3 tutorials
are added to start:

 * An introduction which serves as a both an intro to graph theory
   concepts and on how to interact with retworkx
 * An introduction to directed acyclic graphs and their applications
 * A guide on how to calculate the betweenness centrality and visualize
   it

The expectation is that once w
[…]
