+++ docs/source/api.rst
@@ -14,6 +14,8 @@ Graph Classes
     retworkx.PyDiGraph
     retworkx.PyDAG
 
+.. _algorithm_api:
+
 Algorithm Functions
 ===================
 
@@ -145,6 +147,18 @@ Connectivity and Cycles
    retworkx.biconnected_components
    retworkx.chain_decomposition
 
+.. _graph-ops:
+
+Graph Operations
+----------------
+
+.. autosummary::
+   :toctree: stubs
+
+   retworkx.complement
+   retworkx.union
+   retworkx.cartesian_product
+
 .. _other-algorithms:
 
 Other Algorithm Functions
@@ -153,15 +167,14 @@ Other Algorithm Functions
 .. autosummary::
    :toctree: apiref
 
-   retworkx.complement
    retworkx.adjacency_matrix
    retworkx.all_simple_paths
    retworkx.transitivity
    retworkx.core_number
    retworkx.graph_greedy_color
-   retworkx.union
    retworkx.metric_closure
-   retworkx.cartesian_product
+
+.. _generator_funcs:
 
 Generators
 ==========
@@ -191,6 +204,8 @@ Generators
     retworkx.generators.generalized_petersen_graph
     retworkx.generators.barbell_graph
     retworkx.generators.full_rary_tree
+
+.. _random_generators:
 
 Random Graph Generator Functions
 ================================
