### 2022-01-19 32786b2
Apply suggestions from code review
Co-authored-by: Jake Lishman <jake@binhbar.com>
