+++ releasenotes/notes/0.11/average-path-length-disconnected-56b118a964af085a.yaml
@@ -9,13 +9,13 @@ upgrade:
   - |
     The default behavior for how the
     :func:`~retworkx.unweighted_average_shortest_path_length` function handles
-    disconnected graphs has been changed Previously, disconnected pairs of nodes
+    disconnected graphs has been changed. Previously, disconnected pairs of nodes
     was assumed to have zero distance which is arguably incorrect/unexpected
     behavior. To make this more consistent with user expectations this has been
     changed to an infinite value. In addition, an extra kwarg ``disconnected``
     was added where, if set to ``True``, the average is taken only over connected
     node pairs. By default, it's set to ``False``. If the previous behavior of
-    treating disconnected pairs of nodes as have a distance of zero is desired,
+    treating disconnected pairs of nodes as having a distance of zero is desired,
     it can be reproduced using the rest of retworkx API like:
 
     .. code-block:: python
