+++ releasenotes/notes/0.11/add-cartesian-product-f57b4aeb8cbebee7.yaml
@@ -1,9 +1,9 @@
 ---
 features:
   - |
-    Added a new function, :func:`~retworkx.cartesian_product` (and it's per
+    Added a new function, :func:`~retworkx.cartesian_product` (and its per
     type variants func:`~retworkx.digraph_cartesian_product` and
-    :func:`~retworkx.graph_cartesian_product`), which calculates the cartesian
+    :func:`~retworkx.graph_cartesian_product`), which calculates the Cartesian
     product of two graphs. For example:
     
     .. jupyter-execute::
