+++ releasenotes/notes/0.11/barbell-graph-generator-5cc9df7148b1a09e.yaml
@@ -2,7 +2,7 @@
 features:
   - |
     Added a new generator function, :func:`~retworkx.generators.barbell_graph`,
-    to the :mod:`retworkx.generators` module that will generate a barbell graph.
+    to the ``retworkx.generators`` module that will generate a barbell graph.
     For example:
 
     .. jupyter-execute::
