+++ releasenotes/notes/0.11/fix-dispatch-3596ef110cc68338.yaml
@@ -4,7 +4,7 @@ fixes:
     Fixed the :func:`~retworkx.dfs_edges` function to have the ``source``
     argument default to ``None`` as was documented. Previously, the ``source``
     argument was incorrectly required and had no default value.
-    Fixed `#515 <https://github.com/Qiskit/retworkx/issues/515>`__
+    Fixed `#515 <https://github.com/Qiskit/retworkx/issues/515>`__.
   - |
     Fixed an oversight in the :func:`~retworkx.union` function where
     user-defined values for the ``merge_nodes`` and ``merge_edges`` arguments
