### 2022-01-19 c4ddfb5
Update releasenotes/notes/0.11/add-cartesian-product-f57b4aeb8cbebee7.yaml
