### 2022-01-19 b04ccf6
Apply suggestions from code review
Co-authored-by: Kevin Hartman <kevin@hart.mn>
