### 2022-01-19 39a26a0
Prepare 0.11.0 release (#549)
* Prepare 0.11.0 release

This commit prepares the repository for the 0.11.0 release. It makes 2
primary changes it moves the release notes for this release to a
separate directory to separate the release notes from future releases.
The second change is a adding a release note for the prelude providing
the highlights for the release.

* Fix doc build

* Move and update new release notes

* Move new release notes

* Start editing and fixing release notes

* Fix code example for barbell graph release note

* Fix issues in the remaining release notes

* Add note abou
[…]
