+++ releasenotes/notes/average-path-length-disconnected-56b118a964af085a.yaml
@@ -1,19 +1,0 @@
----
-upgrade:
-  - |
-    Updates the way :func:`~unweighted_average_shortest_path_length` handles
-    disconnected graphs. Previously, disconnected pairs of nodes was assumed
-    to have zero distance and this has been changed to an infinite value.
-    In addition, an extra kwarg ``disconnected`` was added where, if set to
-    ``True``, the average is taken only over connected node pairs. By default,
-    it's set to ``False``. If the old behavior is desired, it can be reproduced
-    using the rest of retworkx API like:
-
-    .. code-block:: python
-
-      import retworkx
-
-      graph = retworkx.undirected_gnm_random_graph(20, 10, seed=42)
-      n = len(graph)
-      d_mat = retworkx.distance_matrix(graph, null_value=0.0)
-      avg_shortest_path = d_mat.sum() / (n * (n - 1.0))
