### 2022-01-20 86cbb3e
Update and unify cibuildwheel configuration and version
This commit updates the cibuildwheel configuration and version to the
latest release, 2.3.1. At the same time the configuration for
cibuildwheel is unified and simplified by leveraging the pyproject.toml
to store common cross-job options. Job specific options are still set as
environment variables in the job configuration (such as for cross build
settings). This should make the configuration easier to manage moving
forward as we only have to update a single location to adjust for global
setting changes.
