### 2022-01-20 4b5941f
Fix toml syntax error in pyproject.toml
