### 2022-01-20 20c0fb4
Replace most occurrences of "indexes" with "indices" (#557)
(cherry picked from commit b50578374346312aa147ff061f07baf676600304)
