+++ docs/source/tutorial/introduction.rst
@@ -1,5 +1,5 @@
 .. This document was adapted from and heavily based on the similar
-   introductutorial in  NetworkX's documentation which can be found here:
+   introduction tutorial in  NetworkX's documentation which can be found here:
    https://networkx.org/documentation/networkx-2.6.2/tutorial.html
 
 ########################
