### 2022-01-21 b26d3e2
Expand on platform support levels
Following the similar change in Qiskit that defined the levels of
platform support (see:
https://qiskit.org/documentation/getting_started.html#platform-support),
this commit does a similar process for retworkx. This should make it a
bit clearer for users the level of support and expectations when running
retworkx on different platforms.
