### 2022-02-02 #3968 @drvinceknight
Apologies for the delay all.

@inakleinbottle I am not sure why you are unable to tick the boxes. Could you confirm that your review is finished and you're happy for the paper to be accepted? (No problem if not, just clarifying for my benefit)

> Regarding https://github.com/openjournals/joss-reviews/issues/3968#issuecomment-995197715, my review should be considered complete.

@szhorvat likewise @szhorvat, are you happy for the paper to be accepted? (No problem if not, just clarifying for my benefit)
