### 2022-02-02 feed732
Set max line length to 100 characters
Previously we had been setting rustfmt and black to use an 80 character
line length. This was primarily because my personal dev workflow is
setup around 80 character-wide windows. However, as the project has a
growing number of contributors my personal archaic taste in dev workflow
should not be subjected upon everyone. This commit bumps the max line
length in rustfmt and black up to 100 which is a more common default
that will work better in most editors and not needlessly wrap lines
everywhere. Most of this diff is just the auto-formatting adjustments
no
[…]
