+++ src/tree.rs
@@ -62,8 +62,7 @@ pub fn minimum_spanning_edges(
     let mut edge_list: Vec<(f64, EdgeReference<PyObject>)> =
         Vec::with_capacity(graph.graph.edge_count());
     for edge in graph.graph.edge_references() {
-        let weight =
-            weight_callable(py, &weight_fn, edge.weight(), default_weight)?;
+        let weight = weight_callable(py, &weight_fn, edge.weight(), default_weight)?;
         if weight.is_nan() {
             return Err(PyValueError::new_err("NaN found as an edge weight"));
         }
