+++ docs/source/conf.py
@@ -24,9 +24,9 @@ copyright = u'2021, retworkx Contributors'
 
 
 # The short X.Y version.
-version = '0.11.0'
+version = '0.12.0'
 # The full version, including alpha/beta/rc tags.
-release = '0.11.0'
+release = '0.12.0'
 
 extensions = ['sphinx.ext.autodoc',
               'sphinx.ext.autosummary',
