+++ src/steiner_tree.rs
@@ -83,23 +83,16 @@ fn _metric_closure_edges(
     }
     let mut out_vec = Vec::with_capacity(node_count * (node_count - 1) / 2);
     let mut distances = HashMap::with_capacity(graph.graph.node_count());
-    let paths = all_pairs_dijkstra_shortest_paths(
-        py,
-        &graph.graph,
-        weight_fn,
-        Some(&mut distances),
-    )?
-    .paths;
-    let mut nodes: HashSet<usize> =
-        graph.graph.node_indices().map(|x| x.index()).collect();
+    let paths =
+        all_pairs_dijkstra_shortest_paths(py, &graph.graph, weight_fn, Some(&mut distances))?.paths;
+    let mut nodes: HashSet<usize> = graph.graph.node_indices().map(|x| x.index()).collect();
     let first_node = graph
         .graph
         .node_indices()
         .map(|x| x.index())
         .next()
         .unwrap();
-    let path_keys: HashSet<usize> =
-        paths[&first_node].paths.keys().copied().collect();
+    let path_keys: HashSet<usize> = paths[&first_node].paths.keys().copied().collect();
     // first_node will always be missing from path_keys so if the difference
     // is > 1 with nodes that means there is another node in the graph that
     // first_node doesn't have a path to.
@@ -153,8 +146,7 @@ fn fast_metric_edges(
 
     let cost_fn = |edge: EdgeReference<'_, PyObject>| -> PyResult<f64> {
         if edge.source() != dummy && edge.target() != dummy {
-            let weight: f64 =
-                weight_fn.call1(py, (edge.weight(),))?.extract(py)?;
+            let weight: f64 = weight_fn.call1(py, (edge.weight(),))?.extract(py)?;
             is_valid_weight(weight)
         } else {
             Ok(0.0)
@@ -169,15 +161,13 @@ fn fast_metric_edges(
     graph.graph.remove_node(dummy);
 
     // ``partition[u]`` holds the terminal node closest to node ``u``.
-    let mut partition: Vec<usize> =
-        vec![std::usize::MAX; graph.graph.node_bound()];
+    let mut partition: Vec<usize> = vec![std::usize::MAX; graph.graph.node_bound()];
     for (u, path) in paths.iter() {
         let u = u.index();
         partition[u] = path[1].index();
     }
 
-    let mut out_edges: Vec<MetricClosureEdge> =
-        Vec::with_capacity(graph.graph.edge_count());
+    let mut out_edges: Vec<MetricClosureEdge> = Vec::with_capacity(graph.graph.edge_count());
 
     for edge in graph.graph.edge_references() {
         let source = edge.source();
@@ -185,8 +175,7 @@ fn fast_metric_edges(
         // assert that ``source`` is reachable from a terminal node.
         if distance.contains_key(&source) {
             let weight = distance[&source] + cost_fn(edge)? + distance[&target];
-            let mut path: Vec<usize> =
-                paths[&source].iter().skip(1).map(|x| x.index()).collect();
+            let mut path: Vec<usize> = paths[&source].iter().skip(1).map(|x| x.index()).collect();
             path.append(
                 &mut paths[&target]
                     .iter()
@@ -278,8 +267,7 @@ pub fn steiner_tree(
     terminal_nodes: Vec<usize>,
     weight_fn: PyObject,
 ) -> PyResult<graph::PyGraph> {
-    let mut edge_list =
-        fast_metric_edges(py, graph, &terminal_nodes, &weight_fn)?;
+    let mut edge_list = fast_metric_edges(py, graph, &terminal_nodes, &weight_fn)?;
     let mut subgraphs = UnionFind::<usize>::new(graph.graph.node_bound());
     edge_list.par_sort_unstable_by(|a, b| {
[…]
