+++ tests/test_custom_return_types.py
@@ -33,9 +33,7 @@ class TestBFSSuccessorsComparisons(unittest.TestCase):
         self.assertFalse(retworkx.bfs_successors(self.dag, 0) == [("a", ["c"])])
 
     def test__eq__different_length(self):
-        self.assertFalse(
-            retworkx.bfs_successors(self.dag, 0) == [("a", ["b"]), ("b", ["c"])]
-        )
+        self.assertFalse(retworkx.bfs_successors(self.dag, 0) == [("a", ["b"]), ("b", ["c"])])
 
     def test__eq__invalid_type(self):
         with self.assertRaises(TypeError):
@@ -51,9 +49,7 @@ class TestBFSSuccessorsComparisons(unittest.TestCase):
         self.assertTrue(retworkx.bfs_successors(self.dag, 0) != [("a", ["c"])])
 
     def test__ne__different_length(self):
-        self.assertTrue(
-            retworkx.bfs_successors(self.dag, 0) != [("a", ["b"]), ("b", ["c"])]
-        )
+        self.assertTrue(retworkx.bfs_successors(self.dag, 0) != [("a", ["b"]), ("b", ["c"])])
 
     def test__ne__invalid_type(self):
         with self.assertRaises(TypeError):
@@ -158,24 +154,16 @@ class TestNodesCountMapping(unittest.TestCase):
         self.dag.add_child(node_a, "b", "Edgy")
 
     def test__eq__match(self):
-        self.assertTrue(
-            retworkx.num_shortest_paths_unweighted(self.dag, 0) == {1: 1}
-        )
+        self.assertTrue(retworkx.num_shortest_paths_unweighted(self.dag, 0) == {1: 1})
 
     def test__eq__not_match_keys(self):
-        self.assertFalse(
-            retworkx.num_shortest_paths_unweighted(self.dag, 0) == {2: 1}
-        )
+        self.assertFalse(retworkx.num_shortest_paths_unweighted(self.dag, 0) == {2: 1})
 
     def test__eq__not_match_values(self):
-        self.assertFalse(
-            retworkx.num_shortest_paths_unweighted(self.dag, 0) == {1: 2}
-        )
+        self.assertFalse(retworkx.num_shortest_paths_unweighted(self.dag, 0) == {1: 2})
 
     def test__eq__different_length(self):
-        self.assertFalse(
-            retworkx.num_shortest_paths_unweighted(self.dag, 0) == {1: 1, 2: 2}
-        )
+        self.assertFalse(retworkx.num_shortest_paths_unweighted(self.dag, 0) == {1: 1, 2: 2})
 
     def test_eq__same_type(self):
         self.assertEqual(
@@ -184,39 +172,25 @@ class TestNodesCountMapping(unittest.TestCase):
         )
 
     def test__eq__invalid_type(self):
-        self.assertFalse(
-            retworkx.num_shortest_paths_unweighted(self.dag, 0) == ["a", None]
-        )
+        self.assertFalse(retworkx.num_shortest_paths_unweighted(self.dag, 0) == ["a", None])
 
     def test__eq__invalid_inner_type(self):
-        self.assertFalse(
-            retworkx.num_shortest_paths_unweighted(self.dag, 0) == {0: "a"}
-        )
+        self.assertFalse(retworkx.num_shortest_paths_unweighted(self.dag, 0) == {0: "a"})
 
     def test__ne__match(self):
-        self.assertFalse(
-            retworkx.num_shortest_paths_unweighted(self.dag, 0) != {1: 1}
-        )
+        self.assertFalse(retworkx.num_shortest_paths_unweighted(self.dag, 0) != {1: 1})
 
     def test__ne__not_match(self):
-        self.assertTrue(
[…]

[…]
