### 2022-02-03 caeded5
Bump package version, raise MSRV to 1.48 and drop Python 3.6 support (#560)
Now that the 0.11.0 release is out the door this commit prepares the
main branch for development of the next minor version release, 0.12.0.
To do this it makes three changes, the first is to bump the package
version strings to the new version. This is primarily to differentiate
the development branch against the released 0.11.0. Second, this raises
the minimum supported rust version (MSRV) to 1.48. As was documented in
the release notes the 0.11.0 release was the last version to have an
MSRV of 1.41. Moving forward we'
[…]
