### 2022-02-03 2448e04
Format retworkx-core with line-width=100
