+++ retworkx-core/rustfmt.toml
@@ -1,1 +1,1 @@
-max_width = 80
+max_width = 100
