### 2022-02-03 05dba63
Format retworkx-core with line-width=100 (#561)
Follow up to #490

This PR formats retworkx-core with line width of 100 characters.
On #490, we accidentally left the line width for the retworkx-core as
80. The PR will fix the inconsistency
