### 2022-02-07 #3968 @szhorvat
> @szhorvat likewise @szhorvat, are you happy for the paper to be accepted? (No problem if not, just clarifying for my benefit)

Here's the gist of my opinion:

> My opinion is that the submission to JOSS was premature. This package certainly shows promise, and I think it would make sense to resubmit it in the future, maybe in a year's time. However, trying to work out all the issues within the timeframe of the review does not seem realistic. It would put too much pressure on the authors, as well as on the reviewers.

The package shows good promise and I certainly hope that it will be published in JOSS _eventually_. However, the quality needs to be improved before then. I am not sure how much times JOSS is willing to give authors to make changes. The reason why I recommended later resubmission was to avoid putting time pressure on all those involved (especially the authors).

I should note that the authors have made many improvements since these reviews were written and are working on a revised version of the paper. https://github.com/Qiskit/retworkx/pull/545
