+++ tests/graph/test_tensor_product.py
@@ -57,6 +57,4 @@ class TestTensorProduct(unittest.TestCase):
         graph_2.add_edge(0, 1, "w_2")
 
         graph_product, _ = retworkx.graph_tensor_product(graph_1, graph_2)
-        self.assertEqual(
-            [("w_1", "w_2"), ("w_2", "w_1")], graph_product.edges()
-        )
+        self.assertEqual([("w_1", "w_2"), ("w_2", "w_1")], graph_product.edges())
