### 2022-02-14 b110689
Add suggestions from the code review
