+++ paper.bib
@@ -310,3 +310,16 @@ abstract = {This paper presents a largely improved version of the VF2 algorithm 
       archivePrefix={arXiv},
       primaryClass={quant-ph}
 }
+
+@article{Preskill2018,
+   title={Quantum Computing in the NISQ era and beyond},
+   volume={2},
+   ISSN={2521-327X},
+   url={http://dx.doi.org/10.22331/q-2018-08-06-79},
+   DOI={10.22331/q-2018-08-06-79},
+   journal={Quantum},
+   publisher={Verein zur Forderung des Open Access Publizierens in den Quantenwissenschaften},
+   author={Preskill, John},
+   year={2018},
+   month={Aug},
+   pages={79} }
