### 2022-02-16 444fa48
add for loop for clarity
