+++ tox.ini
@@ -76,10 +76,9 @@ commands = black {posargs} '../retworkx' '../tests'
 basepython = python3
 envdir = .tox/stubs
 deps =
-  pytest
-  mypy==0.910
-  pytest-mypy-testing==0.0.8
-  dataclasses; python_version<'3.7'
+  pytest~=6.2
+  mypy==0.931
+  pytest-mypy-testing==0.0.9
 commands = python -m pytest '../stubs-tests'
 
 [flake8]
