### 2022-02-16 b35bb02
Correct pytest-mypy-testing version
