+++ tox.ini
@@ -78,7 +78,7 @@ envdir = .tox/stubs
 deps =
   pytest~=6.2
   mypy==0.931
-  pytest-mypy-testing==0.0.9
+  pytest-mypy-testing==0.0.8
 commands = python -m pytest '../stubs-tests'
 
 [flake8]
