### 2022-02-16 bafcb5c
add with_capacity for final_graph
