+++ .github/workflows/main.yml
@@ -2,9 +2,9 @@
 name: CI Jobs
 on:
   push:
-    branches: [ main ]
+    branches: [ main, 'stable/*' ]
   pull_request:
-    branches: [ main ]
+    branches: [ main, 'stable/*' ]
 concurrency:
   group: ${{ github.repository }}-${{ github.ref }}-${{ github.head_ref }}
   cancel-in-progress: true
