### 2022-02-18 c2795ad
add node_map and edge assertions
