### 2022-02-19 b4661e6
Change to assert equals
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
