+++ tests/digraph/test_tensor_product.py
@@ -29,7 +29,7 @@ class TestTensorProduct(unittest.TestCase):
 
         graph_product, node_map = retworkx.digraph_tensor_product(graph_1, graph_2)
         expected_node_map = {(0, 0): 0, (0, 1): 1, (1, 0): 2, (1, 1): 3}
-        self.assertTrue(node_map == expected_node_map)
+        self.assertEqual(node_map, expected_node_map)
 
         expected_edges = [(0, 3)]
         self.assertEqual(graph_product.num_nodes(), 4)
