### 2022-02-19 8e69414
Update assert equal
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
