### 2022-02-19 b6c255b
Add Tensor product (#472)
This adds the tensor product of graphs for PyDiGraph and PyGraph.

This PR is related to #440.

* add tensor product

* lint

* fix release notes

* add ProductNodeMap to tensor product

* Update releasenotes/notes/add-tensor-product-651db371a96d650c.yaml

* lint

* black lint

* black lint

* fix docs

* add for loop for clarity

* add with_capacity for final_graph

* add node_map and edge assertions

* lint

* Change to assert equals

* Update assert equal
