+++ src/lib.rs
@@ -26,6 +26,7 @@ mod matching;
 mod random_graph;
 mod shortest_path;
 mod steiner_tree;
+mod tensor_product;
 mod toposort;
 mod transitivity;
 mod traversal;
@@ -43,6 +44,7 @@ use matching::*;
 use random_graph::*;
 use shortest_path::*;
 use steiner_tree::*;
+use tensor_product::*;
 use transitivity::*;
 use traversal::*;
 use tree::*;
@@ -353,6 +355,8 @@ fn retworkx(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_wrapped(wrap_pyfunction!(graph_astar_shortest_path))?;
     m.add_wrapped(wrap_pyfunction!(digraph_astar_shortest_path))?;
     m.add_wrapped(wrap_pyfunction!(graph_greedy_color))?;
+    m.add_wrapped(wrap_pyfunction!(graph_tensor_product))?;
+    m.add_wrapped(wrap_pyfunction!(digraph_tensor_product))?;
     m.add_wrapped(wrap_pyfunction!(directed_gnp_random_graph))?;
     m.add_wrapped(wrap_pyfunction!(undirected_gnp_random_graph))?;
     m.add_wrapped(wrap_pyfunction!(directed_gnm_random_graph))?;
