+++ docs/source/api.rst
@@ -279,6 +279,7 @@ the functions from the explicitly typed based on the data type.
    retworkx.digraph_core_number
    retworkx.digraph_complement
    retworkx.digraph_union
+   retworkx.digraph_tensor_product
    retworkx.digraph_cartesian_product
    retworkx.digraph_random_layout
    retworkx.digraph_bipartite_layout
@@ -325,6 +326,7 @@ typed API based on the data type.
    retworkx.graph_core_number
    retworkx.graph_complement
    retworkx.graph_union
+   retworkx.graph_tensor_product
    retworkx.graph_cartesian_product
    retworkx.graph_random_layout
    retworkx.graph_bipartite_layout
