### 2022-02-23 1242af8
Add attribution comment to introduction tutorial (#554) (#564)
In #537 we added the tutorials to the retworkx documentation. The
introduction tutorial was heavily based (and used some phrases
verbatim from that document) from the similar tutorial in NetworkX.
While the final result is different as it's specific to retworkx the
basic outline and some of the high level details are the same so we
should provide credit to the original source. I noticed reviewing the
published documentation for the 0.11.0 that I neglected to provide
proper attribution of that as part of #537. This corrects that ove
[…]
