### 2022-02-25 60a6d4b
Use new DOI for cited paper that just got published
