+++ paper.bib
@@ -284,7 +284,9 @@ abstract = {This paper presents a largely improved version of the VF2 algorithm 
       year={2021},
       eprint={2104.14722},
       archivePrefix={arXiv},
-      primaryClass={quant-ph}
+      primaryClass={quant-ph},
+      doi={10.48550/arXiv.2104.14722},
+      url={https://doi.org/10.48550/arXiv.2104.14722}
 }
 
 @article{Li2021,
@@ -302,14 +304,20 @@ abstract = {This paper presents a largely improved version of the VF2 algorithm 
    pages={1777–1788}
 }
 
-@misc{Bharti2021,
-      title={Noisy intermediate-scale quantum (NISQ) algorithms}, 
-      author={Kishor Bharti and Alba Cervera-Lierta and Thi Ha Kyaw and Tobias Haug and Sumner Alperin-Lea and Abhinav Anand and Matthias Degroote and Hermanni Heimonen and Jakob S. Kottmann and Tim Menke and Wai-Keong Mok and Sukin Sim and Leong-Chuan Kwek and Alán Aspuru-Guzik},
-      year={2021},
-      eprint={2101.08448},
-      archivePrefix={arXiv},
-      primaryClass={quant-ph}
+@article{Bharti2021,
+   title={Noisy intermediate-scale quantum algorithms},
+   volume={94},
+   ISSN={1539-0756},
+   url={http://dx.doi.org/10.1103/RevModPhys.94.015004},
+   DOI={10.1103/revmodphys.94.015004},
+   number={1},
+   journal={Reviews of Modern Physics},
+   publisher={American Physical Society (APS)},
+   author={Bharti, Kishor and Cervera-Lierta, Alba and Kyaw, Thi Ha and Haug, Tobias and Alperin-Lea, Sumner and Anand, Abhinav and Degroote, Matthias and Heimonen, Hermanni and Kottmann, Jakob S. and Menke, Tim and Mok, Wai-Keong and Sim, Sukin and Kwek, Leong-Chuan and Aspuru-Guzik, Alán},
+   year={2022},
+   month={Feb} 
 }
+
 
 @article{Preskill2018,
    title={Quantum Computing in the NISQ era and beyond},
