### 2022-02-26 a57bd15
retworkx Paper V2 (#545)
This is a rewrite of the retworkx paper. It has a completely rewritten Use Cases section,
which is based on Qiskit's usage of retworkx. It also enchances the Statement of Need
and Related Work sections and adds a Limitations section.

* Change references for v2

* More references

* Rewrite Related Work section

* More rework

* More writing

* Small rewording

* Add Graph Creation, Manipulation, and Traversal

* Update references

* Add Subgraph Isomorphism

* Add limitations section

* Remove one TODO

* Place TODOs for images

* Add image for dag circuit

* Tweak im
[…]
