+++ paper.bib
@@ -40,11 +40,11 @@
                   Christa Zoufal and
                   Donny Greenberg and
                   Stefan Woerner},
-  title        = {Qiskit/qiskit-terra: Qiskit Terra  0.18.3},
+  title        = {Qiskit/qiskit-terra: Qiskit Terra  0.19.1},
   month        = sep,
   year         = 2021,
   publisher    = {Zenodo},
-  version      = {0.18.3},
+  version      = {0.19.1},
   doi          = {10.5281/zenodo.2583252},
   url          = {https://doi.org/10.5281/zenodo.2583252}
 }
@@ -210,46 +210,6 @@
   url = {https://github.com/PyO3/pyo3}
 }
 
-@book{Demetrescu2009,
-  doi = {10.1090/dimacs/074},
-  url = {https://doi.org/10.1090/dimacs/074},
-  year = {2009},
-  month = {jul},
-  publisher = {American Mathematical Society},
-  address = {Providence, RI, USA},
-  editor = {Camil Demetrescu and Andrew V. Goldberg and David Johnson},
-  title = {The Shortest Path Problem: Ninth DIMACS Implementation Challenge}
-}
-
-@Article{Dijkstra1959ANO,
-  title={A note on two problems in connexion with graphs},
-  author={Edsger W. Dijkstra},
-  journal={Numerische Mathematik},
-  year={1959},
-  volume={1},
-  doi={10.1007/BF01386390},
-  pages={269-271}
-}
-
-@article{DeSanto2003,
-author = {De Santo, M. and Foggia, P. and Sansone, C. and Vento, M.},
-title = {A large database of graphs and its use for benchmarking graph isomorphism algorithms},
-journal = {Pattern Recogn. Lett.},
-issue_date = {May 2003},
-volume = {24},
-number = {8},
-month = may,
-year = {2003},
-issn = {0167-8655},
-pages = {1067–1079},
-numpages = {13},
-url = {http://dx.doi.org/10.1016/S0167-8655(02)00253-2},
-doi = {10.1016/S0167-8655(02)00253-2},
-acmid = {763457},
-publisher = {Elsevier Science Inc.},
-address = {New York, NY, USA},
-}
-
 @ARTICLE{Cordella2004,
   author={Cordella, L.P. and Foggia, P. and Sansone, C. and Vento, M.},
   journal={IEEE Transactions on Pattern Analysis and Machine Intelligence}, 
@@ -275,38 +235,99 @@ keywords = {Computational biology, Subgraph isomorphism problem},
 abstract = {This paper presents a largely improved version of the VF2 algorithm for the Subgraph Isomorphism Problem. The improvements are twofold. Firstly, it is based on a new approach for determining the matching order of the nodes, and secondly, more efficient — nevertheless easier to compute — cutting rules are proposed. They together reduce the search space significantly. In addition to the usual Subgraph Isomorphism Problem, the paper also presents specialized algorithms for the Induced Subgraph Isomorphism and for the Graph Isomorphism Problems. Finally, an extensive experimental evaluation is provided using a wide range of inputs, including both real-life biological and chemical datasets and standard randomly generated graph series. The results show major and consistent running time improvements over the other known methods. The C++ implementations of the algorithms are available open-source as part of the LEMON graph and network optimization library.}
 }
 
-@article{Johnson1977,
-author = {Johnson, Donald B.},
-title = {Efficient Algorithms for Shortest Paths in Sparse Networks},
-year = {1977},
[…]
