### 2022-02-26 #3968 @IvanIsCoding
@drvinceknight  We have revised the software paper and want to generate a new version, however we are struggling with the bot. Can you help setting the branch to [joss-paper](https://github.com/qiskit/retworkx/tree/joss-paper)?
