### 2022-02-28 3349be4
Add section to README about slack and irc channels (#501) (#568)
This commit adds a section to the bottom of the README that mentions our
slack and irc channels for retworkx if people would like to reach out
about the library.

(cherry picked from commit 8dd39a42e4384834b22a4a114e41917062c3dd46)

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
