+++ src/iterators.rs
@@ -30,9 +30,7 @@
 //      from `usize` to `f64`.
 //
 // You should always implement `PyGCProtocol` for the new custom return type. If you
-// don't store any python object, just use the macro `default_pygc_protocol_impl`.
-//
-// e.g `default_pygc_protocol_impl!(MyReadOnlyType);`
+// don't store any python object, just use `impl PyGCProtocol for MyReadOnlyType {}`.
 //
 // Types `T, K, V` above should implement `PyHash`, `PyEq`, `PyDisplay` traits.
 // These are arleady implemented for many primitive rust types and `PyObject`.
@@ -46,12 +44,10 @@ use std::hash::Hasher;
 use num_bigint::BigUint;
 use retworkx_core::dictmap::*;
 
-use pyo3::class::iter::{IterNextOutput, PyIterProtocol};
-use pyo3::class::{PyMappingProtocol, PyObjectProtocol, PySequenceProtocol};
+use pyo3::class::iter::IterNextOutput;
 use pyo3::exceptions::{PyIndexError, PyKeyError, PyNotImplementedError};
-use pyo3::gc::{PyGCProtocol, PyVisit};
+use pyo3::gc::PyVisit;
 use pyo3::prelude::*;
-use pyo3::types::PySequence;
 use pyo3::PyTraverseError;
 
 macro_rules! last_type {
@@ -294,39 +290,18 @@ where
     }
 }
 
-impl<A> PyEq<PySequence> for Vec<A>
-where
-    A: PyEq<PyAny>,
-{
-    #[inline]
-    fn eq(&self, other: &PySequence, py: Python) -> PyResult<bool> {
-        if other.len()? as usize != self.len() {
-            return Ok(false);
-        }
-
-        for (i, item) in self.iter().enumerate() {
-            let other_raw = other.get_item(i)?;
-            if !PyEq::eq(item, other_raw, py)? {
-                return Ok(false);
-            }
-        }
-        Ok(true)
-    }
-}
-
-impl<K, V> PyEq<PyObject> for DictMap<K, V>
+impl<K, V> PyEq<PyAny> for DictMap<K, V>
 where
     for<'p> K: PyEq<K> + Clone + pyo3::ToBorrowedObject,
     for<'p> V: PyEq<PyAny>,
 {
     #[inline]
-    fn eq(&self, other: &PyObject, py: Python) -> PyResult<bool> {
-        let other_ref = other.as_ref(py);
-        if other_ref.len()? != self.len() {
+    fn eq(&self, other: &PyAny, py: Python) -> PyResult<bool> {
+        if other.len()? != self.len() {
             return Ok(false);
         }
         for (key, value) in self {
-            match other_ref.get_item(key) {
+            match other.get_item(key) {
                 Ok(other_raw) => {
                     if !PyEq::eq(value, other_raw, py)? {
                         return Ok(false);
@@ -423,23 +398,18 @@ impl<K: PyDisplay, V: PyDisplay> PyDisplay for DictMap<K, V> {
     }
 }
 
-macro_rules! default_pygc_protocol_impl {
-    ($name:ident) => {
-        #[pyproto]
-        impl PyGCProtocol for $name {
-            fn __traverse__(&self, _: PyVisit) -> Result<(), PyTraverseError> {
-                Ok(())
-            }
+trait PyGCProtocol {
+    fn __traverse__(&self, _: PyVisit) -> Result<(), PyTraverseError> {
+        Ok(())
+    }
 
-            fn __clear__(&mut self) {}
-        }
-    };
+    fn __clear__(&mut self) {}
 }
 
 macro_rules! custom_vec_iter_impl {
     ($name:ident, $data:ident, $T:ty, $doc:literal) => {
         #[doc = $doc]
-        #[pyclass(module = "retworkx", gc)]
[…]
