+++ Cargo.toml
@@ -32,7 +32,7 @@ num-complex = "0.4"
 retworkx-core = { path = "./retworkx-core" }
 
 [dependencies.pyo3]
-version = "0.16.0"
+version = "0.16.1"
 features = ["extension-module", "hashbrown", "num-bigint", "num-complex", "indexmap"]
 
 [dependencies.hashbrown]
