### 2022-03-07 #3968 @drvinceknight
> @editorialbot set joss-paper as branch

Thank you @arfon.

Apologies for the delay @IvanIsCoding, let me know if there's anything else you need at this stage.
