### 2022-03-10 2448878
add difference and tests
