+++ src/lib.rs
@@ -15,6 +15,7 @@ mod centrality;
 mod coloring;
 mod connectivity;
 mod dag_algo;
+mod difference;
 mod digraph;
 mod dot_utils;
 mod generators;
@@ -38,6 +39,7 @@ use centrality::*;
 use coloring::*;
 use connectivity::*;
 use dag_algo::*;
+use difference::*;
 use isomorphism::*;
 use layout::*;
 use matching::*;
@@ -325,6 +327,8 @@ fn retworkx(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_wrapped(wrap_pyfunction!(graph_union))?;
     m.add_wrapped(wrap_pyfunction!(digraph_cartesian_product))?;
     m.add_wrapped(wrap_pyfunction!(graph_cartesian_product))?;
+    m.add_wrapped(wrap_pyfunction!(digraph_difference))?;
+    m.add_wrapped(wrap_pyfunction!(graph_difference))?;
     m.add_wrapped(wrap_pyfunction!(topological_sort))?;
     m.add_wrapped(wrap_pyfunction!(descendants))?;
     m.add_wrapped(wrap_pyfunction!(ancestors))?;
