+++ releasenotes/notes/add-graph-difference-9916bf3d612f0b1a.yaml
@@ -1,0 +1,31 @@
+---
+features:
+  - |
+    Add two new functions which calculates the difference of two graphs :func:`~retworkx.graph_difference`
+    for undirected graphs and :func:`~retworkx.digraph_difference` for directed graphs. For example:
+    
+    .. jupyter-execute::
+
+      import retworkx
+      from retworkx.visualization import mpl_draw
+      
+      graph_1 = retworkx.PyGraph()
+      graph_1.add_nodes_from(["a_1", "a_2", "a_3", "a_4"])
+      graph_1.extend_from_weighted_edge_list([(0, 1, "e_1"),
+                                              (1, 2, "e_2"), 
+                                              (2, 3, "e_3"),
+                                              (3, 0, "e_4"),
+                                              (0, 2, "e_5"),
+                                              (1, 3, "e_6"),
+                                              ])
+      graph_2 = retworkx.PyGraph()
+      graph_2.add_nodes_from(["a_1", "a_2", "a_3", "a_4"])
+      graph_2.extend_from_weighted_edge_list([(0, 1, "e_1"),
+                                              (1, 2, "e_2"), 
+                                              (2, 3, "e_3"),
+                                              (3, 0, "e_4"),
+                                              ])
+        
+      graph_difference = retworkx.graph_difference(graph_1, graph_2)
+      
+      mpl_draw(graph_difference)
