+++ retworkx/__init__.py
@@ -1815,6 +1815,41 @@ def _graph_cartesian_product(
 
 
 @functools.singledispatch
+def difference(
+    first,
+    second,
+):
+    """Return a new PyGraph that is the difference from two input
+    graph objects
+
+    :param first: The first graph object
+    :param second: The second graph object
+
+    :returns: A new graph object that is the difference of ``second`` and
+        ``first``. It's worth noting the weight/data payload objects are
+        passed by reference from ``first`` to this new object.
+
+    :rtype: :class:`~retworkx.PyGraph` or :class:`~retworkx.PyDiGraph`
+    """
+    raise TypeError("Invalid Input Type %s for graph" % type(first))
+
+
+@difference.register(PyDiGraph)
+def _digraph_difference(
+    first,
+    second,
+):
+    return digraph_difference(first, second)
+
+
+@difference.register(PyGraph)
+def _graph_difference(
+    first,
+    second,
+):
+    return graph_difference(first, second)
+
+@functools.singledispatch
 def bfs_search(graph, source, visitor):
     """Breadth-first traversal of a directed/undirected graph.
 
