+++ Cargo.toml
@@ -22,7 +22,7 @@ crate-type = ["cdylib"]
 ahash = { version = "0.7.6", default-features = false }
 petgraph = "0.6.0"
 fixedbitset = "0.4.1"
-numpy = "0.16.0"
+numpy = "0.16.2"
 rand = "0.8"
 rand_pcg = "0.3"
 rayon = "1.5"
@@ -32,7 +32,7 @@ num-complex = "0.4"
 retworkx-core = { path = "./retworkx-core" }
 
 [dependencies.pyo3]
-version = "0.16.1"
+version = "0.16.2"
 features = ["extension-module", "hashbrown", "num-bigint", "num-complex", "indexmap"]
 
 [dependencies.hashbrown]
