+++ src/transitivity.rs
@@ -70,7 +70,7 @@ fn _graph_triangles(graph: &graph::PyGraph, node: usize) -> (usize, usize) {
 /// :rtype: float
 #[pyfunction]
 #[pyo3(text_signature = "(graph, /)")]
-fn graph_transitivity(graph: &graph::PyGraph) -> f64 {
+pub fn graph_transitivity(graph: &graph::PyGraph) -> f64 {
     let node_indices: Vec<NodeIndex> = graph.graph.node_indices().collect();
     let (triangles, triples) = node_indices
         .par_iter()
@@ -171,7 +171,7 @@ fn _digraph_triangles(graph: &digraph::PyDiGraph, node: usize) -> (usize, usize)
 ///    Physical Review E, 76(2), 026107 (2007)
 #[pyfunction]
 #[pyo3(text_signature = "(graph, /)")]
-fn digraph_transitivity(graph: &digraph::PyDiGraph) -> f64 {
+pub fn digraph_transitivity(graph: &digraph::PyDiGraph) -> f64 {
     let node_indices: Vec<NodeIndex> = graph.graph.node_indices().collect();
     let (triangles, triples) = node_indices
         .par_iter()
