### 2022-03-28 d45dc17
First setup of planar layout for retworkx
