### 2022-03-30 72c7c91
add generalized Petersen graph as test cases
