### 2022-03-30 b6f4a92
remove old release note
