+++ releasenotes/notes/dfs-search-6083680bf62356b0.yaml
@@ -1,28 +1,0 @@
----
-features:
-  - |
-    Added a new :func:`~retworkx.dfs_search` (and it's per type variants
-    :func:`~retworkx.graph_dfs_search` and :func:`~retworkx.digraph_dfs_search`)
-    that traverses the graph in a depth-first manner and emits events at specified
-    points. The events are handled by a visitor object that subclasses
-    :class:`~retworkx.visit.DFSVisitor` through the appropriate callback functions.
-    For example:
-
-    .. jupyter-execute::
-
-      import retworkx
-      from retworkx.visit import DFSVisitor
-
-      class TreeEdgesRecorder(DFSVisitor):
-
-          def __init__(self):
-              self.edges = []
-
-          def tree_edge(self, edge):
-              self.edges.append(edge)
-
-      graph = retworkx.PyGraph()
-      graph.extend_from_edge_list([(1, 3), (0, 1), (2, 1), (0, 2)])
-      vis = TreeEdgesRecorder()
-      retworkx.dfs_search(graph, [0], vis)
-      print('Tree edges:', vis.edges)
