### 2022-04-05 871a5f4
Bump pyo3 from 0.16.2 to 0.16.3
Bumps [pyo3](https://github.com/pyo3/pyo3) from 0.16.2 to 0.16.3.
- [Release notes](https://github.com/pyo3/pyo3/releases)
- [Changelog](https://github.com/PyO3/pyo3/blob/main/CHANGELOG.md)
- [Commits](https://github.com/pyo3/pyo3/compare/v0.16.2...v0.16.3)

---
updated-dependencies:
- dependency-name: pyo3
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
