+++ src/graphml.rs
@@ -1,0 +1,684 @@
+// Licensed under the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License. You may obtain
+// a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations
+// under the License.
+
+use std::io::BufRead;
+use std::iter::FromIterator;
+use std::path::Path;
+
+use hashbrown::HashMap;
+use indexmap::IndexMap;
+
+use quick_xml::events::{BytesStart, Event};
+use quick_xml::Reader;
+
+use petgraph::algo;
+use petgraph::{Directed, Undirected};
+
+use pyo3::prelude::*;
+
+use self::error::Error;
+use crate::{digraph::PyDiGraph, graph::PyGraph, StablePyGraph};
+
+mod error {
+    use std::convert::From;
+    use std::num::{ParseFloatError, ParseIntError};
+    use std::str::ParseBoolError;
+
+    use pyo3::exceptions::PyException;
+    use pyo3::PyErr;
+
+    use quick_xml::Error as XmlError;
+
+    #[derive(Debug)]
+    pub enum Error {
+        Xml(String),
+        ParseValue(String),
+        NotFound(String),
+        UnSupportedAttribute(String),
+        InvalidDoc(String),
+    }
+
+    impl From<XmlError> for Error {
+        #[inline]
+        fn from(e: XmlError) -> Error {
+            Error::Xml(format!("Xml document not well-formed: {}", e))
+        }
+    }
+
+    impl From<ParseBoolError> for Error {
+        #[inline]
+        fn from(e: ParseBoolError) -> Error {
+            Error::ParseValue(format!("Failed conversion to 'bool': {}", e))
+        }
+    }
+
+    impl From<ParseIntError> for Error {
+        #[inline]
+        fn from(e: ParseIntError) -> Error {
+            Error::ParseValue(format!("Failed conversion to 'int': {}", e))
+        }
+    }
+
+    impl From<ParseFloatError> for Error {
+        #[inline]
+        fn from(e: ParseFloatError) -> Error {
+            Error::ParseValue(format!("Failed conversion to 'float' or 'double': {}", e))
+        }
+    }
+
+    impl From<Error> for PyErr {
+        #[inline]
+        fn from(error: Error) -> PyErr {
+            match error {
+                Error::Xml(msg)
+                | Error::ParseValue(msg)
+                | Error::NotFound(msg)
+                | Error::UnSupportedAttribute(msg)
+                | Error::InvalidDoc(msg) => PyException::new_err(msg),
+            }
+        }
+    }
+}
+
+fn xml_attribute<'a, B: BufRead>(
+    reader: &Reader<B>,
+    element: &'a BytesStart<'a>,
+    key: &[u8],
+) -> Result<String, Error> {
+    element
+        .attributes()
+        .find_map(|a| {
+            if let Ok(a) = a {
+                if a.key == key {
+                    let decoded = a.unescape_and_decode_value(reader).map_err(Error::from);
+                    return Some(decoded);
+                }
+            }
+            None
+        })
+        .unwrap_or_else(|| {
+            Err(Error::NotFound(format!(
+                "Attribute '{}' not found.",
+                String::from_utf8_lossy(key)
+            )))
+        })
+}
+
[…]
