+++ tests/test_graphml.py
@@ -1,0 +1,399 @@
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+import unittest
+import tempfile
+import numpy
+
+import retworkx
+
+
+class TestGraphML(unittest.TestCase):
+    HEADER = """
+        <?xml version="1.0" encoding="UTF-8"?>
+        <graphml xmlns="http://graphml.graphdrawing.org/xmlns"
+                xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                xsi:schemaLocation="http://graphml.graphdrawing.org/xmlns
+                http://graphml.graphdrawing.org/xmlns/1.0/graphml.xsd">
+        {}
+        </graphml>
+    """
+
+    def assertDictEqual(self, xs, ys):
+        self.assertEqual(len(xs), len(ys))
+        for key, va in xs.items():
+            vb = ys.get(key, None)
+            self.assertTrue(
+                (isinstance(va, float) and isinstance(vb, float) and numpy.allclose(va, vb))
+                or (va == vb)
+            )
+
+    def assertGraphEqual(self, graph, nodes, edges, directed=True):
+        self.assertTrue(isinstance(graph, retworkx.PyDiGraph if directed else retworkx.PyGraph))
+        self.assertEqual(len(graph), len(nodes))
+        for node_a, node_b in zip(graph.nodes(), nodes):
+            self.assertDictEqual(node_a, node_b)
+
+        for ((s, t, data), edge) in zip(graph.weighted_edge_list(), edges):
+            assert (graph[s]["id"], graph[t]["id"]) == (edge[0], edge[1])
+            self.assertDictEqual(data, edge[2])
+
+    def assertGraphMLRaises(self, graph_xml):
+        with tempfile.NamedTemporaryFile("wt") as fd:
+            fd.write(graph_xml)
+            fd.flush()
+            with self.assertRaises(Exception):
+                retworkx.read_graphml(fd.name)
+
+    def test_simple(self):
+        graph_xml = self.HEADER.format(
+            """
+            <key id="d0" for="node" attr.name="color" attr.type="string">
+            <default>yellow</default>
+            </key>
+            <key id="d1" for="edge" attr.name="fidelity" attr.type="float">
+            <default>0.95</default>
+            </key>
+            <graph id="G" edgedefault="undirected">
+            <node id="n0">
+                <data key="d0">blue</data>
+            </node>
+            <node id="n1"/>
+            <node id="n2">
+                <data key="d0">green</data>
+            </node>
+            <edge source="n0" target="n1">
+                <data key="d1">0.98</data>
+            </edge>
+            <edge source="n0" target="n2"/>
+            </graph>
+            """
+        )
+
+        with tempfile.NamedTemporaryFile("wt") as fd:
+            fd.write(graph_xml)
+            fd.flush()
+            graphml = retworkx.read_graphml(fd.name)
+            graph = graphml[0]
+            nodes = [
+                {"id": "n0", "color": "blue"},
+                {"id": "n1", "color": "yellow"},
+                {"id": "n2", "color": "green"},
+            ]
+            edges = [
+                ("n0", "n1", {"fidelity": 0.98}),
[…]
