+++ src/lib.rs
@@ -19,6 +19,7 @@ mod digraph;
 mod dot_utils;
 mod generators;
 mod graph;
+mod graphml;
 mod isomorphism;
 mod iterators;
 mod layout;
@@ -38,6 +39,7 @@ use centrality::*;
 use coloring::*;
 use connectivity::*;
 use dag_algo::*;
+use graphml::*;
 use isomorphism::*;
 use layout::*;
 use matching::*;
@@ -405,6 +407,7 @@ fn retworkx(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_wrapped(wrap_pyfunction!(articulation_points))?;
     m.add_wrapped(wrap_pyfunction!(biconnected_components))?;
     m.add_wrapped(wrap_pyfunction!(chain_decomposition))?;
+    m.add_wrapped(wrap_pyfunction!(read_graphml))?;
     m.add_class::<digraph::PyDiGraph>()?;
     m.add_class::<graph::PyGraph>()?;
     m.add_class::<toposort::TopologicalSorter>()?;
