+++ docs/source/api.rst
@@ -235,6 +235,16 @@ Layout Functions
    retworkx.spiral_layout
 
 
+.. _graphml:
+
+GraphML
+==========
+
+.. autosummary::
+   :toctree: apiref
+
+   retworkx.read_graphml
+
 .. _converters:
 
 Converters
