+++ Cargo.toml
@@ -29,6 +29,7 @@ rayon = "1.5"
 num-traits = "0.2"
 num-bigint = "0.4"
 num-complex = "0.4"
+quick-xml = "0.22.0"
 retworkx-core = { path = "./retworkx-core" }
 
 [dependencies.pyo3]
