### 2022-04-05 355a2a4
remove unused derived debug trait and redundant `finalize_graph` logic + annotate unreachable code parts
