+++ src/iterators.rs
@@ -1113,6 +1113,170 @@ impl PyDisplay for PathMapping {
     }
 }
 
+/// A custom class for the return multiple paths to target nodes
+///
+/// The class is a read-only mapping of node indices to a list of node indices
+/// representing a path of the form::
+///
+///     {node_c: [[node_a, node_b, node_c], [node_a, node_c]]}
+///
+/// where ``node_a``, ``node_b``, and ``node_c`` are integer node indices.
+///
+/// This class is a container class for the results of functions that
+/// return a mapping of target nodes and paths. It implements the Python
+/// mapping protocol. So you can treat the return as a read-only
+/// mapping/dict.
+#[pyclass(mapping, module = "retworkx")]
+#[derive(Clone)]
+pub struct MultiplePathMapping {
+    pub paths: DictMap<usize, Vec<Vec<usize>>>,
+}
+
+#[pymethods]
+impl MultiplePathMapping {
+    #[new]
+    fn new() -> MultiplePathMapping {
+        MultiplePathMapping {
+            paths: DictMap::new(),
+        }
+    }
+
+    fn __getstate__(&self) -> DictMap<usize, Vec<Vec<usize>>> {
+        self.paths.clone()
+    }
+
+    fn __setstate__(&mut self, state: DictMap<usize, Vec<Vec<usize>>>) {
+        self.paths = state;
+    }
+
+    fn keys(&self) -> MultiplePathMappingKeys {
+        MultiplePathMappingKeys {
+            path_keys: self.paths.keys().copied().collect(),
+            iter_pos: 0,
+        }
+    }
+
+    fn values(&self) -> MultiplePathMappingValues {
+        MultiplePathMappingValues {
+            path_values: self
+                .paths
+                .values()
+                .map(|paths| {
+                    paths
+                        .iter()
+                        .map(|v| NodeIndices { nodes: v.to_vec() })
+                        .collect()
+                })
+                .collect(),
+            iter_pos: 0,
+        }
+    }
+
+    fn items(&self) -> MultiplePathMappingItems {
+        let items: Vec<(usize, Vec<NodeIndices>)> = self
+            .paths
+            .iter()
+            .map(|(k, paths)| {
+                let out_paths: Vec<NodeIndices> = paths
+                    .iter()
+                    .map(|v| NodeIndices { nodes: v.to_vec() })
+                    .collect();
+
+                (*k, out_paths)
+            })
+            .collect();
+        MultiplePathMappingItems {
+            path_items: items,
+            iter_pos: 0,
+        }
+    }
+
+    fn __richcmp__(&self, other: &PyAny, op: pyo3::basic::CompareOp) -> PyResult<bool> {
+        let compare = |other: &PyAny| -> PyResult<bool> {
+            Python::with_gil(|py| PyEq::eq(&self.paths, other, py))
+        };
+        match op {
+            pyo3::basic::CompareOp::Eq => compare(other),
+            pyo3::basic::CompareOp::Ne => match compare(other) {
+                Ok(res) => Ok(!res),
+                Err(err) => Err(err),
+            },
+            _ => Err(PyNotImplementedError::new_err("Comparison not implemented")),
+        }
+    }
+
+    fn __str__(&self) -> PyResult<String> {
+        Python::with_gil(|py| Ok(format!("MultiplePathMapping{}", self.paths.str(py)?)))
+    }
+
+    fn __hash__(&self) -> PyResult<u64> {
[…]
