+++ tests/graph/test_all_simple_paths.py
@@ -176,3 +176,2608 @@ class TestGraphAllSimplePaths(unittest.TestCase):
         dag.add_node(0)
         dag.add_node(1)
         self.assertRaises(TypeError, retworkx.graph_all_simple_paths, (dag, 0, 1))
+
+
+class TestGraphAllSimplePathsAllPairs(unittest.TestCase):
+    def setUp(self):
+        super().setUp()
+        self.edges = [
+            (0, 1),
+            (0, 2),
+            (0, 3),
+            (1, 2),
+            (1, 3),
+            (2, 3),
+            (2, 4),
+            (3, 2),
+            (3, 4),
+            (4, 2),
+            (4, 5),
+            (5, 2),
+            (5, 3),
+        ]
+
+    def test_all_simple_paths(self):
+        dag = retworkx.PyGraph()
+        for i in range(6):
+            dag.add_node(i)
+        dag.add_edges_from_no_data(self.edges)
+        paths = retworkx.all_pairs_all_simple_paths(dag)
+        expected = {
+            0: {
+                0: [
+                    [0, 3, 4, 5, 2, 0],
+                    [0, 3, 4, 2, 1, 0],
+                    [0, 3, 4, 2, 0],
+                    [0, 3, 4, 2, 1, 0],
+                    [0, 3, 4, 2, 0],
+                    [0, 3, 2, 1, 0],
+                    [0, 3, 2, 0],
+                    [0, 3, 5, 2, 1, 0],
+                    [0, 3, 5, 2, 0],
+                    [0, 3, 5, 4, 2, 0],
+                    [0, 3, 5, 4, 2, 0],
+                    [0, 3, 2, 1, 0],
+                    [0, 3, 2, 0],
+                    [0, 3, 1, 2, 0],
+                    [0, 3, 1, 0],
+                    [0, 3, 0],
+                    [0, 2, 4, 5, 3, 0],
+                    [0, 2, 4, 3, 1, 0],
+                    [0, 2, 4, 3, 0],
+                    [0, 2, 3, 1, 0],
+                    [0, 2, 3, 0],
+                    [0, 2, 5, 3, 1, 0],
+                    [0, 2, 5, 3, 0],
+                    [0, 2, 5, 4, 3, 0],
+                    [0, 2, 4, 5, 3, 0],
+                    [0, 2, 4, 3, 1, 0],
+                    [0, 2, 4, 3, 0],
+                    [0, 2, 3, 1, 0],
+                    [0, 2, 3, 0],
+                    [0, 2, 1, 3, 0],
+                    [0, 2, 1, 0],
+                    [0, 2, 0],
+                    [0, 1, 3, 4, 2, 0],
+                    [0, 1, 3, 4, 2, 0],
+                    [0, 1, 3, 2, 0],
+                    [0, 1, 3, 5, 2, 0],
+                    [0, 1, 3, 2, 0],
+                    [0, 1, 3, 0],
+                    [0, 1, 2, 4, 3, 0],
+                    [0, 1, 2, 3, 0],
+                    [0, 1, 2, 5, 3, 0],
+                    [0, 1, 2, 4, 3, 0],
+                    [0, 1, 2, 3, 0],
+                    [0, 1, 2, 0],
+                    [0, 1, 0],
+                ],
+                1: [
+                    [0, 3, 4, 5, 2, 1],
+                    [0, 3, 4, 2, 1],
+                    [0, 3, 4, 2, 1],
+                    [0, 3, 2, 1],
+                    [0, 3, 5, 2, 1],
+                    [0, 3, 5, 4, 2, 1],
+                    [0, 3, 5, 4, 2, 1],
+                    [0, 3, 2, 1],
+                    [0, 3, 1],
[…]
