### 2022-04-07 2043bf0
Fix function signature typos
