+++ retworkx-core/src/centrality.rs
@@ -168,7 +168,7 @@ where
 }
 
 fn _rescale(
-    betweenness: &mut Vec<Option<f64>>,
+    betweenness: &mut [Option<f64>],
     node_count: usize,
     normalized: bool,
     directed: bool,
