### 2022-04-07 4e0e217
use `assertEqual` not `assert`
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
