### 2022-04-08 e49eb76
raise if hyperedges or ports
