+++ src/connectivity/mod.rs
@@ -633,10 +633,10 @@ pub fn digraph_all_simple_paths(
 /// with the ``RAYON_NUM_THREADS`` environment variable. For example, setting
 /// ``RAYON_NUM_THREADS=4`` would limit the thread pool to 4 threads.
 ///
-/// :param PyGraph graph: The graph to find all simple paths in
+/// :param PyDiGraph graph: The graph to find all simple paths in
 /// :param int min_depth: The minimum depth of the path to include in the output
 ///     list of paths. By default all paths are included regardless of depth,
-///     sett to 0 will behave like the default.
+///     setting to 0 will behave like the default.
 /// :param int cutoff: The maximum depth of path to include in the output list
 ///     of paths. By default includes all paths regardless of depth, setting to
 ///     0 will behave like default.
@@ -664,7 +664,7 @@ pub fn digraph_all_pairs_all_simple_paths(
 /// :param PyGraph graph: The graph to find all simple paths in
 /// :param int min_depth: The minimum depth of the path to include in the output
 ///     list of paths. By default all paths are included regardless of depth,
-///     sett to 0 will behave like the default.
+///     setting to 0 will behave like the default.
 /// :param int cutoff: The maximum depth of path to include in the output list
 ///     of paths. By default includes all paths regardless of depth, setting to
 ///     0 will behave like default.//
