### 2022-04-09 7936dca
Simplify and deduplicate multigraph handling in PyGraph (#580)
This commit simplifies and deduplicates the PyGraph method definitions
that add edges to the graph object. When reviewing the multigraph=False
path to debug an issue I found that the PyGraph code around edge
handling was duplicated between every method that adds edges and also
was needlessly returning a PyResult<T>. This commit abstracts the edge
adding to an inner private method that does the multigraph flag handling
and also drops the PyResult return as these functions can't raise an
error. The return type for the functions which
[…]
