+++ tox.ini
@@ -45,7 +45,7 @@ extras =
 basepython = python3
 envdir = .tox/lint
 deps =
-  black==21.5b0
+  black~=22.0
   flake8
   setuptools-rust
 whitelist_externals=cargo
@@ -69,7 +69,7 @@ commands =
 basepython = python3
 envdir = .tox/lint
 deps =
-  black==21.5b0
+  black~=22.0
 commands = black {posargs} '../retworkx' '../tests'
 
 [flake8]
