### 2022-04-11 6e0794c
Update test_binomial_tree.py
