+++ tests/generators/test_binomial_tree.py
@@ -196,8 +196,8 @@ class TestBinomialTreeGraph(unittest.TestCase):
                 graph = retworkx.generators.directed_binomial_tree_graph(
                     n, bidirectional=True
                 )
-                self.assertEqual(len(graph), 2 ** n)
-                self.assertEqual(len(graph.edges()), 2 * (2 ** n - 1))
+                self.assertEqual(len(graph), 2**n)
+                self.assertEqual(len(graph.edges()), 2 * (2**n - 1))
                 self.assertEqual(list(graph.edge_list()), expected_edges[n])
 
     def test_overflow_binomial_tree(self):
