+++ retworkx-core/src/max_weight_matching.rs
@@ -127,7 +127,7 @@ fn scan_blossom(
     in_blossoms: &[usize],
     blossom_base: &[Option<usize>],
     endpoints: &[usize],
-    labels: &mut Vec<Option<usize>>,
+    labels: &mut [Option<usize>],
     label_ends: &[Option<usize>],
     mate: &HashMap<usize, usize>,
 ) -> Option<usize> {
@@ -181,21 +181,21 @@ fn scan_blossom(
 fn add_blossom<E>(
     base: usize,
     edge: usize,
-    blossom_children: &mut Vec<Vec<usize>>,
+    blossom_children: &mut [Vec<usize>],
     num_nodes: usize,
     edges: &[(usize, usize, i128)],
-    in_blossoms: &mut Vec<usize>,
-    dual_var: &mut Vec<i128>,
-    labels: &mut Vec<Option<usize>>,
-    label_ends: &mut Vec<Option<usize>>,
-    best_edge: &mut Vec<Option<usize>>,
+    in_blossoms: &mut [usize],
+    dual_var: &mut [i128],
+    labels: &mut [Option<usize>],
+    label_ends: &mut [Option<usize>],
+    best_edge: &mut [Option<usize>],
     queue: &mut Vec<usize>,
-    blossom_base: &mut Vec<Option<usize>>,
+    blossom_base: &mut [Option<usize>],
     endpoints: &[usize],
-    blossom_endpoints: &mut Vec<Vec<usize>>,
+    blossom_endpoints: &mut [Vec<usize>],
     unused_blossoms: &mut Vec<usize>,
-    blossom_best_edges: &mut Vec<Vec<usize>>,
-    blossom_parents: &mut Vec<Option<usize>>,
+    blossom_best_edges: &mut [Vec<usize>],
+    blossom_parents: &mut [Option<usize>],
     neighbor_endpoints: &[Vec<usize>],
     mate: &HashMap<usize, usize>,
 ) -> Result<(), E> {
