### 2022-04-11 2fa5977
Update coverage job to use stable rust (#579) (#581)
In the recent Rust 1.60.0 release the instrument-coverage compiler flag
was stabilized. This means it no longer requires a nightly build of the
rust compiler to use it. This commit updates our coverage CI job to use
stable rust so we don't have to rely on a nightly build anymore.

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
(cherry picked from commit 52759d04f987ac72be42d84d2276e82e1d1d6460)

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.nor
[…]
