### 2022-04-11 f4490cd
Raise a ValueError on valid input for min_depth or cutoff
