### 2022-04-11 4b8af81
Revert version bump from main
This commit reverts the version bump from main that was part of the
automated backport. We don't want to have anything retworkx at
version 0.12.x on the Stable/0.11 branch.
