+++ retworkx-core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "retworkx-core"
-version = "0.12.0"
+version = "0.11.0"
 edition = "2018"
 authors = ["Matthew Treinish <mtreinish@kortar.org>"]
 description = "Rust APIs used for retworkx algorithms"
