### 2022-04-11 0bf06a8
Update workspace configuration to be tied to specific versions (backport #583) (#584)
* Update workspace configuration to be tied to specific versions (#583)

This commit updates the retworkx-core crate dependency to have a fixed
version which matches the retworkx crates version. We're not going to be
in a situation where we we release retworkx-core separately from the
main retworkx crate we only want to support the main crate with the core
crate at the same version. This commit makes this relationship
explicitly. At the same time it fixes the version string in the
retworkx-core crate as it wa
[…]
