+++ retworkx-core/src/connectivity/mod.rs
@@ -12,8 +12,10 @@
 
 //! Module for connectivity and cut algorithms.
 
+mod all_simple_paths;
 mod biconnected;
 mod chain;
 
+pub use all_simple_paths::all_simple_paths_multiple_targets;
 pub use biconnected::articulation_points;
 pub use chain::chain_decomposition;
