### 2022-04-14 d74cfc8
Avoid temp Vec when visited.len() == max_length
