### 2022-04-14 d80c11f
Add additional check to avoid extra branch
Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>
