### 2022-04-14 0647e15
Revamp undirected graph tests
