### 2022-04-14 edda11d
Move all_simple_paths under connectivity functions in docs
