### 2022-04-15 #3968 @IvanIsCoding
> What is the situation with this paper now @drvinceknight, @IvanIsCoding, @mtreinish, @inakleinbottle?
>
> Are we ready for a new review round?

@szhorvat  @inakleinbottle The software paper is ready for a new round of review, there are substantial changes compared to the first version.

Addressing the issues opened on our repository is still WIP, we [have closed 13 out of the 24 issues with tag joss](https://github.com/Qiskit/retworkx/issues?q=joss+is%3Aissue).
