### 2022-04-17 f3bc9f3
fix issue #585 that pickling graph & digraph do not preserve original edge index
