### 2022-04-18 49986bf
fix clippy lints - collapsible_else_if
