### 2022-04-20 c0f7cd9
Add PlanarEmbedding structure
