+++ src/layout/planar.rs
@@ -1,56 +1,155 @@
-use petgraph::EdgeType;
+use petgraph::{EdgeType, Directed};
+use petgraph::prelude::*;
+use pyo3::{PyObject};
+use pyo3::prelude::*;
 
+use super::spring::{recenter, rescale, Point};
 use crate::iterators::Pos2DMapping;
 use crate::StablePyGraph;
+use crate::connected_components;
+use retworkx_core::dictmap::*;
 
 pub fn planar_layout<Ty: EdgeType>(
+    py: Python,
     graph: &StablePyGraph<Ty>,
     scale: Option<f64>,
-    center: Option<[f64; 2]>,
+    center: Option<Point>,
 ) -> Pos2DMapping {
 
-    match scale {
-        Some(scale) => scale,
-        None => 1.0,
-    };
-    println!("HELLO WORLD!");
-    Pos2DMapping {
-        pos_map: graph
-            .node_indices()
-            .map(|n| {
-                let random_tuple: [f64; 2] = [5.0, 5.0];
-                match center {
-                    Some(center) => (
-                        n.index(),
-                        [random_tuple[0] + center[0], random_tuple[1] + center[1]],
-                    ),
-                    None => (n.index(), random_tuple),
-                }
-            })
-            .collect(),
+    let node_num = graph.node_count();
+    if node_num == 0 {
+        return Pos2DMapping {
+            pos_map: DictMap::new(),
+        };
+    }
+    let mut pos: Vec<Point> = Vec::with_capacity(node_num);
+    let mut planar = PlanarEmbedding::new();
+
+    if !is_planar(graph) {
+        return Pos2DMapping {
+            pos_map: DictMap::new(),
+        };
+    } else {
+        create_embedding(graph, &mut planar.embedding);
+        combinitorial_embedding_to_pos(&planar.embedding, &mut pos);
+
+        if let Some(scale) = scale {
+            rescale(&mut pos, scale, (0..node_num).collect());
+        }
+        if let Some(center) = center {
+            recenter(&mut pos, center);
+        }
+        Pos2DMapping {
+            pos_map: planar.embedding.node_indices().map(|n| n.index()).zip(pos).collect(),
+        }
     }
 }
 
-//     import numpy as np
-// 
-//     if dim != 2:
-//         raise ValueError("can only handle 2 dimensions")
-// 
-//     G, center = _process_params(G, center, dim)
-// 
-//     if len(G) == 0:
-//         return {}
-// 
-//     if isinstance(G, nx.PlanarEmbedding):
-//         embedding = G
-//     else:
-//         is_planar, embedding = nx.check_planarity(G)
-//         if not is_planar:
-//             raise nx.NetworkXException("G is not planar.")
-//     pos = nx.combinatorial_embedding_to_pos(embedding)
-//     node_list = list(embedding)
-//     pos = np.row_stack([pos[x] for x in node_list])
-//     pos = pos.astype(np.float64)
-//     pos = rescale_layout(pos, scale=scale) + center
-//     return dict(zip(node_list, pos))
-// 
+pub fn is_planar<Ty: EdgeType>(
+    graph: &StablePyGraph<Ty>,
+) -> bool {
+    true
+}
+
+pub fn create_embedding<Ty: EdgeType>(
+    graph: &StablePyGraph<Ty>,
+    embedding: &mut StablePyGraph<Directed>,
+) -> bool {
+    // DEBUG CODE FOR TESTING BASIC EMBEDDING
+    for v in graph.node_indices() {
+        if v.index() < 5 {
[…]
