### 2022-04-23 6e819b5
Initial implementation of closeness_centrality
