+++ tests/graph/test_centrality.py
@@ -56,6 +56,16 @@ class TestCentralityGraph(unittest.TestCase):
         expected = {0: 0.0, 1: 2.0, 2: 2.0, 3: 0.0}
         self.assertEqual(expected, betweenness)
 
+    def test_closeness_centrality(self):
+        betweenness = retworkx.graph_closeness_centrality(self.graph)
+        expected = {0: 0.5, 1: 0.75, 2: 0.75, 3: 0.5}
+        self.assertEqual(expected, betweenness)
+
+    def test_closeness_centrality_wf_improved(self):
+        betweenness = retworkx.graph_closeness_centrality(self.graph, wf_improved=False)
+        expected = {0: 0.5, 1: 0.75, 2: 0.75, 3: 0.5}
+        self.assertEqual(expected, betweenness)
+
 
 class TestCentralityGraphDeletedNode(unittest.TestCase):
     def setUp(self):
@@ -98,4 +108,14 @@ class TestCentralityGraphDeletedNode(unittest.TestCase):
             self.graph, endpoints=False, normalized=False
         )
         expected = {0: 0.0, 1: 2.0, 2: 2.0, 4: 0.0}
+        self.assertEqual(expected, betweenness)
+
+    def test_closeness_centrality(self):
+        betweenness = retworkx.graph_closeness_centrality(self.graph)
+        expected = {0: 0.5, 1: 0.75, 2: 0.75, 4: 0.5}
+        self.assertEqual(expected, betweenness)
+
+    def test_closeness_centrality_wf_improved(self):
+        betweenness = retworkx.graph_closeness_centrality(self.graph, wf_improved=False)
+        expected = {0: 0.5, 1: 0.75, 2: 0.75, 4: 0.5}
         self.assertEqual(expected, betweenness)
