+++ tests/digraph/test_centrality.py
@@ -85,6 +85,16 @@ class TestCentralityDiGraph(unittest.TestCase):
         expected = {0: 0.0, 1: 2.0, 2: 2.0, 3: 0.0}
         self.assertEqual(expected, betweenness)
 
+    def test_closeness_centrality(self):
+        betweenness = retworkx.digraph_closeness_centrality(self.graph)
+        expected = {0: 0.0, 1: 1.0 / 3.0, 2: 4.0 / 9.0, 3: 0.5}
+        self.assertEqual(expected, betweenness)
+
+    def test_closeness_centrality_wf_improved(self):
+        betweenness = retworkx.digraph_closeness_centrality(self.graph, wf_improved=False)
+        expected = {0: 0.0, 1: 1.0, 2: 2.0 / 3.0, 3: 0.5}
+        self.assertEqual(expected, betweenness)
+
 
 class TestCentralityDiGraphDeletedNode(unittest.TestCase):
     def setUp(self):
@@ -127,4 +137,14 @@ class TestCentralityDiGraphDeletedNode(unittest.TestCase):
             self.graph, endpoints=False, normalized=False
         )
         expected = {0: 0.0, 1: 2.0, 2: 2.0, 4: 0.0}
+        self.assertEqual(expected, betweenness)
+
+    def test_closeness_centrality(self):
+        betweenness = retworkx.digraph_closeness_centrality(self.graph)
+        expected = {0: 0.0, 1: 1.0 / 3.0, 2: 4.0 / 9.0, 4: 0.5}
+        self.assertEqual(expected, betweenness)
+
+    def test_closeness_centrality_wf_improved(self):
+        betweenness = retworkx.digraph_closeness_centrality(self.graph, wf_improved=False)
+        expected = {0: 0.0, 1: 1.0, 2: 2.0 / 3.0, 4: 0.5}
         self.assertEqual(expected, betweenness)
