### 2022-04-23 fab350a
Add embedding functions
