### 2022-04-23 d409a07
add connected_components to core
