### 2022-04-25 bb74b5e
Add test_connectivity.rs
