### 2022-04-26 6fe44e5
Finish tests for connected components in core
