+++ retworkx-core/src/connectivity/test_connectivity.rs
@@ -10,20 +10,44 @@
 // License for the specific language governing permissions and limitations
 // under the License.
 
-use hashbrown::HashSet;
-use petgraph::graph::Graph;
-use petgraph::Undirected;
+#[cfg(test)]
+mod tests {
+    use hashbrown::HashSet;
+    use petgraph::graph::Graph;
+    use petgraph::graph::NodeIndex;
+    use petgraph::Undirected;
 
-use crate::connected_components;
+    use crate::connectivity::conn_components::{connected_components, number_connected_components};
 
-#[test]
-fn test_connected_components() {
-    let mut graph = Graph::<(), (), Undirected>::from_edges(&[
-        (0, 1), (1, 2), (2, 3), (3, 0), (4, 5), (5, 6), (6, 7), (7, 4)
-    ]);
-    let components = connected_components(&mut graph);
-    let exp1: HashSet<usize> = [0, 1, 3, 2].iter().cloned().collect();
-    let exp2: HashSet<usize> = [7, 5, 4, 6].iter().cloned().collect();
-    let expected: Vec<_> = vec![exp1, exp2];
-    assert_eq!(expected, components);
+    #[test]
+    fn test_number_connected() {
+        let mut graph = Graph::<(), (), Undirected>::from_edges([(0, 1), (1, 2), (3, 4)]);
+        assert_eq!(number_connected_components(&mut graph), 2);
+    }
+
+    #[test]
+    fn test_number_node_holes() {
+        let mut graph = Graph::<(), (), Undirected>::from_edges([(0, 1), (1, 2)]);
+        graph.remove_node(NodeIndex::new(1));
+        assert_eq!(number_connected_components(&mut graph), 2);
+    }
+
+    #[test]
+    fn test_connected_components() {
+        let mut graph = Graph::<(), (), Undirected>::from_edges(&[
+            (0, 1),
+            (1, 2),
+            (2, 3),
+            (3, 0),
+            (4, 5),
+            (5, 6),
+            (6, 7),
+            (7, 4),
+        ]);
+        let components = connected_components(&mut graph);
+        let exp1: HashSet<usize> = [0, 1, 3, 2].iter().cloned().collect();
+        let exp2: HashSet<usize> = [7, 5, 4, 6].iter().cloned().collect();
+        let expected: Vec<_> = vec![exp1, exp2];
+        assert_eq!(expected, components);
+    }
 }
