### 2022-04-26 94e3376
Add return all simple paths between all pairs of nodes (#577)
* Add return all simple paths between all pairs of nodes

This commit adds a new function all_pairs_all_simple_paths
which is used to return all simple paths between all nodes of a
particular length. It basically is a multithreaded loop around the
existing all_simple_paths function that runs the function between all
pairs of nodes in the graph.

* Fix function signature typos

* Typo fixes

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Raise a ValueError on valid input for min_depth or cutoff

* Add 
[…]
