### 2022-04-26 cb53f6a
Convert return from usize to NodeId
