+++ retworkx-core/src/connectivity/test_connectivity.rs
@@ -47,8 +47,12 @@ mod tests {
         let components = connected_components(&graph);
         let mut nodes1 = Vec::<NodeIndex>::new();
         let mut nodes2 = Vec::<NodeIndex>::new();
-        for i in [0, 1, 3, 2] {nodes1.push(NodeIndex::new(i));}
-        for i in [7, 5, 4, 6] {nodes2.push(NodeIndex::new(i));}
+        for i in [0, 1, 3, 2] {
+            nodes1.push(NodeIndex::new(i));
+        }
+        for i in [7, 5, 4, 6] {
+            nodes2.push(NodeIndex::new(i));
+        }
         let exp1: HashSet<NodeIndex> = nodes1.iter().cloned().collect();
         let exp2: HashSet<NodeIndex> = nodes2.iter().cloned().collect();
         let expected = vec![exp1, exp2];
