### 2022-04-27 8e780ac
Fix handling of negative length int index
Previously if an integer index was passed to `__getitem__`
with a value equal to the negative length of the sequence
we would incorrectly raise an except when instead the first
element should have been returned. This commit fixes the
handling of that edge case.

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>
