+++ retworkx-core/tests/centrality.rs
@@ -1,3 +1,15 @@
+// Licensed under the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License. You may obtain
+// a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations
+// under the License.
+
 use petgraph::visit::Reversed;
 use petgraph::Graph;
 use retworkx_core::centrality::closeness_centrality;
@@ -25,26 +37,26 @@ fn test_wf_improved() {
     let c = closeness_centrality(&g, true);
     assert_eq!(
         vec![
-            Some(1. / 4.),
-            Some(3. / 8.),
-            Some(3. / 8.),
-            Some(1. / 4.),
-            Some(2. / 9.),
-            Some(1. / 3.),
-            Some(2. / 9.)
+            Some(1.0 / 4.0),
+            Some(3.0 / 8.0),
+            Some(3.0 / 8.0),
+            Some(1.0 / 4.0),
+            Some(2.0 / 9.0),
+            Some(1.0 / 3.0),
+            Some(2.0 / 9.0)
         ],
         c
     );
     let cwf = closeness_centrality(&g, false);
     assert_eq!(
         vec![
-            Some(1. / 2.),
-            Some(3. / 4.),
-            Some(3. / 4.),
-            Some(1. / 2.),
-            Some(2. / 3.),
-            Some(1.),
-            Some(2. / 3.)
+            Some(1.0 / 2.0),
+            Some(3.0 / 4.0),
+            Some(3.0 / 4.0),
+            Some(1.0 / 2.0),
+            Some(2.0 / 3.0),
+            Some(1.0),
+            Some(2.0 / 3.0)
         ],
         cwf
     );
@@ -85,7 +97,7 @@ fn test_k5() {
 fn test_path() {
     let g = UnGraph::<i32, ()>::from_edges(&[(0, 1), (1, 2)]);
     let c = closeness_centrality(&g, true);
-    assert_eq!(vec![Some(2. / 3.), Some(1.), Some(2. / 3.)], c);
+    assert_eq!(vec![Some(2.0 / 3.0), Some(1.0), Some(2.0 / 3.0)], c);
 }
 
 #[test]
