+++ releasenotes/notes/closeness-centrality-459c5c7e35cb2e63.yaml
@@ -1,0 +1,29 @@
+---
+features:
+  - |
+    Added a new function, :func:`~retworkx.closeness_centrality` to compute
+    closeness centrality of all nodes in a :class:`~retworkx.PyGraph` or
+    :class:`~retworkx.PyDiGraph` object. 
+
+    The closeness centrality of a node :math:`u` is the reciprocal of the
+    average shortest path distance to :math:`u` over all :math:`n-1` reachable
+    nodes.
+  
+    .. math::
+  
+        C(u) = \frac{n - 1}{\sum_{v=1}^{n-1} d(v, u)},
+  
+    where :math:`d(v, u)` is the shortest-path distance between :math:`v` and
+    :math:`u`, and :math:`n` is the number of nodes that can reach :math:`u`.
+    
+    Wasserman and Faust propose an improved formula for graphs with more than
+    one connected component. The result is "a ratio of the fraction of actors
+    in the group who are reachable, to the average distance" from the reachable
+    actors.
+    
+    .. math::
+
+        C_{WF}(u) = \frac{n-1}{N-1} \frac{n - 1}{\sum_{v=1}^{n-1} d(v, u)},
+    
+    where :math:`N` is the number of nodes in the graph. By default, this is
+    enabled.
