### 2022-04-27 fa549b3
Fix handling of negative steps in slices
This commit fixes the handling of negative slices in sequence custom
return types. Previouly using a negative step would cause a panic
instead of correctly returning a reversed slice from the container. This
commit corrects this so we can deal with negative slice steps too.
