+++ docs/source/api.rst
@@ -155,7 +155,7 @@ Graph Operations
 ----------------
 
 .. autosummary::
-   :toctree: stubs
+   :toctree: apiref
 
    retworkx.complement
    retworkx.union
