+++ releasenotes/notes/slices-for-sequences-a3b31c70f5d896b4.yaml
@@ -12,7 +12,7 @@ fixes:
 
     now correctly handle slice inputs to ``__getitem__``. Previously if you
     tried to access a slice from one of these objects it would raise a
-    ``TypeError. For example, if you had a :class:`~.NodeIndices` object named
+    ``TypeError``. For example, if you had a :class:`~.NodeIndices` object named
     ``nodes`` containing ``[0, 1, 3, 4, 5]`` if you did something like::
 
         nodes[0:3]
