### 2022-04-28 26a1e29
Fix stray reference to old autosummary dir (#598) (#599)
In #544 we updated the autosummary directives output directory from
'stubs' to 'apiref' so that the urls of the hosted documentation made
more sense. Then a redirect was added to ensure that stale links to
stubs were redirected to the new location. However, as part of this
migration one autosummary directory was missed causing it to still
publish to 'stubs'. With the redirects in place this means that the
documentation is not actually reachable anymore. This commit fixes this
oversight by fixing the autosummary directory so the documenta
[…]
