+++ src/layout/planar.rs
@@ -56,21 +56,44 @@ impl<T> CwCcw<T> {
         self.cw.as_mut()
     }
 
-    fn cc_as_mut(&mut self) -> Option<&mut T> {
+    fn ccw_as_mut(&mut self) -> Option<&mut T> {
         self.ccw.as_mut()
     }
 }
 
-pub struct FirstNbr {
-    first_nbr: Option<NodeIndex>,
+pub struct FirstNbr<T> {
+    first_nbr: Option<T>,
 }
 
-impl Default for FirstNbr {
+impl<T> Default for FirstNbr<T>  {
     fn default() -> Self {
         FirstNbr { first_nbr: None }
     }
 }
 
+impl<T> FirstNbr<T> {
+    fn new(first_nbr: T) -> Self {
+        FirstNbr {
+            first_nbr: Some(first_nbr),
+        }
+    }
+
+    fn first_nbr_is_empty(&self) -> bool {
+        self.first_nbr.is_none()
+    }
+
+    fn first_nbr_unwrap(self) -> T {
+        self.first_nbr.unwrap()
+    }
+
+    fn first_nbr_as_ref(&mut self) -> Option<&T> {
+        self.first_nbr.as_ref()
+    }
+
+    fn first_nbr_as_mut(&mut self) -> Option<&mut T> {
+        self.first_nbr.as_mut()
+    }
+}
 pub fn planar_layout<Ty: EdgeType>(
     py: Python,
     graph: &StablePyGraph<Ty>,
@@ -115,7 +138,7 @@ pub fn is_planar<Ty: EdgeType>(
 
 pub fn create_embedding<Ty: EdgeType>(
     graph: &StablePyGraph<Ty>,
-    embedding: &mut Graph<FirstNbr, CwCcw<NodeIndex>, Directed>,
+    embedding: &mut Graph<FirstNbr<NodeIndex>, CwCcw<NodeIndex>, Directed>,
 ) -> bool {
     // DEBUG CODE FOR TESTING BASIC EMBEDDING
     // for v in graph.node_indices() {
@@ -130,7 +153,7 @@ pub fn create_embedding<Ty: EdgeType>(
 }
 
 pub fn combinitorial_embedding_to_pos (
-    embedding: &Graph<FirstNbr, CwCcw<NodeIndex>, Directed>,
+    embedding: &Graph<FirstNbr<NodeIndex>, CwCcw<NodeIndex>, Directed>,
     pos: &mut Vec<Point>,
 ){
     if embedding.node_count() < 4 {
@@ -143,7 +166,7 @@ pub fn combinitorial_embedding_to_pos (
 }
 
 pub fn triangulate_embedding (
-    embedding: &Graph<FirstNbr, CwCcw<NodeIndex>, Directed>,
+    embedding: &Graph<FirstNbr<NodeIndex>, CwCcw<NodeIndex>, Directed>,
     fully_triangulate: bool,
 ) -> Vec<NodeIndex> {
     if embedding.node_count() <= 1 {
@@ -156,20 +179,20 @@ pub fn triangulate_embedding (
 }
 
 struct PlanarEmbedding {
-    embedding: Graph::<FirstNbr, CwCcw<NodeIndex>, Directed>,
+    embedding: Graph::<FirstNbr<NodeIndex>, CwCcw<NodeIndex>, Directed>,
 }
 
 impl Default for PlanarEmbedding {
     fn default () -> Self {
         PlanarEmbedding {
-            embedding: Graph::<FirstNbr, CwCcw<NodeIndex>, Directed>::new(),
+            embedding: Graph::<FirstNbr<NodeIndex>, CwCcw<NodeIndex>, Directed>::new(),
         }
     }
 }
[…]
