### 2022-04-29 af5b5ff
Finish doc tests and set up core test dir
