### 2022-04-30 816acd4
fix implementation of tensor product for undirected graphs
