+++ retworkx-core/src/connectivity/mod.rs
@@ -22,3 +22,4 @@ pub use biconnected::articulation_points;
 pub use chain::chain_decomposition;
 pub use conn_components::connected_components;
 pub use conn_components::number_connected_components;
+pub use conn_components::bfs_undirected;
