+++ retworkx-core/src/shortest_path/mod.rs
@@ -16,9 +16,11 @@
 //! shortest path of a graph.
 
 mod astar;
+mod bellman_ford;
 mod dijkstra;
 mod k_shortest_path;
 
 pub use astar::astar;
+pub use bellman_ford::bellman_ford;
 pub use dijkstra::dijkstra;
 pub use k_shortest_path::k_shortest_path;
