+++ tests/digraph/test_bellman_ford.py
@@ -1,0 +1,165 @@
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+import unittest
+
+import retworkx
+
+
+class TestBellmanFordDiGraph(unittest.TestCase):
+    def setUp(self):
+        self.graph = retworkx.PyDiGraph()
+        self.a = self.graph.add_node("A")
+        self.b = self.graph.add_node("B")
+        self.c = self.graph.add_node("C")
+        self.d = self.graph.add_node("D")
+        self.e = self.graph.add_node("E")
+        self.f = self.graph.add_node("F")
+        edge_list = [
+            (self.a, self.b, 7),
+            (self.c, self.a, 9),
+            (self.a, self.d, 14),
+            (self.b, self.c, 10),
+            (self.d, self.c, 2),
+            (self.d, self.e, 9),
+            (self.b, self.f, 15),
+            (self.c, self.f, 11),
+            (self.e, self.f, 6),
+        ]
+        self.graph.add_edges_from(edge_list)
+
+    def test_bellman_ford(self):
+        path = retworkx.digraph_bellman_ford_shortest_path_lengths(
+            self.graph, self.a, lambda x: float(x)
+        )
+        expected = {1: 7.0, 2: 16.0, 3: 14.0, 4: 23.0, 5: 22.0}
+        self.assertEqual(expected, path)
+
+    def test_bellman_ford_length_with_no_path(self):
+        g = retworkx.PyDiGraph()
+        a = g.add_node("A")
+        b = g.add_node("B")
+        path_lenghts = retworkx.digraph_bellman_ford_shortest_path_lengths(g, a, edge_cost_fn=float)
+        expected = {}
+        self.assertEqual(expected, path_lenghts)
+
+    def test_bellman_ford_path(self):
+        paths = retworkx.digraph_bellman_ford_shortest_paths(self.graph, self.a)
+        expected = {
+            # a -> b
+            1: [0, 1],
+            # a -> c: a, d, c
+            2: [0, 3, 2],
+            # a -> d
+            3: [0, 3],
+            # a -> e: a, d, e
+            4: [0, 3, 4],
+            # a -> f: a, b, f
+            5: [0, 1, 5],
+        }
+        self.assertEqual(expected, paths)
+
+    def test_bellman_ford_path_with_weight_fn(self):
+        paths = retworkx.digraph_bellman_ford_shortest_paths(
+            self.graph, self.a, weight_fn=lambda x: x
+        )
+        expected = {
+            1: [0, 1],
+            2: [0, 3, 2],
+            3: [0, 3],
+            4: [0, 3, 4],
+            5: [0, 1, 5],
+        }
+        self.assertEqual(expected, paths)
+
+    def test_bellman_ford_path_undirected(self):
+        paths = retworkx.digraph_bellman_ford_shortest_paths(self.graph, self.a, as_undirected=True)
+        expected = {
+            1: [0, 1],
+            2: [0, 2],
+            3: [0, 3],
+            4: [0, 3, 4],
+            5: [0, 1, 5],
+        }
+        self.assertEqual(expected, paths)
+
+    def test_bellman_ford_path_undirected_with_weight_fn(self):
[…]
