+++ tests/graph/test_bellman_ford.py
@@ -1,0 +1,138 @@
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+import unittest
+
+import retworkx
+
+
+class TestBellmanFordGraph(unittest.TestCase):
+    def setUp(self):
+        self.graph = retworkx.PyGraph()
+        self.a = self.graph.add_node("A")
+        self.b = self.graph.add_node("B")
+        self.c = self.graph.add_node("C")
+        self.d = self.graph.add_node("D")
+        self.e = self.graph.add_node("E")
+        self.f = self.graph.add_node("F")
+        self.graph.add_edge(self.a, self.b, 7)
+        self.graph.add_edge(self.c, self.a, 9)
+        self.graph.add_edge(self.a, self.d, 14)
+        self.graph.add_edge(self.b, self.c, 10)
+        self.graph.add_edge(self.d, self.c, 2)
+        self.graph.add_edge(self.d, self.e, 9)
+        self.graph.add_edge(self.b, self.f, 15)
+        self.graph.add_edge(self.c, self.f, 11)
+        self.graph.add_edge(self.e, self.f, 6)
+
+    def test_bellman_ford(self):
+        path = retworkx.graph_bellman_ford_shortest_path_lengths(
+            self.graph, self.a, lambda x: float(x)
+        )
+        path_dijkstra = retworkx.graph_dijkstra_shortest_path_lengths(
+            self.graph, self.a, lambda x: float(x)
+        )
+        print(path)
+        print(path_dijkstra)
+        self.assertEqual(path_dijkstra, path)
+
+    def test_bellman_ford_path(self):
+        path = retworkx.graph_bellman_ford_shortest_paths(
+            self.graph, self.a, weight_fn=lambda x: float(x)
+        )
+        # a -> d -> e = 23
+        # a -> c -> d -> e = 20
+        expected = retworkx.graph_dijkstra_shortest_paths(
+            self.graph, self.a, weight_fn=lambda x: float(x)
+        )
+        self.assertEqual(expected, path)
+
+    def test_bellman_ford_with_no_goal_set(self):
+        path = retworkx.graph_bellman_ford_shortest_path_lengths(self.graph, self.a, lambda x: 1)
+        expected = {1: 1.0, 2: 1.0, 3: 1.0, 4: 2.0, 5: 2.0}
+        self.assertEqual(expected, path)
+
+    def test_bellman_ford_length_with_no_path(self):
+        g = retworkx.PyGraph()
+        a = g.add_node("A")
+        b = g.add_node("B")
+        path_lenghts = retworkx.graph_bellman_ford_shortest_path_lengths(g, a, edge_cost_fn=float)
+        expected = {}
+        self.assertEqual(expected, path_lenghts)
+
+    def test_bellman_ford_path_with_no_goal_set(self):
+        path = retworkx.graph_bellman_ford_shortest_paths(self.graph, self.a)
+        expected = {
+            1: [0, 1],
+            2: [0, 2],
+            3: [0, 3],
+            4: [0, 3, 4],
+            5: [0, 1, 5],
+        }
+        self.assertEqual(expected, path)
+
+    def test_bellman_ford_with_no_path(self):
+        g = retworkx.PyGraph()
+        a = g.add_node("A")
+        g.add_node("B")
[…]
