### 2022-05-03 6bf4d35
Fix bug in cycle detection
