### 2022-05-03 529d553
Cache edge weights
