### 2022-05-03 186dd0d
Handle negative cycles
