### 2022-05-03 a55a79b
Add negative_edge_cycle
