+++ retworkx-core/src/shortest_path/mod.rs
@@ -21,6 +21,6 @@ mod dijkstra;
 mod k_shortest_path;
 
 pub use astar::astar;
-pub use bellman_ford::bellman_ford;
+pub use bellman_ford::{bellman_ford, negative_cycle_finder};
 pub use dijkstra::dijkstra;
 pub use k_shortest_path::k_shortest_path;
