+++ retworkx-core/src/shortest_path/bellman_ford.rs
@@ -11,7 +11,6 @@
 // under the License.
 
 use std::collections::VecDeque;
-use std::error::Error;
 use std::hash::Hash;
 
 use fixedbitset::FixedBitSet;
@@ -39,7 +38,6 @@ where
     F: FnMut(G::EdgeRef) -> Result<K, E>,
     K: Measure + Copy,
     S: DistanceMap<G::NodeId, K>,
-    E: Error,
 {
     let node_count = graph.node_count();
     let mut in_queue = FixedBitSet::with_capacity(graph.node_bound());
@@ -104,6 +102,70 @@ where
     Ok(Some(scores))
 }
 
+/// Find negative cycle
+/// using SPFA
+pub fn negative_cycle_finder<G, F, K, E>(
+    graph: G,
+    mut edge_cost: F,
+) -> Result<Option<Vec<G::NodeId>>, E>
+where
+    G: IntoEdges + Visitable + NodeIndexable + NodeCount + IntoNodeIdentifiers,
+    G::NodeId: Eq + Hash + IndexType,
+    F: FnMut(G::EdgeRef) -> Result<K, E>,
+    K: Measure + Copy,
+{
+    let node_count = graph.node_count();
+    let mut in_queue = FixedBitSet::with_capacity(graph.node_bound());
+    let zero_score = K::default();
+    let mut scores: Vec<K> = vec![zero_score; graph.node_bound()];
+    let mut predecessor: Vec<Option<G::NodeId>> = vec![None; graph.node_bound()];
+    let mut visit_next = VecDeque::with_capacity(graph.node_bound());
+    let mut relaxation_count: usize = 0;
+
+    for node in graph.node_identifiers() {
+        visit_next.push_back(node);
+        in_queue.set(node.index(), true);
+    }
+
+    while let Some(node) = visit_next.pop_front() {
+        in_queue.set(node.index(), false);
+        let node_score = scores[node.index()];
+
+        for edge in graph.edges(node) {
+            let next = edge.target();
+            let current_score = scores[next.index()];
+            let cost = edge_cost(edge)?;
+            let next_score = node_score + cost;
+
+            if next_score < current_score {
+                scores[next.index()] = next_score;
+                predecessor[next.index()] = Some(node);
+                relaxation_count += 1;
+
+                if relaxation_count == node_count {
+                    relaxation_count = 0;
+
+                    if check_for_negative_cycle(
+                        predecessor.iter().map(|x| x.map(|y| y.index())).collect(),
+                    ) {
+                        return Ok(Some(recover_negative_cycle_from_predecessors(
+                            graph,
+                            predecessor,
+                        )));
+                    }
+                }
+
+                if !in_queue.contains(next.index()) {
+                    visit_next.push_back(next);
+                    in_queue.set(next.index(), true);
+                }
+            }
+        }
+    }
+
+    Ok(None)
+}
+
 fn check_for_negative_cycle(predecessor: Vec<Option<usize>>) -> bool {
     let mut path_graph =
         StableDiGraph::<usize, ()>::with_capacity(predecessor.len(), predecessor.len());
@@ -119,4 +181,16 @@ fn check_for_negative_cycle(predecessor: Vec<Option<usize>>) -> bool {
     }
 
     is_cyclic_directed(&path_graph)
+}
+
+fn recover_negative_cycle_from_predecessors<G>(
+    _graph: G,
+    _predecessor: Vec<Option<G::NodeId>>,
+) -> Vec<G::NodeId>
+where
+    G: IntoEdges + Visitable + NodeIndexable + NodeCount + IntoNodeIdentifiers,
+    G::NodeId: Eq + Hash + IndexType,
+{
+    // TODO: replace with actual code to find cycle
+    Vec::new()
 }
