+++ tests/digraph/test_bellman_ford.py
@@ -47,7 +47,7 @@ class TestBellmanFordDiGraph(unittest.TestCase):
     def test_bellman_ford_length_with_no_path(self):
         g = retworkx.PyDiGraph()
         a = g.add_node("A")
-        b = g.add_node("B")
+        g.add_node("B")
         path_lenghts = retworkx.digraph_bellman_ford_shortest_path_lengths(g, a, edge_cost_fn=float)
         expected = {}
         self.assertEqual(expected, path_lenghts)
