### 2022-05-03 d01e666
support for a multigraph
