+++ Cargo.toml
@@ -36,7 +36,7 @@ version = "0.16.4"
 features = ["extension-module", "hashbrown", "num-bigint", "num-complex", "indexmap"]
 
 [dependencies.hashbrown]
-version = "0.11"
+version = "0.12"
 features = ["rayon"]
 
 [dependencies.ndarray]
