+++ retworkx-core/Cargo.toml
@@ -16,7 +16,7 @@ petgraph = "0.6.0"
 rayon = "1.5"
 
 [dependencies.hashbrown]
-version = "0.11"
+version = "0.12"
 features = ["rayon"]
 
 [dependencies.indexmap]
