### 2022-05-03 54e3734
Avoid code duplication
