### 2022-05-03 8634f1c
Add new functions to documentation
