+++ docs/source/api.rst
@@ -31,6 +31,9 @@ Shortest Paths
    retworkx.dijkstra_shortest_path_lengths
    retworkx.all_pairs_dijkstra_shortest_paths
    retworkx.all_pairs_dijkstra_path_lengths
+   retworkx.bellman_ford_shortest_paths
+   retworkx.bellman_ford_shortest_path_lengths
+   retworkx.negative_edge_cycle
    retworkx.distance_matrix
    retworkx.floyd_warshall
    retworkx.floyd_warshall_numpy
@@ -273,6 +276,8 @@ the functions from the explicitly typed based on the data type.
    retworkx.digraph_all_pairs_dijkstra_shortest_paths
    retworkx.digraph_dijkstra_shortest_path_lengths
    retworkx.digraph_all_pairs_dijkstra_path_lengths
+   retworkx.digraph_bellman_ford_shortest_path_lengths
+   retworkx.digraph_bellman_ford_shortest_path_lengths
    retworkx.digraph_k_shortest_path_lengths
    retworkx.digraph_dfs_edges
    retworkx.digraph_dfs_search
@@ -323,6 +328,8 @@ typed API based on the data type.
    retworkx.graph_all_pairs_dijkstra_shortest_paths
    retworkx.graph_k_shortest_path_lengths
    retworkx.graph_all_pairs_dijkstra_path_lengths
+   retworkx.graph_bellman_ford_shortest_path_lengths
+   retworkx.graph_bellman_ford_shortest_path_lengths
    retworkx.graph_dfs_edges
    retworkx.graph_dfs_search
    retworkx.graph_transitivity
