### 2022-05-03 4e9355e
Add find_negative_cycle
