+++ src/lib.rs
@@ -381,6 +381,7 @@ fn retworkx(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_wrapped(wrap_pyfunction!(graph_bellman_ford_shortest_path_lengths))?;
     m.add_wrapped(wrap_pyfunction!(digraph_bellman_ford_shortest_path_lengths))?;
     m.add_wrapped(wrap_pyfunction!(negative_edge_cycle))?;
+    m.add_wrapped(wrap_pyfunction!(find_negative_cycle))?;
     m.add_wrapped(wrap_pyfunction!(digraph_all_pairs_dijkstra_path_lengths))?;
     m.add_wrapped(wrap_pyfunction!(digraph_all_pairs_dijkstra_shortest_paths))?;
     m.add_wrapped(wrap_pyfunction!(graph_all_pairs_dijkstra_path_lengths))?;
