### 2022-05-04 e2f1e6a
Document retworkx-core code
