+++ retworkx-core/src/shortest_path/bellman_ford.rs
@@ -26,8 +26,76 @@ use petgraph::visit::{
 use crate::dictmap::*;
 use crate::distancemap::DistanceMap;
 
-/// Bellman-Ford shortest path algorithm
-/// using SPFA
+/// Bellman-Ford shortest path algorithm with the SPFA heuristic.
+///
+/// Compute the length of the shortest path from `start` to every reachable
+/// node.
+///
+/// The graph should be [`Visitable`] and implement [`IntoEdges`]. The function
+/// `edge_cost` should return the cost for a particular edge, which is used
+/// to compute path costs. Edge costs can be negative, as long as there is no
+/// negative cycle.
+///
+///
+/// If `path` is not [`None`], then the algorithm will mutate the input
+/// [`DictMap`] to insert an entry where the index is the dest node index
+/// the value is a Vec of node indices of the path starting with `start` and
+/// ending at the index.
+///
+/// Returns a [`DistanceMap`] that maps `NodeId` to path cost if there are no
+/// negative cycles. If there are negative cycles, the Result is None.
+/// # Example
+/// ```rust
+/// use retworkx_core::petgraph::Graph;
+/// use retworkx_core::petgraph::prelude::*;
+/// use retworkx_core::dictmap::DictMap;
+/// use retworkx_core::shortest_path::bellman_ford;
+/// use retworkx_core::Result;
+///
+/// let mut graph : Graph<(),(),Directed>= Graph::new();
+/// let a = graph.add_node(()); // node with no weight
+/// let b = graph.add_node(());
+/// let c = graph.add_node(());
+/// let d = graph.add_node(());
+/// let e = graph.add_node(());
+/// let f = graph.add_node(());
+/// let g = graph.add_node(());
+/// let h = graph.add_node(());
+/// // z will be in another connected component
+/// let z = graph.add_node(());
+///
+/// graph.extend_with_edges(&[
+///     (a, b),
+///     (b, c),
+///     (c, d),
+///     (d, a),
+///     (e, f),
+///     (b, e),
+///     (f, g),
+///     (g, h),
+///     (h, e)
+/// ]);
+/// // a ----> b ----> e ----> f
+/// // ^       |       ^       |
+/// // |       v       |       v
+/// // d <---- c       h <---- g
+///
+/// let expected_res: DictMap<NodeIndex, usize> = [
+///      (a, 3),
+///      (b, 0),
+///      (c, 1),
+///      (d, 2),
+///      (e, 1),
+///      (f, 2),
+///      (g, 3),
+///      (h, 4)
+///     ].iter().cloned().collect();
+/// let res: Result<Option<DictMap<NodeIndex, usize>>> = bellman_ford(
+///     &graph, b, |_| Ok(1), None
+/// );
+/// assert_eq!(res.unwrap().unwrap(), expected_res);
+/// // z is not inside res because there is not path from b to z.
+/// ```
 pub fn bellman_ford<G, F, K, E, S>(
     graph: G,
     start: G::NodeId,
@@ -52,6 +120,7 @@ where
     scores.put_item(start, zero_score);
     visit_next.push_back(start);
 
+    // SPFA heuristic: relax only nodes that need to be relaxed
     while let Some(node) = visit_next.pop_front() {
         in_queue.set(node.index(), false);
         let node_score = *scores.get_item(node).unwrap();
@@ -67,6 +136,8 @@ where
                 predecessor[next.index()] = Some(node);
                 relaxation_count += 1;
 
+                // We do the negative cycle check every O(|V|)
+                // iterations to amortize the cost, as it costs O(|V|) to run it
                 if relaxation_count == node_count {
                     relaxation_count = 0;
 
@@ -77,6 +148,7 @@ where
                     }
                 }
 
[…]
