### 2022-05-04 fe6b0db
More PlanarEmbedding functions
