+++ src/layout/planar.rs
@@ -211,25 +211,25 @@ impl PlanarEmbedding {
         ref_nbr: Option<NodeIndex>
     ){
         let cw_weight = CwCcw::<NodeIndex>::default();
+        let first_nbr = FirstNbr::<NodeIndex>::default();
         let new_edge = self.embedding.add_edge(start_node, end_node, cw_weight);
         if ref_nbr.is_none() {
             // The start node has no neighbors
-            //cw_weight.cw = Some(end_node.clone());
-            //self.embedding.update_edge(start_node, end_node, cw_weight);
+            self.update_edge_weight(start_node, end_node, end_node, true);
             self.update_edge_weight(start_node, end_node, end_node, false);
-            // self.embedding[new_edge].ccw = Some(end_node.clone());
-            // self.embedding[start_node].first_nbr = Some(end_node.clone());
-            // return
+            self.embedding[start_node].first_nbr = Some(end_node);
+            return
         }
-        // // if ref_nbr not in self[start_node] error
-        // let ref_nbr_edge = self.embedding.find_edge(start_node, ref_nbr.unwrap().clone());
-        // let cw_ref_node = ref_nbr_edge.unwrap().cw;
-        // let cw_ref_edge = self.embedding.find_edge(start_node, cw_ref_node);
+        // if ref_nbr not in self[start_node] error
+        let ref_nbr_node = ref_nbr.unwrap();
+        let cw_ref_edge = self.embedding.find_edge(start_node, ref_nbr_node).unwrap();
+        let cw_ref_node = self.embedding.edge_weight_mut(cw_ref_edge).unwrap().cw.unwrap();
 
-        // self.embedding[ref_nbr_edge].cw = Some(end_node.clone());
-        // self.embedding[new_edge].cw = Some(cw_ref_node.clone());
-        // self.embedding[cw_ref_edge].ccw = Some(end_node.clone());
-        // self.embedding[new_edge].ccw = ref_nbr.unwrap().clone();
+        // Alter half-edge data structures
+        self.update_edge_weight(start_node, ref_nbr_node, end_node, true);
+        self.update_edge_weight(start_node, end_node, cw_ref_node, true);
+        self.update_edge_weight(start_node, cw_ref_node, end_node, false);
+        self.update_edge_weight(start_node, end_node, ref_nbr_node, false);
     }
 
     fn add_half_edge_ccw (
@@ -238,26 +238,31 @@ impl PlanarEmbedding {
         end_node: NodeIndex,
         ref_nbr: Option<NodeIndex>
     ){
-        let mut cw_weight = CwCcw::<NodeIndex>::default();
-        let new_edge = self.embedding.add_edge(start_node, end_node, cw_weight);
-        if !ref_nbr.is_none() {
-            self.embedding[new_edge].cw = Some(end_node.clone());
-            self.embedding[new_edge].ccw = Some(end_node.clone());
-            self.embedding[start_node].first_nbr = Some(end_node.clone());
-            return
+        if ref_nbr.is_none() {
+            let cw_weight = CwCcw::<NodeIndex>::default();
+            let new_edge = self.embedding.add_edge(start_node, end_node, cw_weight);
+            self.update_edge_weight(start_node, end_node, end_node, true);
+            self.update_edge_weight(start_node, end_node, end_node, false);
+            self.embedding[start_node].first_nbr = Some(end_node);
         } else {
-            return
-            //let ccw_ref = self.embedding.find_edge(start_node, ref_nbr.unwrap().clone()).ccw;
-            //self.add_half_edge_cw(start_node, end_node, ccw_ref);
-
[…]
