### 2022-05-04 7a28716
Try to fix docstrings
