### 2022-05-04 731d8da
Use example with negative weights
