### 2022-05-04 0819535
Use predecessors to build cycle
