### 2022-05-05 e7a7cfc
Add goal and target to functions
