+++ src/shortest_path/mod.rs
@@ -1208,13 +1208,16 @@ pub fn graph_unweighted_average_shortest_path_length(
 /// :param edge_cost_fn: A python callable that will take in 1 parameter, an
 ///     edge's data object and will return a float that represents the
 ///     cost/weight of that edge. It can be negative.
+/// :param int goal: An optional node index to use as the end of the path.
+///     When specified the output dictionary will only have a single entry with
+///     the length of the shortest path to the goal node.
 ///
 /// :returns: A read-only dictionary of the shortest paths from the provided node where
 ///     the key is the node index of the end of the path and the value is the
 ///     cost/sum of the weights of path
 /// :rtype: PathLengthMapping
 ///
-/// :raises :class:`~retworkx.NegativeCycle`: when there is a negative cycle and the shortest
+/// :raises: :class:`~retworkx.NegativeCycle`: when there is a negative cycle and the shortest
 ///     path is not defined.
 #[pyfunction]
 #[pyo3(text_signature = "(graph, node, edge_cost_fn, /)")]
@@ -1223,6 +1226,7 @@ pub fn digraph_bellman_ford_shortest_path_lengths(
     graph: &digraph::PyDiGraph,
     node: usize,
     edge_cost_fn: PyObject,
+    goal: Option<usize>,
 ) -> PyResult<PathLengthMapping> {
     let edge_weights: Vec<Option<f64>> =
         edge_weights_from_callable(py, &graph.graph, &Some(edge_cost_fn), 1.0)?;
@@ -1245,6 +1249,19 @@ pub fn digraph_bellman_ford_shortest_path_lengths(
     }
 
     let res = res.unwrap();
+
+    if let Some(goal_usize) = goal {
+        return Ok(PathLengthMapping {
+            path_lengths: match res[goal_usize] {
+                Some(goal_length) => {
+                    let mut ans = DictMap::new();
+                    ans.insert(goal_usize, goal_length);
+                    ans
+                }
+                None => DictMap::new(),
+            },
+        });
+    }
 
     Ok(PathLengthMapping {
         path_lengths: res
@@ -1270,13 +1287,16 @@ pub fn digraph_bellman_ford_shortest_path_lengths(
 /// :param edge_cost_fn: A python callable that will take in 1 parameter, an
 ///     edge's data object and will return a float that represents the
 ///     cost/weight of that edge. It can be negative.
+/// :param int goal: An optional node index to use as the end of the path.
+///     When specified the output dictionary will only have a single entry with
+///     the length of the shortest path to the goal node.
 ///
 /// :returns: A read-only dictionary of the shortest paths from the provided node where
 ///     the key is the node index of the end of the path and the value is the
 ///     cost/sum of the weights of path
 /// :rtype: PathLengthMapping
 ///
-/// :raises :class:`~retworkx.NegativeCycle`: when there is a negative cycle and the shortest
+/// :raises: :class:`~retworkx.NegativeCycle`: when there is a negative cycle and the shortest
 ///     path is not defined.
 #[pyfunction]
 #[pyo3(text_signature = "(graph, node, edge_cost_fn, /)")]
@@ -1285,6 +1305,7 @@ pub fn graph_bellman_ford_shortest_path_lengths(
     graph: &graph::PyGraph,
     node: usize,
     edge_cost_fn: PyObject,
+    goal: Option<usize>,
 ) -> PyResult<PathLengthMapping> {
     let edge_weights: Vec<Option<f64>> =
         edge_weights_from_callable(py, &graph.graph, &Some(edge_cost_fn), 1.0)?;
@@ -1308,6 +1329,19 @@ pub fn graph_bellman_ford_shortest_path_lengths(
 
     let res = res.unwrap();
 
+    if let Some(goal_usize) = goal {
[…]
