+++ tests/digraph/test_bellman_ford.py
@@ -110,6 +110,34 @@ class TestBellmanFordDiGraph(unittest.TestCase):
         expected = {1: 1.0, 2: 2.0, 3: 1.0, 4: 2.0, 5: 2.0}
         self.assertEqual(expected, path)
 
+    def test_bellman_path(self):
+        path = retworkx.digraph_bellman_ford_shortest_paths(
+            self.graph, self.a, weight_fn=lambda x: float(x), target=self.e
+        )
+        expected = retworkx.digraph_dijkstra_shortest_paths(
+            self.graph, self.a, weight_fn=lambda x: float(x), target=self.e
+        )
+        self.assertEqual(expected, path)
+
+    def test_bellman_path_lengths(self):
+        path = retworkx.digraph_bellman_ford_shortest_path_lengths(
+            self.graph, self.a, lambda x: float(x), goal=self.e
+        )
+        expected = retworkx.digraph_dijkstra_shortest_path_lengths(
+            self.graph, self.a, lambda x: float(x), goal=self.e
+        )
+        self.assertEqual(expected, path)
+
+    def test_bellman_ford_length_with_no_path_and_goal(self):
+        g = retworkx.PyDiGraph()
+        a = g.add_node("A")
+        b = g.add_node("B")
+        path_lenghts = retworkx.digraph_bellman_ford_shortest_path_lengths(
+            g, a, edge_cost_fn=float, goal=b
+        )
+        expected = retworkx.digraph_dijkstra_shortest_path_lengths(g, a, edge_cost_fn=float, goal=b)
+        self.assertEqual(expected, path_lenghts)
+
     def test_bellman_ford_with_no_path(self):
         g = retworkx.PyDiGraph()
         a = g.add_node("A")
