### 2022-05-05 fe24cd4
Reduce code duplication in Bellman-Ford
