+++ retworkx-core/src/shortest_path/bellman_ford.rs
@@ -25,10 +25,21 @@ use petgraph::visit::{
 use crate::dictmap::*;
 use crate::distancemap::DistanceMap;
 
+struct BellmanFordData<G, S>
+where
+    G: IntoEdges + Visitable + NodeIndexable + NodeCount + IntoNodeIdentifiers,
+{
+    negative_cycle: bool,
+    scores: S,
+    predecessor: Vec<Option<G::NodeId>>,
+}
+
 /// Bellman-Ford shortest path algorithm with the SPFA heuristic.
 ///
 /// Compute the length of the shortest path from `start` to every reachable
-/// node.
+/// node. This implementation differs from petgraph's implementation because it
+/// has an expected time complexity of O(kE) with 1 < k <= |V|. For random
+/// graphs, we expect the SPFA heuristic to be very efficient.
 ///
 /// The graph should be [`Visitable`] and implement [`IntoEdges`]. The function
 /// `edge_cost` should return the cost for a particular edge, which is used
@@ -98,7 +109,7 @@ use crate::distancemap::DistanceMap;
 pub fn bellman_ford<G, F, K, E, S>(
     graph: G,
     start: G::NodeId,
-    mut edge_cost: F,
+    edge_cost: F,
     mut path: Option<&mut DictMap<G::NodeId, Vec<G::NodeId>>>,
 ) -> Result<Option<S>, E>
 where
@@ -108,50 +119,17 @@ where
     K: Measure + Copy,
     S: DistanceMap<G::NodeId, K>,
 {
-    let node_count = graph.node_count();
-    let mut in_queue = FixedBitSet::with_capacity(graph.node_bound());
-    let mut scores: S = S::build(graph.node_bound());
-    let mut predecessor: Vec<Option<G::NodeId>> = vec![None; graph.node_bound()];
-    let mut visit_next = VecDeque::with_capacity(graph.node_bound());
-    let zero_score = K::default();
-    let mut relaxation_count: usize = 0;
+    let res: BellmanFordData<G, S> = inner_bellman_ford(graph, vec![start], edge_cost)?;
 
-    scores.put_item(start, zero_score);
-    visit_next.push_back(start);
+    let BellmanFordData {
+        negative_cycle,
+        scores,
+        predecessor,
+    } = res;
 
-    // SPFA heuristic: relax only nodes that need to be relaxed
-    while let Some(node) = visit_next.pop_front() {
-        in_queue.set(node.index(), false);
-        let node_score = *scores.get_item(node).unwrap();
-
-        for edge in graph.edges(node) {
-            let next = edge.target();
-            let current_score = scores.get_item(next);
-            let cost = edge_cost(edge)?;
-            let next_score = node_score + cost;
-
-            if current_score.is_none() || next_score < *current_score.unwrap() {
-                scores.put_item(next, next_score);
-                predecessor[next.index()] = Some(node);
-                relaxation_count += 1;
-
-                // We do the negative cycle check every O(|V|)
-                // iterations to amortize the cost, as it costs O(|V|) to run it
-                if relaxation_count == node_count {
-                    relaxation_count = 0;
-
-                    if check_for_negative_cycle(graph, &predecessor) {
-                        return Ok(None);
-                    }
-                }
-
-                // Node needs to be relaxed on a future iteration
-                if !in_queue.contains(next.index()) {
-                    visit_next.push_back(next);
-                    in_queue.set(next.index(), true);
-                }
-            }
-        }
+    // Shortest path is not defined
+    if negative_cycle {
+        return Ok(None);
     }
 
     // Build path from predecessors
@@ -224,7 +202,7 @@ where
 /// ```
 pub fn negative_cycle_finder<G, F, K, E>(
     graph: G,
[…]
