+++ releasenotes/notes/bellman-ford-ebeb580222000f56.yaml
@@ -8,4 +8,4 @@ features:
      * :func:`retworkx.find_negative_cycle`
      * :func:`retworkx.bellman_ford_path_lengths`
      * :func:`retworkx.bellman_ford_shortest_paths`
-     * :func:`retworkx.edge_negative_cycle`
+     * :func:`retworkx.negative_edge_cycle`
