+++ retworkx-core/tests/centrality.rs
@@ -11,7 +11,6 @@
 // under the License.
 
 use petgraph::visit::Reversed;
-use petgraph::Graph;
 use retworkx_core::centrality::closeness_centrality;
 use retworkx_core::petgraph::graph::{DiGraph, UnGraph};
 
@@ -98,27 +97,4 @@ fn test_path() {
     let g = UnGraph::<i32, ()>::from_edges(&[(0, 1), (1, 2)]);
     let c = closeness_centrality(&g, true);
     assert_eq!(vec![Some(2.0 / 3.0), Some(1.0), Some(2.0 / 3.0)], c);
-}
-
-#[test]
-fn test_weighted_closeness() {
-    let mut g = Graph::new();
-    let s = g.add_node(0);
-    let u = g.add_node(0);
-    let x = g.add_node(0);
-    let v = g.add_node(0);
-    let y = g.add_node(0);
-    g.add_edge(s, u, 10.);
-    g.add_edge(s, x, 5.);
-    g.add_edge(u, v, 1.);
-    g.add_edge(u, x, 2.);
-    g.add_edge(v, y, 1.);
-    g.add_edge(x, u, 3.);
-    g.add_edge(x, v, 5.);
-    g.add_edge(x, y, 2.);
-    g.add_edge(y, s, 7.);
-    g.add_edge(y, v, 6.);
-    let c = closeness_centrality(&g, true);
-    println!("{:?}", c);
-    assert_eq!(0, 0)
 }
