+++ tox.ini
@@ -76,10 +76,10 @@ commands = black {posargs} '../retworkx' '../tests'
 basepython = python3
 envdir = .tox/stubs
 deps =
-  pytest~=6.2
-  mypy==0.931
-  pytest-mypy-testing==0.0.8
-commands = python -m pytest '../stubs-tests'
+  pytest~=7.1
+  mypy==0.950
+  pytest-mypy-testing==0.0.10
+commands = python -m pytest '../tests/stubs-tests'
 
 [flake8]
 # E125 is deliberately excluded. See https://github.com/jcrocholl/pep8/issues/126
