+++ tox.ini
@@ -87,7 +87,9 @@ commands = python -m pytest '../tests/stubs-tests'
 # E129 skipped because it is too limiting when combined with other rules
 # E711 skipped because sqlalchemy filter() requires using == instead of is
 # max-line-length, E203, W503 are added for black compatibility
+# builtins includes reveal_type because mypy injects it
 max-line-length = 110
 ignore = E125,E123,E129,E711
 extend-ignore = E203, W503
 exclude = .venv,.git,.tox,dist,doc,*egg,build
+builtins=reveal_type
