+++ tox.ini
@@ -79,7 +79,7 @@ deps =
   pytest~=7.1
   mypy==0.950
   pytest-mypy-testing==0.0.10
-commands = python -m pytest '../tests/stubs-tests'
+commands = python -m pytest '../tests/stubs'
 
 [flake8]
 # E125 is deliberately excluded. See https://github.com/jcrocholl/pep8/issues/126
