+++ tox.ini
@@ -88,8 +88,11 @@ commands = python -m pytest '../tests/stubs'
 # E711 skipped because sqlalchemy filter() requires using == instead of is
 # max-line-length, E203, W503 are added for black compatibility
 # builtins includes reveal_type because mypy injects it
+# pytest-mypy-testing needs reveal_type to be a one-liner hence we ignore lenght warnings
 max-line-length = 110
 ignore = E125,E123,E129,E711
 extend-ignore = E203, W503
 exclude = .venv,.git,.tox,dist,doc,*egg,build
 builtins=reveal_type
+per-file-ignores =
+  tests/stubs/*:E501
