### 2022-05-06 ed8ea1c
Do not test stubs on Python 3.7
