+++ .github/workflows/main.yml
@@ -90,7 +90,7 @@ jobs:
         run: tox -epy
       - name: 'Run stubs-tests'
         run: tox -estubs
-        if: runner.os == 'Linux'
+        if: runner.os == 'Linux' && ${{ matrix.python-version }} != 3.7
   coverage:
     needs: [tests]
     name: Coverage
