### 2022-05-06 63c87a6
Add test cases handling restricted methods
