+++ tests/digraph/test_centrality.py
@@ -86,14 +86,14 @@ class TestCentralityDiGraph(unittest.TestCase):
         self.assertEqual(expected, betweenness)
 
     def test_closeness_centrality(self):
-        betweenness = retworkx.digraph_closeness_centrality(self.graph)
+        closeness = retworkx.digraph_closeness_centrality(self.graph)
         expected = {0: 0.0, 1: 1.0 / 3.0, 2: 4.0 / 9.0, 3: 0.5}
-        self.assertEqual(expected, betweenness)
+        self.assertEqual(expected, closeness)
 
     def test_closeness_centrality_wf_improved(self):
-        betweenness = retworkx.digraph_closeness_centrality(self.graph, wf_improved=False)
+        closeness = retworkx.digraph_closeness_centrality(self.graph, wf_improved=False)
         expected = {0: 0.0, 1: 1.0, 2: 2.0 / 3.0, 3: 0.5}
-        self.assertEqual(expected, betweenness)
+        self.assertEqual(expected, closeness)
 
 
 class TestCentralityDiGraphDeletedNode(unittest.TestCase):
@@ -140,11 +140,11 @@ class TestCentralityDiGraphDeletedNode(unittest.TestCase):
         self.assertEqual(expected, betweenness)
 
     def test_closeness_centrality(self):
-        betweenness = retworkx.digraph_closeness_centrality(self.graph)
+        closeness = retworkx.digraph_closeness_centrality(self.graph)
         expected = {0: 0.0, 1: 1.0 / 3.0, 2: 4.0 / 9.0, 4: 0.5}
-        self.assertEqual(expected, betweenness)
+        self.assertEqual(expected, closeness)
 
     def test_closeness_centrality_wf_improved(self):
-        betweenness = retworkx.digraph_closeness_centrality(self.graph, wf_improved=False)
+        closeness = retworkx.digraph_closeness_centrality(self.graph, wf_improved=False)
         expected = {0: 0.0, 1: 1.0, 2: 2.0 / 3.0, 4: 0.5}
-        self.assertEqual(expected, betweenness)
+        self.assertEqual(expected, closeness)
