### 2022-05-06 797366b
GraphML parser. (#575)
* GraphML parser.

This commit implements a parser that generates retworkx graph objects from files written in
GraphML format. Since retworkx can't express mixed graphs (directed and unidirected edges
together), hyperedges, nested graphs, or ports, these GraphML features are not supported.

* cargo fmt

* try fixing build error in earlier rustc versions

* remove unused derived debug trait and redundant `finalize_graph` logic + annotate unreachable code parts

* use `assertEqual` not `assert`

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

* remove `error` mod

[…]
