### 2022-05-07 23c9a6d
Make PyDAG be generic
