### 2022-05-07 53d5c7d
Temporarily disable typehints in function signatures
