### 2022-05-07 64895a2
Cleanup PlanarEmbedding
