+++ src/layout/planar.rs
@@ -1,22 +1,26 @@
-use petgraph::{EdgeType, Directed};
 use petgraph::prelude::*;
+use petgraph::{Directed, EdgeType};
+use petgraph::visit::NodeIndexable;
 use pyo3::prelude::*;
 
 use super::spring::{recenter, rescale, Point};
-use crate::iterators::Pos2DMapping;
-use crate::StablePyGraph;
-use crate::Graph;
 use crate::connected_components;
+use crate::iterators::Pos2DMapping;
+use crate::Graph;
+use crate::StablePyGraph;
 use retworkx_core::dictmap::*;
 
-pub struct CwCcw<T>  {
+pub struct CwCcw<T> {
     cw: Option<T>,
     ccw: Option<T>,
 }
 
-impl<T> Default for CwCcw<T>  {
+impl<T> Default for CwCcw<T> {
     fn default() -> Self {
-        CwCcw { cw: None, ccw: None }
+        CwCcw {
+            cw: None,
+            ccw: None,
+        }
     }
 }
 
@@ -65,7 +69,7 @@ pub struct FirstNbr<T> {
     first_nbr: Option<T>,
 }
 
-impl<T> Default for FirstNbr<T>  {
+impl<T> Default for FirstNbr<T> {
     fn default() -> Self {
         FirstNbr { first_nbr: None }
     }
@@ -100,7 +104,6 @@ pub fn planar_layout<Ty: EdgeType>(
     scale: Option<f64>,
     center: Option<Point>,
 ) -> Pos2DMapping {
-
     let node_num = graph.node_count();
     if node_num == 0 {
         return Pos2DMapping {
@@ -125,14 +128,17 @@ pub fn planar_layout<Ty: EdgeType>(
             recenter(&mut pos, center);
         }
         Pos2DMapping {
-            pos_map: planar.embedding.node_indices().map(|n| n.index()).zip(pos).collect(),
+            pos_map: planar
+                .embedding
+                .node_indices()
+                .map(|n| n.index())
+                .zip(pos)
+                .collect(),
         }
     }
 }
 
-pub fn is_planar<Ty: EdgeType>(
-    graph: &StablePyGraph<Ty>,
-) -> bool {
+pub fn is_planar<Ty: EdgeType>(graph: &StablePyGraph<Ty>) -> bool {
     true
 }
 
@@ -140,22 +146,14 @@ pub fn create_embedding<Ty: EdgeType>(
     graph: &StablePyGraph<Ty>,
     embedding: &mut Graph<FirstNbr<NodeIndex>, CwCcw<NodeIndex>, Directed>,
 ) -> bool {
-    // DEBUG CODE FOR TESTING BASIC EMBEDDING
-    // for v in graph.node_indices() {
-    //     if v.index() < 5 {
-    //         println!("GRAPH {:?} {}", v, graph[v].clone());
-    //         embedding.add_node(graph[v].clone());
-    //     } else {
-    //         break;
-    //     }
-    // }
+
     true
 }
 
-pub fn combinitorial_embedding_to_pos (
+pub fn combinitorial_embedding_to_pos(
     embedding: &Graph<FirstNbr<NodeIndex>, CwCcw<NodeIndex>, Directed>,
     pos: &mut Vec<Point>,
-){
+) {
     if embedding.node_count() < 4 {
         *pos = [[0.0, 0.0], [2.0, 0.0], [1.0, 1.0]].to_vec()
     }
@@ -165,7 +163,7 @@ pub fn combinitorial_embedding_to_pos (
     *pos = [[0.4, 0.5]].to_vec()
 }
 
-pub fn triangulate_embedding (
[…]
