### 2022-05-07 3ef2747
Trick sphinx into thinking it is not a double dispatch
