+++ docs/source/conf.py
@@ -156,6 +156,17 @@ def _get_version_label(current_version):
     else:
         return "Development"
 
+def avoid_duplicate_in_dispatch(app, obj, bound_method):
+    if hasattr(obj, 'dispatch') and hasattr(obj, 'register') and obj.dispatch.__module__ == 'functools':
+        # TODO: disable this trick once https://github.com/Qiskit/qiskit_sphinx_theme/issues/21 is fixed
+        # Basically, to avoid signatures being duplicated, we want to disable the singledispatch function
+        # from Sphinx's autodoc. But if we unregister the function, our jupyter notebook executions
+        # will fail. Hence we just trick the check in sphinx/util/inspect/is_singledispatch_function
+        # that checks for obj.dispatch.__module__ == 'functools'. This should be harmless as
+        # that property is only used on __repr__ and not to dispatch the function itself
+        obj.dispatch.__module__ = "retworkx"
+
 
 def setup(app):
     app.connect('config-inited', _get_versions)
+    app.connect('autodoc-before-process-signature', avoid_duplicate_in_dispatch)
