### 2022-05-09 504c188
Add graph attributes to PyGraph and PyDiGraph
This commit adds graph attributes to the retworkx graph classes. This is
done with a new object attribute, ``attrs``, which can contain any
Python object. This enables users to specify any python object for
storing metadata associated with the graph attribute.

Part of #607
