+++ src/graph.rs
@@ -109,17 +109,35 @@ use petgraph::visit::{
 /// if a method call is made that would add a parallel edge it will instead
 /// update the existing edge's weight/data payload.
 ///
+/// Each ``PyGraph`` object has an :attr:`~.PyGraph.attrs` attribute which is
+/// used to contain additional attributes/metadata of the graph instance. By
+/// default this is set to ``None`` but can optionally be specified by using the
+/// ``attrs` keyword argument when constructing a new graph::
+///
+///     graph = retworkx.PyGraph(attrs=dict(source_path='/tmp/graph.csv'))
+///
+/// This attribute can be set to any Python object. Additionally, you can access
+/// and modify this attribute after creating an object. For example::
+///
+///     source_path = graph.attrs
+///     graph.attrs = {'new_path': '/tmp/new.csv', 'old_path': source_path}
+///
 /// :param bool multigraph: When this is set to ``False`` the created PyGraph
 ///     object will not be a multigraph. When ``False`` if a method call is
 ///     made that would add parallel edges the the weight/weight from that
 ///     method call will be used to update the existing edge in place.
+/// :param attrs: An optional attributes payload to assign to the
+///     :attrs:`~.PyGraph.attrs` attribute. This can be any Python object. If
+///     it is not specified :attrs:`~.PyGraph.attrs` will be set to ``None``.
 #[pyclass(mapping, module = "retworkx", subclass)]
-#[pyo3(text_signature = "(/, multigraph=True)")]
+#[pyo3(text_signature = "(/, multigraph=True, attrs=None)")]
 #[derive(Clone)]
 pub struct PyGraph {
     pub graph: StablePyGraph<Undirected>,
     pub node_removed: bool,
     pub multigraph: bool,
+    #[pyo3(get, set)]
+    pub attrs: PyObject,
 }
 
 impl GraphBase for PyGraph {
@@ -158,11 +176,16 @@ impl PyGraph {
 impl PyGraph {
     #[new]
     #[args(multigraph = "true")]
-    fn new(multigraph: bool) -> Self {
+    fn new(py: Python, multigraph: bool, attrs: Option<PyObject>) -> Self {
+        let graph_attrs = match attrs {
+            Some(g_attr) => g_attr,
+            None => py.None(),
+        };
         PyGraph {
             graph: StablePyGraph::<Undirected>::default(),
             node_removed: false,
             multigraph,
+            attrs: graph_attrs,
         }
     }
 
@@ -173,6 +196,7 @@ impl PyGraph {
         out_dict.set_item("nodes", node_dict)?;
         out_dict.set_item("nodes_removed", self.node_removed)?;
         out_dict.set_item("multigraph", self.multigraph)?;
+        out_dict.set_item("attrs", self.attrs.clone_ref(py))?;
         for node_index in self.graph.node_indices() {
             let node_data = self.graph.node_weight(node_index).unwrap();
             node_dict.set_item(node_index.index(), node_data)?;
@@ -204,6 +228,11 @@ impl PyGraph {
             .unwrap()
             .downcast::<PyBool>()?;
         self.multigraph = multigraph_raw.extract()?;
+        let attrs = match dict_state.get_item("attrs") {
+            Some(attr) => attr.into(),
+            None => py.None(),
+        };
+        self.attrs = attrs;
 
         let mut node_indices: Vec<usize> = Vec::new();
         for raw_index in nodes_dict.keys() {
@@ -1053,6 +1082,7 @@ impl PyGraph {
             cycle_state: algo::DfsSpace::default(),
             check_cycle: false,
             multigraph: self.multigraph,
+            attrs: py.None(),
         }
     }
 
[…]
