+++ src/digraph.rs
@@ -147,7 +147,7 @@ use super::dag_algo::is_directed_acyclic_graph;
 /// Each ``PyDiGraph`` object has an :attr:`~.PyDiGraph.attrs` attribute which is
 /// used to contain additional attributes/metadata of the graph instance. By
 /// default this is set to ``None`` but can optionally be specified by using the
-/// ``attrs` keyword argument when constructing a new graph::
+/// ``attrs`` keyword argument when constructing a new graph::
 ///
 ///     graph = retworkx.PyDiGraph(attrs=dict(source_path='/tmp/graph.csv'))
 ///
@@ -164,8 +164,8 @@ use super::dag_algo::is_directed_acyclic_graph;
 ///     method call is made that would add parallel edges the the weight/weight
 ///     from that method call will be used to update the existing edge in place.
 /// :param attrs: An optional attributes payload to assign to the
-///     :attrs:`~.PyDiGraph.attrs` attribute. This can be any Python object. If
-///     it is not specified :attrs:`~.PyDiGraph.attrs` will be set to ``None``.
+///     :attr:`~.PyDiGraph.attrs` attribute. This can be any Python object. If
+///     it is not specified :attr:`~.PyDiGraph.attrs` will be set to ``None``.
 #[pyclass(mapping, module = "retworkx", subclass)]
 #[pyo3(text_signature = "(/, check_cycle=False, multigraph=True, attrs=None)")]
 #[derive(Clone)]
