### 2022-05-09 fff8f72
Add graph attributes to PyGraph and PyDiGraph (#608)
* Add graph attributes to PyGraph and PyDiGraph

This commit adds graph attributes to the retworkx graph classes. This is
done with a new object attribute, ``attrs``, which can contain any
Python object. This enables users to specify any python object for
storing metadata associated with the graph attribute.

Part of #607

* Apply suggestions from code review

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Fix clippy

* Fix doc typo

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>
