+++ tests/graph/test_connected_components.py
@@ -22,6 +22,14 @@ class TestConnectedComponents(unittest.TestCase):
         graph.add_edge(0, 1, None)
         self.assertEqual(retworkx.number_connected_components(graph), 2)
 
+    def test_number_connected_direct(self):
+        graph = retworkx.PyDiGraph()
+        graph.add_nodes_from([0, 1, 2, 3])
+        graph.add_edges_from_no_data([
+            (3, 2), (2, 1), (1, 0)
+        ])
+        self.assertEqual(len(retworkx.weakly_connected_components(graph)), 1)
+    
     def test_number_connected_node_holes(self):
         graph = retworkx.PyGraph()
         graph.add_nodes_from([0, 1, 2])
