+++ src/graphml.rs
@@ -110,6 +110,8 @@ fn xml_attribute<'a, B: BufRead>(
 enum Domain {
     Node,
     Edge,
+    Graph,
+    All,
 }
 
 enum Type {
@@ -187,19 +189,24 @@ struct Graph {
     dir: Direction,
     nodes: Vec<Node>,
     edges: Vec<Edge>,
+    attributes: HashMap<String, Value>,
 }
 
 impl Graph {
-    fn new(dir: Direction) -> Self {
+    fn new<'a, I>(dir: Direction, default_attrs: I) -> Self
+    where
+        I: Iterator<Item = &'a Key>,
+    {
         Self {
             dir,
             nodes: Vec::new(),
             edges: Vec::new(),
+            attributes: HashMap::from_iter(
+                default_attrs.map(|key| (key.name.clone(), key.default.clone())),
+            ),
         }
     }
-}
 
-impl Graph {
     fn add_node<'a, B: BufRead, I>(
         &mut self,
         reader: &Reader<B>,
@@ -300,8 +307,7 @@ impl IntoPy<PyObject> for Graph {
                     graph,
                     node_removed: false,
                     multigraph: true,
-                    // TODO: Add graph attributes here
-                    attrs: py.None(),
+                    attrs: self.attributes.into_py(py),
                 };
 
                 out.into_py(py)
@@ -317,8 +323,7 @@ impl IntoPy<PyObject> for Graph {
                     check_cycle: false,
                     node_removed: false,
                     multigraph: true,
-                    // TODO: Add graph attributes here
-                    attrs: py.None(),
+                    attrs: self.attributes.into_py(py),
                 };
 
                 out.into_py(py)
@@ -334,6 +339,7 @@ enum State {
     Edge,
     DataForNode,
     DataForEdge,
+    DataForGraph,
     Key,
     DefaultForKey,
 }
@@ -356,6 +362,8 @@ struct GraphML {
     graphs: Vec<Graph>,
     key_for_nodes: IndexMap<String, Key>,
     key_for_edges: IndexMap<String, Key>,
+    key_for_graph: IndexMap<String, Key>,
+    key_for_all: IndexMap<String, Key>,
 }
 
 impl Default for GraphML {
@@ -364,6 +372,8 @@ impl Default for GraphML {
             graphs: Vec::new(),
             key_for_nodes: IndexMap::new(),
             key_for_edges: IndexMap::new(),
+            key_for_graph: IndexMap::new(),
+            key_for_all: IndexMap::new(),
         }
     }
 }
@@ -384,7 +394,10 @@ impl GraphML {
             }
         };
 
-        self.graphs.push(Graph::new(dir));
+        self.graphs.push(Graph::new(
+            dir,
+            self.key_for_graph.values().chain(self.key_for_all.values()),
+        ));
 
         Ok(())
     }
@@ -395,7 +408,11 @@ impl GraphML {
         element: &'a BytesStart<'a>,
     ) -> Result<(), Error> {
         if let Some(graph) = self.graphs.last_mut() {
-            graph.add_node(reader, element, self.key_for_nodes.values())?;
+            graph.add_node(
+                reader,
+                element,
+                self.key_for_nodes.values().chain(self.key_for_all.values()),
+            )?;
         }
 
         Ok(())
@@ -407,7 +424,11 @@ impl GraphML {
         element: &'a BytesStart<'a>,
     ) -> Result<(), Error> {
         if let Some(graph) = self.graphs.last_mut() {
-            graph.add_edge(reader, element, self.key_for_edges.values())?;
+            graph.add_edge(
+                reader,
+                element,
+                self.key_for_edges.values().chain(self.key_for_all.values()),
+            )?;
         }
 
[…]
