### 2022-05-09 2a0e5b8
Support graph attributes in GraphML reader. (#610)
GraphML documents may declare attributes/keys for `node`, `edge`, `graph` or `all` graph elements.
This commit adds support for attributes with domain `graph` and `all` by populating `Py(Di)Graph.attrs`
field with a dictionary since the first iteration of our GraphML reader would raise if users tried to
load a graphml file with keys declared for these domains.
