+++ README.md
@@ -71,6 +71,18 @@ these, retworkx is probably still installable, but it’s not tested and you wil
 need a Rust compiler and have to build retworkx (and likely Numpy too) from
 source.
 
+Operating System | CPU Architecture | Support Tier | Notes |
+---------------- | ---------------- | ------------ | ----- |
+Linux | x86_64 | Tier 1 | distributions compatible with the [manylinux 2010](https://peps.python.org/pep-0571/) packaging specification |
+Linux | i686 | Tier 2 (Python < 3.10), Tier 3 (Python >= 3.10) | distributions compatible with the [manylinux 2010](https://peps.python.org/pep-0571/) packaging specification |
+Linux | aarch64 | Tier 2 | distributions compatible with the [manylinux 2014](https://peps.python.org/pep-0599/) packaging specification |
+Linux | pp64le | Tier 3 | distributions compatible with the [manylinux 2014](https://peps.python.org/pep-0599/) packaging specification |
+Linux | s390x | Tier 3 | distributions compatible with the [manylinux 2014](https://peps.python.org/pep-0599/) packaging specification |
+macOS (10.9 or newer) | x86_64 | Tier 1 | |
+macOS (10.15 or newer) | arm64 | Tier 4 | |
+Windows 64bit | x86_64 | Tier 1 | |
+Windows 32bit | i686 | Tier 2 (Python < 3.10), Tier 3 (Python >= 3.10) | |
+
 Additionally, retworkx only supports CPython. Running with other Python
 interpreters isn’t currently supported.
 
@@ -82,28 +94,12 @@ binaries are built, tested, and published to PyPI as part of the release
 process. These platforms are expected to be installable with just a functioning
 Python environment.
 
-Tier 1 platforms are currently:
-
-- Linux x86_64 (distributions compatible with the manylinux 2010 packaging
-  specification)
-- macOS x86_64 (10.9 or newer)
-- Windows 64 bit
-
-
 #### Tier 2
 
 Tier 2 platforms are not tested upstream as part of development process.
 However, pre-compiled binaries are built, tested, and published to PyPI as part
 of the release process and these packages can be expected to be installed with
 just a functioning Python environment.
-
-Tier 2 platforms are currently:
-
-- Linux i686 (distributions compatible with the manylinux 2010 packaging
-  specification) for Python < 3.10
-- Windows 32 bit for Python < 3.10
-- Linux aarch64 (distributions compatible with the manylinux 2014 packaging
-  specification)
 
 #### Tier 3
 
@@ -113,16 +109,6 @@ part of the release process. However, they may not installable with just a
 functioning Python environment and you may be required to build Numpy from
 source, which requires a C/C++ compiler, as part of the installation process.
 
-Tier 3 platforms are currently:
-
-- Linux i686 (distributions compatible with the manylinux 2010 packaging
-  specification) for Python >= 3.10
-- Windows 32bit for Python >= 3.10
-- Linux ppc64le (distributions compatible with the manylinux 2014 packaging
-  specification)
-- Linux s390x (distributions compatible with the manylinux 2014 packaging
-  specification)
-
 #### Tier 4
 
 Tier 4 platforms are not tested upstream as part of the development process.
@@ -131,10 +117,6 @@ process, with no testing at all. They may not be installable with just a
 functioning Python environment and may require a C/C++ compiler or additional
 programs to build dependencies from source as part of the installation process.
 Support for these platforms are best effort only.
-
-Tier 4 platforms are currently:
-
-- macOS arm64 (10.15 or newer)
 
 ### Optional dependencies
 
