### 2022-05-13 a648320
Load lr_planar Create embedding.rs Add embedding functions to embedding.rs
