+++ src/layout/planar.rs
@@ -1,7 +1,4 @@
 use petgraph::prelude::*;
-use petgraph::{Directed, EdgeType};
-use petgraph::visit::NodeIndexable;
-use pyo3::prelude::*;
 
 use super::spring::{recenter, rescale, Point};
 use crate::connected_components;
@@ -9,117 +6,41 @@ use crate::iterators::Pos2DMapping;
 use crate::Graph;
 use crate::StablePyGraph;
 use retworkx_core::dictmap::*;
+use retworkx_core::planar::{PlanarEmbedding, is_planar, create_embedding};
+// use retworkx_core::planar::is_planar;
+// use retworkx_core::planar::create_embedding;
 
-pub struct CwCcw<T> {
-    cw: Option<T>,
-    ccw: Option<T>,
-}
-
-impl<T> Default for CwCcw<T> {
-    fn default() -> Self {
-        CwCcw {
-            cw: None,
-            ccw: None,
-        }
-    }
-}
-
-impl<T> CwCcw<T> {
-    fn new(cw: T, ccw: T) -> Self {
-        CwCcw {
-            cw: Some(cw),
-            ccw: Some(ccw),
-        }
-    }
-
-    fn cw_is_empty(&self) -> bool {
-        self.cw.is_none()
-    }
-
-    fn ccw_is_empty(&self) -> bool {
-        self.ccw.is_none()
-    }
-
-    fn cw_unwrap(self) -> T {
-        self.cw.unwrap()
-    }
-
-    fn ccw_unwrap(self) -> T {
-        self.ccw.unwrap()
-    }
-
-    fn cw_as_ref(&mut self) -> Option<&T> {
-        self.cw.as_ref()
-    }
-
-    fn ccw_as_ref(&mut self) -> Option<&T> {
-        self.ccw.as_ref()
-    }
-
-    fn cw_as_mut(&mut self) -> Option<&mut T> {
-        self.cw.as_mut()
-    }
-
-    fn ccw_as_mut(&mut self) -> Option<&mut T> {
-        self.ccw.as_mut()
-    }
-}
-
-pub struct FirstNbr<T> {
-    first_nbr: Option<T>,
-}
-
-impl<T> Default for FirstNbr<T> {
-    fn default() -> Self {
-        FirstNbr { first_nbr: None }
-    }
-}
-
-impl<T> FirstNbr<T> {
-    fn new(first_nbr: T) -> Self {
-        FirstNbr {
-            first_nbr: Some(first_nbr),
-        }
-    }
-
-    fn first_nbr_is_empty(&self) -> bool {
-        self.first_nbr.is_none()
-    }
-
-    fn first_nbr_unwrap(self) -> T {
-        self.first_nbr.unwrap()
-    }
-
-    fn first_nbr_as_ref(&mut self) -> Option<&T> {
-        self.first_nbr.as_ref()
-    }
-
-    fn first_nbr_as_mut(&mut self) -> Option<&mut T> {
-        self.first_nbr.as_mut()
-    }
-}
-pub fn planar_layout<Ty: EdgeType>(
-    py: Python,
-    graph: &StablePyGraph<Ty>,
+pub fn planar_layout(
+    graph: &StablePyGraph<Undirected>,
     scale: Option<f64>,
     center: Option<Point>,
 ) -> Pos2DMapping {
+
     let node_num = graph.node_count();
+    println!("NODE NUM {}", node_num);
     if node_num == 0 {
         return Pos2DMapping {
             pos_map: DictMap::new(),
         };
     }
-    let mut pos: Vec<Point> = Vec::with_capacity(node_num);
-    let mut planar = PlanarEmbedding::new();
 
-    if !is_planar(graph) {
+    println!("before is");
[…]
