+++ retworkx-core/src/planar/mod.rs
@@ -1,0 +1,19 @@
+// Licensed under the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License. You may obtain
+// a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations
+// under the License.
+
+//! Module for planar graphs.
+
+pub mod lr_planar;
+pub mod embedding;
+
+pub use lr_planar::is_planar;
+pub use embedding::PlanarEmbedding;
