+++ retworkx-core/src/lib.rs
@@ -72,6 +72,7 @@ pub mod centrality;
 pub mod connectivity;
 /// Module for maximum weight matching algorithmss
 pub mod max_weight_matching;
+pub mod planar;
 pub mod shortest_path;
 pub mod traversal;
 // These modules define additional data structures

[…]
