+++ releasenotes/notes/all-pairs-bellman-ford-3e7a3d0ded31427f.yaml
@@ -1,0 +1,9 @@
+---
+features:
+  - |
+    Added new functions to compute the all-pairs shortest path
+    in graphs with negative edge weights using the Bellman-Ford
+    algorithm with the SPFA heuristic:
+
+     * :func:`retworkx.all_pairs_bellman_ford_shortest_path_lengths`
+     * :func:`retworkx.all_pairs_bellman_ford_shortest_paths`
