### 2022-05-14 6bf462b
Use neighbors_directed
