### 2022-05-16 e7409a5
Update to using built-in retworkx macros
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
