### 2022-05-16 a45c147
Take advantage of built-in retworkx macros
