+++ src/tree.rs
@@ -17,14 +17,14 @@ use super::{graph, weight_callable};
 
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
-use pyo3::types::{PyFloat, PyList};
 use pyo3::Python;
 
-use petgraph::algo::{connected_components, is_cyclic_undirected};
 use petgraph::prelude::*;
 use petgraph::stable_graph::EdgeReference;
 use petgraph::unionfind::UnionFind;
 use petgraph::visit::{IntoEdgeReferences, NodeIndexable};
+
+use numpy::PyReadonlyArray1;
 
 use rayon::prelude::*;
 
@@ -165,84 +165,73 @@ pub fn minimum_spanning_tree(
 pub fn balanced_cut_edge(
     py: Python,
     spanning_tree: &graph::PyGraph,
-    py_pops: &PyList,
-    py_pop_target: &PyFloat,
-    py_epsilon: &PyFloat,
+    pops: Vec<f64>,
+    pop_target: f64,
+    epsilon: f64,
 ) -> PyResult<Vec<(usize, Vec<usize>)>> {
-    let epsilon = py_epsilon.value();
-    let pop_target = py_pop_target.value();
-
-    let mut pops: Vec<f64> = vec![]; // not sure if the conversions are needed
-    for i in 0..py_pops.len() {
-        pops.push(py_pops.get_item(i).unwrap().extract::<f64>().unwrap());
+    let mut pops = pops.clone();
+    let spanning_tree_graph = &spanning_tree.graph;
+    let mut same_partition_tracker: Vec<Vec<usize>> =
+        vec![vec![]; spanning_tree_graph.node_count()]; // keeps track of all all the nodes on the same side of the partition
+    let mut node_queue: VecDeque<NodeIndex> = VecDeque::<NodeIndex>::new();
+    for leaf_node in spanning_tree_graph.node_indices() {
+        // todo: filter expr
+        if spanning_tree_graph.neighbors(leaf_node).count() == 1 {
+            node_queue.push_back(leaf_node);
+        }
+        same_partition_tracker[leaf_node.index()].push(leaf_node.index());
     }
 
-    let spanning_tree_graph = &spanning_tree.graph;
+    // eprintln!("leaf nodes: {}", node_queue.len());
 
-    let balanced_nodes = py.allow_threads(move || {
-        let mut same_partition_tracker: Vec<Vec<usize>> =
-            vec![vec![]; spanning_tree_graph.node_count()]; // keeps track of all all the nodes on the same side of the partition
-        let mut node_queue: VecDeque<NodeIndex> = VecDeque::<NodeIndex>::new();
-        for leaf_node in spanning_tree_graph.node_indices() {
-            // todo: filter expr
-            if spanning_tree_graph.neighbors(leaf_node).count() == 1 {
-                node_queue.push_back(leaf_node);
+    // this process can be multithreaded, if the locking overhead isn't too high
+    // (note: locking may not even be needed given the invariants this is assumed to maintain)
+    let mut balanced_nodes: Vec<(usize, Vec<usize>)> = vec![];
+    let mut seen_nodes: Vec<bool> = vec![false; spanning_tree_graph.node_count()]; // todo: perf test this
+    while node_queue.len() > 0 {
+        let node = node_queue.pop_front().unwrap();
+        if seen_nodes[node.index()] {
+            // should not need this
+            // eprintln!("Invalid state! Double vision . . .");
+            continue;
+        }
+        let pop = pops[node.index()];
+
+        // todo: factor out expensive clones
+        // Mark as seen; push to queue if only one unseen neighbor
+        let unseen_neighbors: Vec<NodeIndex> = spanning_tree
+            .graph
[…]
