+++ tests/stubs/pydigraph_test.py
@@ -10,8 +10,7 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-from typing import List, Tuple
-from retworkx import EdgeList, PyDiGraph, NodeIndices, WeightedEdgeList
+from retworkx import PyDiGraph
 
 import pytest
 
