### 2022-05-17 4d93f0e
Use mem::take to save on memory allocs
