### 2022-05-17 f01201c
Revert "Use mem::take to save on memory allocs"
This reverts commit 46d0950b9f8d2222cc8ec44557f94ab22d62d164.
