### 2022-05-17 fdbeaaf
Fix unnecessary clones; address clippy warnings
