### 2022-05-17 517d697
Remove unnecessary dead/commented out code
