### 2022-05-18 2bd4c33
Reduce number of memory allocs
