### 2022-05-18 a360695
Remove no longer relevant todos
