### 2022-05-18 c5d76cb
Switch to using HashSet for seen_nodes tracker
