### 2022-05-20 a9fedab
Finalize structure and start lr_state changes
