### 2022-05-22 629ced6
Add __array__ method that returns empty array
