### 2022-05-22 e48a42b
Make __array__ method work for trivial cases
