### 2022-05-22 11aa9da
Add support for dtype and 1d array of objects
