### 2022-05-23 0c253be
Add connected components to retworkx-core (#595)
* add connected_components to core

* Adjust Graph use

* Complete retworkx-core conn_components and add test

* Add test_connectivity.rs

* Finish tests for connected components in core

* Finish conver to GraphBase

* Convert return from usize to NodeId

* Fmt

* Fix tests and traits

* Make bfs non-pub

* Finish doc tests and set up core test dir

* Add full testing for connected components
and remove retworkx/src/connectivity/conn_components.rs
and modify retworkx/src/connectivity/mod.rs to point
to retworkx/retworkx-core/src/connectivity/co
[…]
