+++ retworkx-core/src/connectivity/mod.rs
@@ -15,7 +15,11 @@
 mod all_simple_paths;
 mod biconnected;
 mod chain;
+mod conn_components;
 
 pub use all_simple_paths::all_simple_paths_multiple_targets;
 pub use biconnected::articulation_points;
 pub use chain::chain_decomposition;
+pub use conn_components::bfs_undirected;
+pub use conn_components::connected_components;
+pub use conn_components::number_connected_components;
