### 2022-05-23 b522daa
Rename functions to biparition_tree and bipartition_graph; update docs
