### 2022-05-23 1e6290d
Make docstrings more descriptive and fix renaming issue
