### 2022-05-23 c283355
Add release notes for bipartition_tree and bipartition_graph
