### 2022-05-23 cbaed09
Add bipartition_graph and bipartition_tree to API section of docs
