### 2022-05-23 41e503a
Shorten description of weight_fn by pointing to minimum_spanning_tree func
