+++ retworkx-core/Cargo.toml
@@ -13,7 +13,7 @@ keywords = ["graph"]
 [dependencies]
 ahash = { version = "0.7.6", default-features = false }
 fixedbitset = "0.4.1"
-petgraph = "0.6.0"
+petgraph = "0.6.1"
 rayon = "1.5"
 
 [dependencies.hashbrown]
