+++ src/iterators.rs
@@ -45,7 +45,7 @@ use num_bigint::BigUint;
 use retworkx_core::dictmap::*;
 
 use ndarray::prelude::*;
-use numpy::{PyArrayDescr, ToPyArray};
+use numpy::{IntoPyArray, PyArrayDescr};
 use pyo3::class::iter::IterNextOutput;
 use pyo3::exceptions::{PyIndexError, PyKeyError, PyNotImplementedError};
 use pyo3::gc::PyVisit;
@@ -423,7 +423,7 @@ macro_rules! py_convert_to_py_array_impl {
     ($($t:ty)*) => ($(
         impl PyConvertToPyArray for Vec<$t> {
             fn convert_to_pyarray(&self, py: Python) -> PyResult<PyObject> {
-                Ok(self.clone().to_pyarray(py).into())
+                Ok(self.clone().into_pyarray(py).into())
             }
         }
     )*)
@@ -434,7 +434,7 @@ macro_rules! py_convert_to_py_array_obj_impl {
         impl PyConvertToPyArray for Vec<$t> {
             fn convert_to_pyarray(&self, py: Python) -> PyResult<PyObject> {
                 let pyobj_vec: Vec<PyObject> = self.iter().map(|x| x.clone().into_py(py)).collect();
-                Ok(pyobj_vec.to_pyarray(py).into())
+                Ok(pyobj_vec.into_pyarray(py).into())
             }
         }
     };
@@ -444,7 +444,6 @@ py_convert_to_py_array_impl! {usize u8 u16 u32 u64 isize i8 i16 i32 i64 f32 f64}
 
 py_convert_to_py_array_obj_impl! {EdgeList}
 py_convert_to_py_array_obj_impl! {(PyObject, Vec<PyObject>)}
-py_convert_to_py_array_obj_impl! {(usize, usize, PyObject)}
 
 impl PyConvertToPyArray for Vec<(usize, usize)> {
     fn convert_to_pyarray(&self, py: Python) -> PyResult<PyObject> {
@@ -455,7 +454,21 @@ impl PyConvertToPyArray for Vec<(usize, usize)> {
             mat[[index, 1]] = element.1;
         }
 
-        Ok(mat.to_pyarray(py).into())
+        Ok(mat.into_pyarray(py).into())
+    }
+}
+
+impl PyConvertToPyArray for Vec<(usize, usize, PyObject)> {
+    fn convert_to_pyarray(&self, py: Python) -> PyResult<PyObject> {
+        let mut mat = Array2::<PyObject>::from_elem((self.len(), 3), py.None());
+
+        for (index, element) in self.iter().enumerate() {
+            mat[[index, 0]] = element.0.into_py(py);
+            mat[[index, 1]] = element.1.into_py(py);
+            mat[[index, 2]] = element.2.clone();
+        }
+
+        Ok(mat.into_pyarray(py).into())
     }
 }
 
