+++ tests/test_custom_return_types.py
@@ -174,6 +174,10 @@ class TestNodeIndicesComparisons(unittest.TestCase):
         self.assertEqual([2, 3], slice_return)
         self.assertEqual([], indices[-1:-2])
 
+    def test_numpy_conversion(self):
+        res = self.dag.node_indexes()
+        np.testing.assert_array_equal(np.asarray(res, dtype=np.uintp), np.array([0, 1]))
+
 
 class TestNodesCountMapping(unittest.TestCase):
     def setUp(self):
@@ -406,11 +410,14 @@ class TestEdgeListComparisons(unittest.TestCase):
         slice_return = edges[0:3:2]
         self.assertEqual([(0, 1), (0, 1)], slice_return)
 
-    def test_numpy_conversion(self):
+    @staticmethod
+    def test_numpy_conversion():
         g = retworkx.generators.directed_star_graph(5)
         res = g.edge_list()
-        # TODO: check dtype argument and assert the result is correct
-        np.asarray(res)
+
+        np.testing.assert_array_equal(
+            np.asarray(res, dtype=np.uintp), np.array([[0, 1], [0, 2], [0, 3], [0, 4]])
+        )
 
 
 class TestWeightedEdgeListComparisons(unittest.TestCase):
@@ -481,6 +488,11 @@ class TestWeightedEdgeListComparisons(unittest.TestCase):
         edges = self.dag.weighted_edge_list()
         slice_return = edges[0:3:2]
         self.assertEqual([(0, 1, "Edgy"), (0, 1, None)], slice_return)
+
+    def test_numpy_conversion(self):
+        np.testing.assert_array_equal(
+            np.asarray(self.dag.weighted_edge_list()), np.array([(0, 1, "Edgy")], dtype=object)
+        )
 
 
 class TestPathMapping(unittest.TestCase):
@@ -1358,6 +1370,10 @@ class TestChainsComparisons(unittest.TestCase):
         self.assertIsInstance(hash_res, int)
         # Assert hash is stable
         self.assertEqual(hash_res, hash(self.chains))
+
+    def test_numpy_conversion(self):
+        # this test assumes the array is 1-dimensional which avoids issues with jagged arrays
+        self.assertTrue(np.asarray(self.chains).shape, (1,))
 
 
 class TestProductNodeMap(unittest.TestCase):
