+++ src/tree.rs
@@ -168,7 +168,6 @@ fn _minimum_spanning_tree<'a>(
 /// :returns: A list of tuples, with each tuple representing a distinct
 /// balanced edge that can be cut. The tuple contains the root of one of the
 /// two partitioned subtrees and the set of nodes making up that subtree.
-///
 #[pyfunction]
 #[pyo3(text_signature = "(spanning_tree, pop, target_pop, epsilon)")]
 pub fn bipartition_tree(
@@ -251,7 +250,6 @@ pub fn bipartition_tree(
 /// :returns: A list of tuples, with each tuple representing a distinct
 /// balanced edge that can be cut. The tuple contains the root of one of the
 /// two partitioned subtrees and the set of nodes making up that subtree.
-///
 #[pyfunction]
 #[pyo3(text_signature = "(graph, weight_fn, pop, target_pop, epsilon)")]
 pub fn bipartition_graph(
