### 2022-05-24 99e5257
Add bipartition tests
