### 2022-05-24 db87924
Fix indent issues in retworkx bipartition docstrings
