### 2022-05-27 f6d807c
Switch to using hashbrown's HashSet impl
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
