### 2022-05-27 46404f3
Make tests deterministic
