### 2022-06-10 Qiskit/retworkx#623 @szhorvat
[joss] Graph size limits
This is a question about the JOSS paper and this sentence:

> retworkx is suitable for modeling graphs ranging from a few nodes scaling up to millions.

What are the actual technical limits (or practical limits) on graph sizes, are these documented, and is there any protection against exceeding them? Is "millions" an actual practical guideline? Any reason it couldn't go to hundreds of millions or even billions?

There seems to be an opportunity for improving the paper here and showing the strengths of retworkx better. I expect this to be a strong point of retworkx. People do work with billion-edge networks these days. Have you tried simple linear-time algorithms such as connected components for such a huge graph?
