### 2022-06-10 Qiskit/retworkx#624 @szhorvat
[joss] Explanation of node indices in paper
This is feedback regarding the following paragraph in the paper:

> A defining characteristic of retworkx graphs is that each node maps to a non-negative integer node index, and similarly, each edge maps to an edge index. Those indices uniquely determine nodes and edges in the graph. Moreover, the indices provide a clear separation between the underlying graph structure and the data associated with payloads

I don't quite understand what the last sentence is meant to communicate, can you clarify please?

And a suggestion regarding the description of indices:

To a reader who does not know retworkx, the statement that there are node indices is likely to seem redundant/uninteresting. Every single graph library I used, with the notable exception of NetworkX, had a well-defined ordering of nodes, and used the concept of node indices, i.e. the `i`th node in the node list.

In retworkx, however, "index" refers to something quite different from what is normally understood (i.e. an ordinal number). retworkx uses an integer ID that can have arbitrary values and is not related to the position of the node in the node order. This way, node IDs are stable upon node removal. If you can convey this information, that would improve the paper. The paragraph, as currently written, doesn't _seem_ to communicate much.

----

On a related note, if I read about this, my first question would be: "This is a good way to uniquely identify nodes, but what about edges? How do we distinguish parallel edges?" It seems to me that not having a consistent way to refer to edges is one of the major shortcomings of the library. But this is well beyond what can be addressed during the review.
