### 2022-06-10 #3968 @szhorvat
The paper is in a much better shape now, and many of the concerns I had were issued. While this library is still young, and has missing features and rough edges, it is making good progress, and show promise. The intended use cases have now been better explained in the paper. I recommend publication in JOSS.

However, I would like to ask the authors to respond about whether they have cleared the name with the NetworkX team. "retworkx" very obviously references "NetworkX", and people will assume a relationship.

----

Some hopefully useful comments to the authors:

Addressing these would improve the paper:

 - https://github.com/Qiskit/retworkx/issues/623
 - https://github.com/Qiskit/retworkx/issues/624

I believe that the lack of a consistent edge representation throughout the API that allows distinguishing parallel edges is a major weekness, and will become a severely limiting factor in the future (unless a decision is made to just not support multigraphs). https://github.com/Qiskit/retworkx/issues/512

Many other suggestions are in the preceding comments and open issues. I hope that at least some of these were useful.
