### 2022-06-10 #3968 @szhorvat
This is a question to the JOSS editors, @drvinceknight

I see several citations of GitHub repos. Does JOSS have guidelines on how to cite repositories, especially how to determine the author list when the project does not provide citation information?
