### 2022-06-10 Qiskit/retworkx#623 @mtreinish
Heh, I raised the same point during the work on the initial draft: https://github.com/Qiskit/retworkx/pull/455#discussion_r737374904. Since that was a while ago I just tested with some of the generators (to eliminate the rng time from that comment) and I can generate a billion node star graph in about 16-25 seconds (taking about 38 GB of RAM) and a billion node square grid graph in ~30 seconds (which took about 60GB of RAM).

Since the nodes are stored in a `Vec` internally the theoretical maximum size of a graph is the `std::usize::MAX - 1`. So on a 64 bit platform it's 2^64 - 1 nodes (it's the same limit for edges too). But, obviously you'll be bound by the available ram on the system well before that number of nodes since each one of those nodes will require at least 64 bits of memory to store the pointer to the python object. When that happens you either get a panic exception for trying to allocate too much memory or the kernel will kill the process with an out of memory error depending on how the allocation is done in the function. Some functions like the [binomal tree generator](https://github.com/Qiskit/retworkx/blob/main/src/generators.rs#L891-L895) will preemptively raise an exception if you try to scale too big where you're encroaching on the limits.

There is some checking in petgraph when we add a node that will also raise a `PanicException` if we exceed the max number of nodes. That being said I think I found a bug in petgraph with this check when trying to make a graph with 10 billion node graph that was causing it to be raised, so that will have to be resolved before scaling larger. But at least things on the order of a billion work today though.

For scaling I ran up to a billion nodes on a cycle graph and ran `connected_components()` on it (and also asserted there was as single component for each step) and it yielded this:

![scaling_v2](https://user-images.githubusercontent.com/2447371/173109616-91f17202-31d6-42fe-82f5-bf8b715fdaed.png)

I also found some inefficiencies in the generators that I'll push a fix up for shortly that should improve the creation time on some of the generators especially at larger scales.
