### 2022-06-10 c4afac1
Optimize performance of some generator functions
This commit makes some improvements to the internals cycle graph, path
graph, star graph, and  mesh graph generator functions to optimize the
performance and reduce the memory overhead when creating the graphs. The
improvements fall into a couple categories, first the graph memory
allocation is only done once now as opposed to on demand as we create
the graph, and second intermediate index lists are no longer used in
places they weren't required to avoid the overhead of having to maintain
dual Vecs.

Related to #623
