### 2022-06-13 29ac996
Use usize for all graph index types
This commit fixes an issue with scaling retworkx graphs above 2**32 - 1
nodes and/or edges. Originally in retworkx we used a usize for all index
types to ensure the graphs could scale as big as the platform would
allow. However, since those earliest retworkx releases we lost the
explicit usize index typing along the way (likely before 0.1.0 and the
library was still very rough and not really used yet). This commit
corrects this oversight and defines a retworkx crate NodeIndex and
EdgeIndex type which are built around a usize and defines StablePyGraph
to be u
[…]
