+++ src/shortest_path/num_shortest_path.rs
@@ -15,12 +15,12 @@ use retworkx_core::dictmap::*;
 use pyo3::exceptions::PyIndexError;
 use pyo3::prelude::*;
 
-use petgraph::graph::NodeIndex;
 use petgraph::visit::{Bfs, NodeIndexable};
 use petgraph::EdgeType;
 
 use num_bigint::{BigUint, ToBigUint};
 
+use crate::NodeIndex;
 use crate::StablePyGraph;
 
 pub fn num_shortest_paths_unweighted<Ty: EdgeType>(
