### 2022-06-17 c1b05cd
Document maximum number of nodes in edges allowed in graph classes
The graph classes use a unsigned 32 bit integer as the index type which
means the maximum allowed size for the number of nodes or edges in a
graph object is std::u32::MAX - 1, roughly 4 billion. We should document
this limitation. In theory we can change the index type to use a usize
(see #628 for where this was attempted) which would increase the upper
limit to 2**64 - 1 (on 64 bit platforms), but there is a significant
memory cost for this and to have graphs so large you will need a
significant amount of local memory. In the 
[…]
