### 2022-06-28 #3968 @drvinceknight
> I believe that the lack of a consistent edge representation throughout the API that allows distinguishing parallel edges is a major weekness, and will become a severely limiting factor in the future (unless a decision is made to just not support multigraphs). [Qiskit/retworkx#512](https://github.com/Qiskit/retworkx/issues/512)
>
> Many other suggestions are in the preceding comments and open issues. I hope that at least some of these were useful.

@IvanIsCoding thanks for your patience, when you have a moment, could you let me know where we are in terms of addressing the remaining comments?
