### 2022-07-02 Qiskit/retworkx#624 @IvanIsCoding
We addressed the comments on the paper, thanks for revieweing the paragraph. With regards to:

>
> On a related note, if I read about this, my first question would be: "This is a good way to uniquely identify nodes, but what about edges? How do we distinguish parallel edges?" It seems to me that not having a consistent way to refer to edges is one of the major shortcomings of the library. But this is well beyond what can be addressed during the review.

I agree that distinguishing parallel edges out-of-the-box is not the strenght of retworkx. It all comes from a design decision and some technical limitations.

In NetworkX, multigraphs are represented by separate classes hence there are `MultiGraph`, `MultiDiGraph` along with the `Graph` and `DiGraph` classes. This allows NetworkX to store the key distinguishing edges for multigraphs. Because of Python is an interpreted language with duck typing, NetworkX methods can accept all the four classes without having to write additional code.

In retworkx, multigraphs are represented by the same classes that represent graphs with non-parallel edges: `PyGraph` and `PyDiGraph`. What differentiates a multigraph from the rest is that the attribute `multigraph` is set to true. This is more convenient for Rust development because Rust is statically typed, hence instead of having internal APIs for 4 types we only have APIs for 2 types (see [API functions for PyGraph](https://qiskit.org/documentation/retworkx/dev/api.html#api-functions-for-pygraph)). The tradeoff is that the implementation for graphs with parallel edges is an extension of the implementation of graphs without parallel edges that did not have key to distinguish between edges.

However, there are some tricks that can help overcome this. For example, if you add edges following this pattern :

```python
payload = {'weight': 1}
edge_index = graph.add_edge(node_a, node_b, payload)
graph.update_edge_by_index(edge_index, {**payload, 'index': edge_index} )
```
You are going to be able to distinguish between parallel edges in a `WeightedEdgeList`
