### 2022-07-02 #3968 @IvanIsCoding
Also, mind me pinging @hagberg, @dschult, @networkx. We are submitting our library, [retworkx](https://github.com/Qiskit/retworkx), to the Journal of Open Source Software. This is the open peer review of our submission.

Before publishing, I wanted to contact you to ask about how you feel about the name of our submission. This came up as one of our reviewer @szhorvat ([igraph](https://github.com/igraph)'s maintainer) pointed out an obvious reference from retworkx to NetworkX. Indeed, the name of our library is an homage to your work (Rust + NetworkX).

Would you be comfortable if we published our library under the current title of retworkx? Or would you prefer if we changed the name of our library before publication?
