### 2022-07-04 #3968 @dschult
I'll bring it up at the weekly core developers meeting this week to say what people feel about it.
My first impression is that the one letter difference and the letters being r and n (which are so close) make it likely that users will get confused between the two. That is not in either of our best interests. There are many other good possibilities. In light of search engine optimization, you should make it distinctive from that perspective.
rustnet?  or maybe something that highlights the type of analysis you focus on for your package?

I'll bring it up at our developer meeting and get back to you soon.
Thanks for asking about it!
