### 2022-07-11 #3968 @mtreinish
@boothby for me obviously not renaming the library would be the best, because trying to manage compatibly and a migration of our existing user base is a lot of effort (it's not exactly small either). But, yeah if you request we rename the library I was thinking something like `rustworkx` or some other name where we can preserve the `rx` shortcut to hopefully ease that transition for people.
