### 2022-07-13 #3968 @dschult
The NetworkX devs would prefer that your package not be named so close to the name networkx.
The names retworkx and networkx are too close for packages that don't try to provide the same API.

Almost all of us also thought retworkx wasn't a very good name. You probably want to make it clear that it is rust. Many also expressed that it would be really cool to have a drop-in Rust version where the API was the same. That way code developed using one package could just be used in the other package by changing the import statement.  But we also understand the subtle choices that have to be made when developing a package. And we appreciate the work toward getting a Rust/Python network analysis tool.

Other good possible names:
RustworkX
Rustworks
RustyNets
NetRust
-- something closer to the type of network analysis you are planning to do.  Or closer to Qiskit.
QisNet
networkQIS
the list could go on...

I personally like QisNet but a lot depends on your long term vision for the package.

The end result is that we would prefer that you change the name.
