### 2022-07-13 Qiskit/retworkx#641 @IvanIsCoding
Proposal Plan To Rename Library
## Proposal Plan To Rename Library

This is a proposal to give a new name for the library. After our submission to the Journal of Open Source Software (https://github.com/openjournals/joss-reviews/issues/3968), one reviewers and the core developers pointed that the name retworkx is very close to NetworkX. As a courtesy to the NetworkX developers, I suggest to change the name of the library.

### Choose a new name

The first step is to choose a new name for the library. Some suggestions from me and :

* reseaux
* ~qisnet~
* ~networkQIS~
* rustworkx
* ~rustynets~
* ~netrust~

We are also open to suggestions.

### Release the next version package under the new name

The second step is to release the next version of the package (which as of now would be 0.12) under the new name. This entails a PR renaming all of our references from retworkx to the new package name we agree on.

This will requires us to:
* Create a new package on PyPI under the new name
* Create a new crate on crates.io under the new name
* Create new documentation links in https://qiskit.org/documentation
* Rename our Github repository
* Publish a new version of the pre-print on arXiv
* Update our record on Zenodo

### Release new versions of retworkx packages depending on the new package

The third step is to provide support for retworkx. Ideally, we should just deprecate the name but continue supporting it for at least some releases  (I suggest two relases, which would be 0.12 and 0.13).

This can be done by making retworkx depend on the new package. For example, retworkx 0.12.0 would depend on the 'newpackagename' 0.12.0', retworkx 0.12.1 would depend on 'newpackagename' 0.12.1 and so on. This can be accomplished by importing from 'newpackage'. Because most of our code is in Rust, we just need to change a couple retworkx Python files to something along the lines of:

```python
from newpackagename import *
```

This step also requires us to:
* Add a disclaimer on README and PyPI that retworkx was renamed and that releases under the old name will stop after a couple versions
* Redirect links from https://qiskit.org/documentation/retworkx to the new documentation
