### 2022-07-13 #3968 @IvanIsCoding
> The NetworkX devs would prefer that your package not be named so close to the name networkx. The names retworkx and networkx are too close for packages that don't try to provide the same API.
>
> Almost all of us also thought retworkx wasn't a very good name. You probably want to make it clear that it is rust. Many also expressed that it would be really cool to have a drop-in Rust version where the API was the same. That way code developed using one package could just be used in the other package by changing the import statement. But we also understand the subtle choices that have to be made when developing a package. And we appreciate the work toward getting a Rust/Python network analysis tool.
>
> Other good possible names: RustworkX Rustworks RustyNets NetRust -- something closer to the type of network analysis you are planning to do. Or closer to Qiskit. QisNet networkQIS the list could go on...
>
> I personally like QisNet but a lot depends on your long term vision for the package.
>
> The end result is that we would prefer that you change the name.

Thanks for the response, we will keep track of the change of name proposal in https://github.com/Qiskit/retworkx/issues/641.

The rollout of this is more complicated than just changing the name of the paper. There is quite a few users depending on us so as Matthew highlighted.

> @boothby for me obviously not renaming the library would be the best, because trying to manage compatibly and a migration of our existing user base (it's not exactly small either) is a lot of effort. But, yeah if you request we rename the library I was thinking something like `rustworkx` or some other name where we can preserve the `rx` shortcut to hopefully ease that transition for people.

I personally suggest `reseaux` for the new package name. It comes from the French [réseaux](https://en.wiktionary.org/wiki/r%C3%A9seau#French), and literally means "networks". The imports would be `import reseaux as rx` which is very close to`import retworkx as rx`
