### 2022-07-13 Qiskit/retworkx#641 @mtreinish
Well for me I think either `reseaux` (which I'm leaning towards) or `rustworkx` are what we should use. That way we can keep the `rx` abbreviated import. So we should decide on a name between those. I definitely don't want to switch to any of the qis names as retworkx is hardly limited to quantum applications.  @georgios-ts @itoko @lcapelluto what are your preferences here?
