### 2022-07-13 7d5f3f6
Rename retworkx to rustworkx
This commit renames the retworkx project to rustworkx. The maintainers of
NetworkX requested that we change the name of the library because they
viewed networkx as being too similar to retworkx. Out of respect for
this request and the NetworkX library which inspired retworkx we are
renaming the library to rustworkx. Rustworkx is a similar name to
retworkx and enables us to keep using the rx abbreviation for imports
which is commonly used.

This commit, migrates all the internal usage and names from retworkx to
rustworkx. This includes the retworkx-core crate which 
[…]
