+++ src/union.rs
@@ -30,7 +30,7 @@ fn extract<T>(x: Entry<T>) -> T {
     match x {
         Entry::Merged(val) => val,
         Entry::Added(val) => val,
-        Entry::None => panic!("Unexpected internal error: called `Entry::extract()` on a `None` value. Please file an issue at https://github.com/Qiskit/retworkx/issues/new/choose with the details on how you encountered this."),
+        Entry::None => panic!("Unexpected internal error: called `Entry::extract()` on a `None` value. Please file an issue at https://github.com/Qiskit/rustworkx/issues/new/choose with the details on how you encountered this."),
     }
 }
 
