+++ rustworkx-core/src/connectivity/conn_components.rs
@@ -1,0 +1,234 @@
+// Licensed under the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License. You may obtain
+// a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations
+// under the License.
+
+use hashbrown::HashSet;
+use std::collections::VecDeque;
+use std::hash::Hash;
+
+use petgraph::visit::{GraphProp, IntoNeighborsDirected, IntoNodeIdentifiers, VisitMap, Visitable};
+use petgraph::{Incoming, Outgoing};
+
+/// Given an graph, a node in the graph, and a visit_map,
+/// return the set of nodes connected to the given node
+/// using breadth first search and treating all edges
+/// as undirected.
+///
+/// Arguments:
+///
+/// * `graph` - The graph object to run the algorithm on
+/// * `node` - The node index to find the connected nodes for
+/// * `discovered()` - The visit map for the graph
+///
+/// # Example
+/// ```rust
+/// use std::iter::FromIterator;
+/// use hashbrown::HashSet;
+/// use petgraph::graph::Graph;
+/// use petgraph::graph::node_index as ndx;
+/// use petgraph::visit::Visitable;
+/// use petgraph::Directed;
+/// use rustworkx_core::connectivity::bfs_undirected;
+///
+/// let graph = Graph::<(), (), Directed>::from_edges(&[
+///     (0, 1),
+///     (1, 2),
+///     (2, 3),
+///     (3, 0),
+///     (4, 5),
+///     (5, 6),
+///     (6, 7),
+///     (7, 4),
+/// ]);
+/// let node_idx = ndx(3);
+/// let component = bfs_undirected(&graph, node_idx, &mut graph.visit_map());
+/// let expected = HashSet::from_iter([ndx(0), ndx(1), ndx(3), ndx(2)]);
+/// assert_eq!(expected, component);
+/// ```
+pub fn bfs_undirected<G>(graph: G, start: G::NodeId, discovered: &mut G::Map) -> HashSet<G::NodeId>
+where
+    G: GraphProp + IntoNeighborsDirected + Visitable,
+    G::NodeId: Eq + Hash,
+{
+    let mut component = HashSet::new();
+    component.insert(start);
+    let mut stack = VecDeque::new();
+    stack.push_front(start);
+
+    while let Some(node) = stack.pop_front() {
+        for succ in graph
+            .neighbors_directed(node, Outgoing)
+            .chain(graph.neighbors_directed(node, Incoming))
+        {
+            if discovered.visit(succ) {
+                stack.push_back(succ);
+                component.insert(succ);
+            }
+        }
+    }
+
+    component
+}
+
+/// Given a graph, return a list of sets of all the
+/// connected components.
+///
+/// Arguments:
+///
+/// * `graph` - The graph object to run the algorithm on
+///
+/// # Example
+/// ```rust
+/// use std::iter::FromIterator;
+/// use hashbrown::HashSet;
+/// use petgraph::graph::Graph;
+/// use petgraph::graph::NodeIndex;
+/// use petgraph::{Undirected, Directed};
+/// use petgraph::graph::node_index as ndx;
+/// use rustworkx_core::connectivity::connected_components;
+///
+/// let graph = Graph::<(), (), Undirected>::from_edges(&[
[…]
