+++ tests/rustworkx_tests/digraph/test_graph_attrs.py
@@ -1,0 +1,31 @@
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+import unittest
+
+import rustworkx
+
+
+class TestAttributes(unittest.TestCase):
+    def test_no_attrs(self):
+        graph = rustworkx.PyDiGraph()
+        self.assertIsNone(graph.attrs)
+
+    def test_attrs_set_at_init(self):
+        graph = rustworkx.PyDiGraph(attrs=dict(foo="bar"))
+        self.assertEqual({"foo": "bar"}, graph.attrs)
+
+    def test_attrs_set_at_init_override(self):
+        graph = rustworkx.PyDiGraph(attrs=dict(foo="bar"))
+        self.assertEqual({"foo": "bar"}, graph.attrs)
+        graph.attrs = "ABC"
+        self.assertEqual("ABC", graph.attrs)

[…]
