### 2022-07-13 Qiskit/retworkx#641 @mtreinish
I put together a commit doing the name migration to reseaux mostly to test the migration path and have idea on how to maintain backwards compatibility during the migration period: https://github.com/mtreinish/retworkx/commit/8a2245bd9558a5f01d7b217a69768d71aeaab36f it's still missing the documentation redirect publishing (which I'll add to the branch later) along with the long tail of renaming things manually (like repo, slack channel, irc channel, etc).

I'd like to wait some more time to collect more opinions on the name to make sure there is a wider consensus on using reseaux before I open a PR doing the migration.
