### 2022-07-14 Qiskit/retworkx#641 @itoko
I like `reseaux` too, (or `rustworkx`, which has a better Google searchability).
