### 2022-07-14 Qiskit/retworkx#641 @IvanIsCoding
> For the specific steps, I think there are a couple of things we'll need to do differently than the plan you outlined. First just doing a compat shim with a `*` import isn't actually sufficient. This is mostly because of imports of subpackages, such as `from retworkx.generators import star_graph` , which won't work with a `*` import. I have a pattern to resolve this though we can use a meta importer to add a redirect from retworkx imports to the new namespace (we'll basically just copy what I did in [Qiskit/qiskit-terra#5089](https://github.com/Qiskit/qiskit-terra/pull/5089) although change the fallback behavior to prefer the redirect). With that hook in place I think it might be easier to just publish both packages side by side with identical code. So for 0.12.0 we publish both retworkx and the new name as identical packages.
>

You are correct about the imports, your solution is better and I am glad we can reuse the code from qiskit_aer.

I still think we should make `retworkx` depend on `reseaux` (or any new name), it has some advantages. This was easy to build on top of the PR you wrote, I just had to change a couple lines: https://github.com/IvanIsCoding/retworkx/blob/9b75cfccc8f6b6c4c6b55740cec204ddf79cfce8/setup.py

Some advantages are:
* `reseaux` will be on the dependency tree of packages that use the latest release of `retworkx`, so doing `pip install reseaux` does not have to download the same code twice
* it allows `retworkx<=0.11` to coexist with `reseaux>=0.12` in the same machine, nothing breaks
* it disallows cases like `retworkx==0.12.0` and `reseaux==0.13` being installed in the same machine, as the version of `retworkx` is tied to `reseaux`
