### 2022-07-14 Qiskit/retworkx#641 @enavarro51
Just thought I'd give you my 2-cents on the name change. The French word reseaux has an acute accent on the first e, so it's pronounced ray-zo. If I didn't see the accent, I might pronounce it more like ru-so. I'm sure there is no need for the accent as a product name, but there might be some French who would object. I also know we have an international following, and I assume most would know a lot more English than French, which might make them (and some American/English) try to pronounce it as ru-zooks, or some such.

I like rustworks or rustworkx, but there apparently is a game called Rust, and there's a website, [rustworks.org](http://rustworks.org/), that is for gamers to meet up and discuss the game. There was also a Github repo called Rustworks that is a Rust language OS for Numworks, whatever that is.

The only items for rustworkx are Rust Worx and Rust Werx, which both seem to be automotive related, so no problem there. I personally like rustworkx. It's pretty descriptive, and it has a gritty, steel mill quality, that might be a good visual for people to remember it. Also pronunciation should be no more of a problem for anyone than networkx is.
