### 2022-07-25 Qiskit/retworkx#641 @mtreinish
So after talking to a number of users I'm thinking we should rename to rustworkx instead of reseaux. Everyone liked the reseaux name once they understood it, but since English is the language for retworkx most people ended up confused by the reseaux name initially, not knowing French. While rustworkx being closer to the original retworkx name was seen as an easier transition to make. I'll start adapting my commit to use rustworkx and push that as a PR.
