+++ tests/rustworkx_tests/digraph/test_edges.py
@@ -1,0 +1,938 @@
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+import unittest
+
+import rustworkx
+
+
+class TestEdges(unittest.TestCase):
+    def test_get_edge_data(self):
+        dag = rustworkx.PyDAG()
+        node_a = dag.add_node("a")
+        node_b = dag.add_child(node_a, "b", "Edgy")
+        res = dag.get_edge_data(node_a, node_b)
+        self.assertEqual("Edgy", res)
+
+    def test_get_all_edge_data(self):
+        dag = rustworkx.PyDAG()
+        node_a = dag.add_node("a")
+        node_b = dag.add_child(node_a, "b", "Edgy")
+        dag.add_edge(node_a, node_b, "b")
+        res = dag.get_all_edge_data(node_a, node_b)
+        self.assertIn("b", res)
+        self.assertIn("Edgy", res)
+
+    def test_no_edge(self):
+        dag = rustworkx.PyDAG()
+        node_a = dag.add_node("a")
+        node_b = dag.add_node("b")
+        self.assertRaises(rustworkx.NoEdgeBetweenNodes, dag.get_edge_data, node_a, node_b)
+
+    def test_num_edges(self):
+        graph = rustworkx.PyDiGraph()
+        graph.add_node(1)
+        graph.add_node(42)
+        graph.add_node(146)
+        graph.add_edges_from_no_data([(0, 1), (1, 2)])
+        self.assertEqual(2, graph.num_edges())
+
+    def test_num_edges_no_edges(self):
+        graph = rustworkx.PyDiGraph()
+        graph.add_node(1)
+        graph.add_node(42)
+        self.assertEqual(0, graph.num_edges())
+
+    def test_update_edge(self):
+        dag = rustworkx.PyDAG()
+        node_a = dag.add_node("a")
+        node_b = dag.add_child(node_a, "b", "not edgy")
+        dag.update_edge(node_a, node_b, "Edgy")
+        self.assertEqual([(0, 1, "Edgy")], dag.weighted_edge_list())
+
+    def test_update_edge_no_edge(self):
+        dag = rustworkx.PyDAG()
+        node_a = dag.add_node("a")
+        node_b = dag.add_node("b")
+        self.assertRaises(rustworkx.NoEdgeBetweenNodes, dag.update_edge, node_a, node_b, None)
+
+    def test_update_edge_by_index(self):
+        dag = rustworkx.PyDAG()
+        node_a = dag.add_node("a")
+        dag.add_child(node_a, "b", "not edgy")
+        dag.update_edge_by_index(0, "Edgy")
+        self.assertEqual([(0, 1, "Edgy")], dag.weighted_edge_list())
+
+    def test_update_edge_invalid_index(self):
+        dag = rustworkx.PyDAG()
+        dag.add_node("a")
+        dag.add_node("b")
+        self.assertRaises(IndexError, dag.update_edge_by_index, 0, None)
+
+    def test_update_edge_parallel_edges(self):
+        graph = rustworkx.PyDiGraph()
+        node_a = graph.add_node("a")
[…]
