### 2022-08-03 #3968 @IvanIsCoding
@drvinceknight @arfon We have changed the name of the library to be rustworkx and need to update the title and repository link. "rustworkx: A High-Performance Graph Library for Python" and https://github.com/Qiskit/rustworkx

 .
