### 2022-10-07 #3968 @IvanIsCoding
@arfon  @drvinceknight We have officially released the library under the name of `rustworkx`.

We have archived the library in Zenodo, can you help set up the archive with editorialbot? https://doi.org/10.5281/zenodo.7158473
