### 2022-10-26 #3968 @IvanIsCoding
> @IvanIsCoding - please update the metadata in your zenodo deposit to change the title to match the title of the software paper

We have updated the title, it matches the paper’s title now
