### 2022-10-28 #3968 @IvanIsCoding
> @IvanIsCoding in your paper.md file, you have 2 figures that use the same label (`\label{fig:dagexample}`) - please change one of them to something different

I have updated the label, it should work now
