### 2022-11-01 #3968 @danielskatz
👋 @IvanIsCoding - I've now proofread the paper, and have made suggestions for small changes in https://github.com/Qiskit/rustworkx/pull/720 - please merge this, or let me know what you disagree with, then we can finish the process
