### 2022-11-01 #3968 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
